+++ .github/workflows/basic_tests_for_aud_csv.yaml
@@ -15,7 +15,7 @@ jobs:
     - name: Install dependencies
       run: |
         sudo apt-get -y install sox libportaudio2
-        python3 -m pip install --upgrade pip
+        # python3 -m pip install --upgrade pip
         pip install -r requirements.txt
         # pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
         pip install --editable .
