+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.91.2"
+VERSION="0.91.3"
 SAMPLING_RATE = 16000
