+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,5 @@
 name: Draft PDF
-on: [push]
+on: [merge]
 
 jobs:
   paper:
