+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.91.3"
+VERSION="0.92.0"
 SAMPLING_RATE = 16000
