### 2024-11-07 #7430 @isdanni
Happy to review this, I'm working on speech technology full-time.
