+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.92.0"
+VERSION="0.92.1"
 SAMPLING_RATE = 16000
