+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.92.1"
+VERSION="0.92.2"
 SAMPLING_RATE = 16000
