+++ nkululeko/data/dataset_csv.py
@@ -53,7 +53,7 @@ class Dataset_CSV(Dataset):
             if audformat.index_type(df.index) == "segmented":
                 file_index = (
                     df.index.levels[0]
-                    .map(lambda x: root + "/" + audio_path + "/" + x)
+                    .map(lambda x: os.path.join(root, audio_path, x))
                     .values
                 )
                 df = df.set_index(df.index.set_levels(file_index, level="file"))
@@ -62,20 +62,20 @@ class Dataset_CSV(Dataset):
                     df = pd.DataFrame(df)
                 df = df.set_index(
                     df.index.to_series().apply(
-                        lambda x: root + "/" + audio_path + "/" + x
+                        lambda x: os.path.join(root, audio_path, x)
                     )
                 )
         else:  # absolute path is True
             if audformat.index_type(df.index) == "segmented":
                 file_index = (
-                    df.index.levels[0].map(lambda x: audio_path + "/" + x).values
+                    df.index.levels[0].map(lambda x: os.path.join(audio_path, x)).values
                 )
                 df = df.set_index(df.index.set_levels(file_index, level="file"))
             else:
                 if not isinstance(df, pd.DataFrame):
                     df = pd.DataFrame(df)
                 df = df.set_index(
-                    df.index.to_series().apply(lambda x: audio_path + "/" + x)
+                    df.index.to_series().apply(lambda x: os.path.join(audio_path, x))
                 )
 
         self.df = df
