+++ nkululeko/feat_extract/feats_wavlm.py
@@ -8,10 +8,11 @@ import pandas as pd
 import torch
 import torchaudio
 from tqdm import tqdm
-from transformers import Wav2Vec2FeatureExtractor, WavLMModel
+from transformers import Wav2Vec2FeatureExtractor
+from transformers import WavLMModel
 
-import nkululeko.glob_conf as glob_conf
 from nkululeko.feat_extract.featureset import Featureset
+import nkululeko.glob_conf as glob_conf
 
 
 class Wavlm(Featureset):
