+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.92.2"
+VERSION="0.93.0"
 SAMPLING_RATE = 16000
