### 2024-11-12 b6a71db
update readme and paper (minors)
