+++ README.md
@@ -57,22 +57,22 @@ And can show the distribution of specific features per category:
 <img src="meta/images/feat_dist.png" width="500px"/>
 
 ### t-SNE plots
-A t-SNE plot can give you an estimate wether your acoustic features are useful at all:
+A t-SNE plot can give you an estimate of whether your acoustic features are useful at all:
 
 <img src="meta/images/tsne.png" width="500px"/>
 
 ### Data distribution
-Sometimes you only want to take a look at your data:
+Sometimes, you only want to take a look at your data:
 
 <img src="meta/images/data_plot.png" width="500px"/>
 
 ### Bias checking
-In cases you might wonder if there's bias in your data. You can try to detect this with automatically estimated speech properties, by visualizing the correlation of target label and predicted labels.
+In some cases, you might wonder if there's bias in your data. You can try to detect this with automatically estimated speech properties by visualizing the correlation of target labels and predicted labels.
 
 <img src="meta/images/emotion-pesq.png" width="500px"/>
 
 ### Uncertainty
-Nkululeko estimates uncertainty of model decision (only for classifiers) with entropy over the class-probabilities or logits per sample.
+Nkululeko estimates the uncertainty of model decisions (only for classifiers) with entropy over the class probabilities or logits per sample.
 
 <img src="meta/images/uncertainty.png" width="500px"/>
 
@@ -95,7 +95,7 @@ appears, please try
 ```
 pip install x
 ```
-For many packages you will need the missing torch package.
+For many packages, you will need the missing torch package.
 If you don't have a GPU (which is probably true if you don't know what that is), please use
 ```
 pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
@@ -176,7 +176,7 @@ All of them take *--config <my_config.ini>* as an argument.
 * **nkululeko.predict**: [predict features](http://blog.syntheticspeech.de/2023/08/16/nkululeko-how-to-predict-labels-for-your-data-from-existing-models-and-check-them/) like SNR, MOS, arousal/valence, age/gender, with DNN models
 * **nkululeko.segment**: [segment a database](http://blog.syntheticspeech.de/2023/07/14/nkululeko-segmenting-a-database/) based on VAD (voice activity detection)
 * **nkululeko.resample**: check on all [sampling rates and change](http://blog.syntheticspeech.de/2023/08/31/how-to-fix-different-sampling-rates-in-a-dataset-with-nkululeko/) to 16kHz 
-* **nkululeko.nkuluflag**: a convenient module to specify configuration parameters on the command-line. Usage:
+* **nkululeko.nkuluflag**: a convenient module to specify configuration parameters on the command line. Usage:
 
   ```bash
   $ python -m nkululeko.nkuluflag.py [-h] [--config CONFIG] [--data [DATA ...]] [--label [LABEL ...]] [--tuning_params [TUNING_PARAMS ...]] [--layers [LAYERS ...]] [--model MODEL] [--feat FEAT] [--set SET] [--with_os WITH_OS] [--target TARGET] [--epochs EPOCHS] [--runs RUNS] [--learning_rate LEARNING_RATE] [--drop DROP]
@@ -193,7 +193,7 @@ There's my [blog](http://blog.syntheticspeech.de/?s=nkululeko) with tutorials:
 * [Combine feature sets](http://blog.syntheticspeech.de/2022/06/30/how-to-combine-feature-sets-with-nkululeko/)
 * [Classifying continuous variables](http://blog.syntheticspeech.de/2022/01/26/nkululeko-classifying-continuous-variables/) 
[…]
