+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.1"
+VERSION="0.93.2"
 SAMPLING_RATE = 16000
