+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.2"
+VERSION="0.93.3"
 SAMPLING_RATE = 16000
