+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.3"
+VERSION="0.93.4"
 SAMPLING_RATE = 16000
