+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.4"
+VERSION="0.93.5"
 SAMPLING_RATE = 16000
