+++ tests/data_roots.ini
@@ -10,6 +10,7 @@ crema-d.split_strategy = specified
 crema-d.colnames = {'sex':'gender'}
 crema-d.files_table = ['files']
 crema-d.target_tables = ['emotion.categories.desired.test','emotion.categories.desired.train', 'emotion.categories.desired.dev']
+crema-d.target_tables_append = True
 crema-d.test_tables = ['emotion.categories.desired.test']
 crema-d.train_tables = ['emotion.categories.desired.train', 'emotion.categories.desired.dev']
 crema-d.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
