+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.6"
+VERSION="0.93.7"
 SAMPLING_RATE = 16000
