+++ nkululeko/plots.py
@@ -242,84 +242,100 @@ class Plots:
 
     def _plot2cont_cat(self, df, cont1, cont2, cat, ylab):
         """Plot relation of two continuous distributions with one categorical."""
+        plot_df = df[[cont1, cont2, cat]].copy()
         if cont2 == "class_label":
-            df.rename(columns={cont2: self.target})
+            plot_df = plot_df.rename(columns={cont2: self.target})
             cont2 = self.target
         if cont1 == "class_label":
-            df.rename(columns={cont1: self.target})
+            plot_df = plot_df.rename(columns={cont1: self.target})
             cont1 = self.target
         if cat == "class_label":
-            df.rename(columns={cat: self.target})
+            plot_df = plot_df.rename(columns={cat: self.target})
             cat = self.target
-        pearson = stats.pearsonr(df[cont1], df[cont2])
+        pearson = stats.pearsonr(plot_df[cont1], plot_df[cont2])
         # trunc to three digits
         pearson = int(pearson[0] * 1000) / 1000
         pearson_string = f"PCC: {pearson}"
         ccc_string = ""
         if self.with_ccc:
-            ccc_val = ccc(df[cont1], df[cont2])
+            ccc_val = ccc(plot_df[cont1], plot_df[cont2])
             ccc_val = int(ccc_val * 1000) / 1000
             ccc_string = f"CCC: {ccc_val}"
-        ax = sns.lmplot(data=df, x=cont1, y=cont2, hue=cat)
-        caption = f"{ylab} {df.shape[0]}. {pearson_string} {ccc_string}"
+        ax = sns.lmplot(data=plot_df, x=cont1, y=cont2, hue=cat)
+        caption = f"{ylab} {plot_df.shape[0]}. {pearson_string} {ccc_string}"
         ax.figure.suptitle(caption)
         return ax, caption
 
     def _plot2cont(self, df, col1, col2, ylab):
         """Plot relation of two continuous distributions."""
+        plot_df = df[[col1, col2]].copy()
         # rename "class_label" to the original target
         if col2 == "class_label":
-            df.rename(columns={col2: self.target})
+            plot_df = plot_df.rename(columns={col2: self.target})
             col2 = self.target
         if col1 == "class_label":
-            df.rename(columns={col1: self.target})
+            plot_df = plot_df.rename(columns={col1: self.target})
             col1 = self.target
-        pearson = stats.pearsonr(df[col1], df[col2])
+        pearson = stats.pearsonr(plot_df[col1], plot_df[col2])
         # trunc to three digits
         pearson = int(pearson[0] * 1000) / 1000
         pearson_string = f"PCC: {pearson}"
         ccc_string = ""
         if self.with_ccc:
-            ccc_val = ccc(df[col1], df[col2])
+            ccc_val = ccc(plot_df[col1], plot_df[col2])
             ccc_val = int(ccc_val * 1000) / 1000
             ccc_string = f"CCC: {ccc_val}"
-        ax = sns.lmplot(data=df, x=col1, y=col2)
-        caption = f"{ylab} {df.shape[0]}. {pearson_string} {ccc_string}"
+        ax = sns.lmplot(data=plot_df, x=col1, y=col2)
+        caption = f"{ylab} {plot_df.shape[0]}. {pearson_string} {ccc_string}"
         ax.figure.suptitle(caption)
         return ax, caption
 
     def plotcatcont(self, df, cat_col, cont_col, xlab, ylab):
         """Plot relation of categorical distribution with continuous."""
         # rename "class_label" to the original target
+        plot_df = df[[cat_col, cont_col]].copy()
         if cat_col == "class_label":
-            df.rename(columns={cat_col: self.target})
[…]
