+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.7"
+VERSION="0.93.8"
 SAMPLING_RATE = 16000
