+++ setup.cfg
@@ -36,8 +36,8 @@ install_requires =
     scipy
     seaborn
     sounddevice
-    tensorflow
-    tensorflow_hub
+    # tensorflow
+    # tensorflow_hub
     torch
     torchvision
     transformers
