+++ nkululeko/models/model_mlp_regression.py
@@ -36,6 +36,10 @@ class MLP_Reg_model(Model):
         else:
             self.util.error(f"unknown loss function: {criterion}")
         self.util.debug(f"training model with {criterion} loss function")
+        manual_seed = eval(self.util.config_val("MODEL", "manual_seed", "True"))
+        if manual_seed:
+            self.util.debug(f"seeding random to {23}")
+            torch.manual_seed(23)
         # set up the model
         cuda = "cuda" if torch.cuda.is_available() else "cpu"
         self.device = self.util.config_val("MODEL", "device", cuda)
