+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.8"
+VERSION="0.93.9"
 SAMPLING_RATE = 16000
