+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.9"
+VERSION="0.93.10"
 SAMPLING_RATE = 16000
