+++ paper/paper.bib
@@ -156,7 +156,7 @@ title = {Pathological Voice Detection From Sustained Vowels: Handcrafted vs. Sel
 year = {2025}
 }
 
-@inproceedings{Atmaja2025b,
+@inproceedings{Atmaja:2025b,
 abstract = {Speech classification has attracted increasing attention due to its wide applications, particularly in classifying physical and mental states. However, these tasks are challenging due to the high variability in speech signals. Ensemble learning has shown promising results when multiple classifiers are combined to improve performance. With recent advancements in hardware development, combining several models is not a limitation in deep learning research and applications. In this paper, we propose an uncertainty-based ensemble learning approach for speech classification. Specifically, we train a set of base features on the same classifier and quantify the uncertainty of their predictions. The predictions are combined using variants of uncertainty calculation to produce the final prediction. The visualization of the effect of uncertainty and its ensemble learning results show potential improvements in speech classification tasks. The proposed method outperforms single models and conventional ensemble learning methods in terms of unweighted accuracy or weighted accuracy.},
 address = {Fukuoka},
 author = {Atmaja, Bagus Tris and Burkhardt, Felix and Sasou, Akira},
