+++ paper/paper.bib
@@ -134,19 +134,20 @@ year = {2022}
 @inproceedings{Atmaja:2024a,
 abstract = {Speech classification has attracted increasing attention due to its wide applications, particularly in classifying physical and mental states. However, these tasks are challenging due to the high variability in speech signals. Ensemble learning has shown promising results when multiple classifiers are combined to improve performance. With recent advancements in hardware development, combining several models is not a limitation in deep learning research and applications. In this paper, we propose an uncertainty-based ensemble learning approach for speech classification. Specifically, we train a set of base features on the same classifier and quantify the uncertainty of their predictions. The predictions are combined using variants of uncertainty calculation to produce the final prediction. The visualization of the effect of uncertainty and its ensemble learning results show potential improvements in speech classification tasks. The proposed method outperforms single models and conventional ensemble learning methods in terms of unweighted accuracy or weighted accuracy.},
 author = {Atmaja, Bagus Tris and Sasou, Akira and Burkhardt, Felix},
-booktitle = {Orient. COCOSDA},
+booktitle = {2024 27th Conference of the Oriental COCOSDA International Committee for the Co-ordination and Standardisation of Speech Databases and Assessment Techniques (O-COCOSDA)},
 pages = {1--6},
-title = {{UNCERTAINTY-BASED ENSEMBLE LEARNING FOR SPEECH CLASSIFICATION}},
-year = {2024}
+title = {Uncertainty-Based Ensemble Learning for Speech Classification},
+year = {2024},
+doi={10.1109/O-COCOSDA64382.2024.10800111}
 }
 
 @inproceedings{Burkhardt:2024,
 author = {Burkhardt, Felix and Atmaja, Bagus Tris and Derington, Anna and Eyben, Florian},
-booktitle = {Orient. COCOSDA},
-file = {:home/bagus/Documents/Mendeley Desktop/Oral Session 1{\_}1.pdf:pdf},
+booktitle = {2024 27th Conference of the Oriental COCOSDA International Committee for the Co-ordination and Standardisation of Speech Databases and Assessment Techniques (O-COCOSDA)},
 pages = {1--6},
-title = {{CHECK YOUR AUDIO DATA : NKULULEKO FOR BIAS DETECTION}},
-year = {2024}
+title = {Check Your Audio Data: Nkululeko for Bias Detection},
+year = {2024},
+doi={10.1109/O-COCOSDA64382.2024.10800580}
 }
 
 @inproceedings{Atmaja:2025,
