### 2024-12-25 #7430 @bagustris
> @felixbur 👋 thanks for your submission. I am now going to check for reviewers. Do you have some ideas yourself? If not, I would like to ask some of the people in the speech community like
>
>     * [@fakufaku](https://github.com/fakufaku?rgh-link-date=2024-11-21T22%3A03%3A13Z)
>
>     * [@popcornell](https://github.com/popcornell?rgh-link-date=2024-11-21T22%3A03%3A13Z)
>
>     * [@hbredin](https://github.com/hbredin?rgh-link-date=2024-11-21T22%3A03%3A13Z)
>
>     * [@boeddeker](https://github.com/boeddeker?rgh-link-date=2024-11-21T22%3A03%3A13Z)
>
>     * [@iver56](https://github.com/iver56?rgh-link-date=2024-11-21T22%3A03%3A13Z)
>
>     * ...
>
>
> Let me know if any of them would flag a COI wrt to our [joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)

Hi @faroit, previously @isdanni is also interested to review this submission.
