### 2024-12-26 #7430 @bagustris
I am also proposing other reviewers below based on the availability in joss reviewers system with keyword `speech` and their Github activity:
- [pzelasko](https://reviewers.joss.theoj.org/reviewers/1266)
- [Yigitcan Özer](https://reviewers.joss.theoj.org/reviewers/4033)
- [Yen-Ju Lu](https://reviewers.joss.theoj.org/reviewers/2624)
- [jorgedch](https://reviewers.joss.theoj.org/reviewers/2112)
