+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.10"
+VERSION="0.93.11"
 SAMPLING_RATE = 16000
