### 2025-01-15 #7430 @yiitozer
Hi @faroit, sorry for my delayed reply. Unfortunately, I won’t be able to make it this time.
