+++ setup.cfg
@@ -4,8 +4,8 @@ author = Felix Burkhardt
 author_email = fxburk@gmail.com
 description = Machine learning audio prediction experiments based on templates
 version = attr: nkululeko.__version__
-long_description = file: README.md, CHANGELOG.md
-long_description_content_type = text/markdown
+# long_description = file: README.md, CHANGELOG.md
+# long_description_content_type = text/markdown
 url = https://github.com/felixbur/nkululeko
 classifiers =
     Programming Language :: Python :: 3
