### 2025-01-22 #7430 @faroit
> the library has already been submitted and accepted at lrec2022...

@felixbur can you please comment on the existing publication from lrec and the changes to the code since then?
