### 2025-01-22 #7430 @felixbur
Sure, since i develop this framework i try to publish at conferences as often as possible to get more attention.
Nonetheless, we constantly add features and change focus, based on our own research interests.

New features since last publications would be, amongst others

* ensemble learning
* Wav2vec2 finetuning
* adding speaker idenitification
* augmentation and segmentation

Here is a list of publications so far

* F. Burkhardt, Bagus Tris Atmaja, Anna Derington, Florian Eyben and Björn Schuller: Check Your Audio Data: Nkululeko for Bias Detection, Proc. Oriental COCOSDA, 2024
* F. Burkhardt, Florian Eyben and Björn Schuller: Nkululeko: Machine Learning Experiments on Speaker Characteristics Without Programming, Proc. Interspeech, 2023
* [F. Burkhardt] Johannes Wagner, Hagen Wierstorf, Florian Eyben and Björn Schuller: Nkululeko: A Tool For Rapid Speaker Characteristics Detection, Proc. [Proc. LREC](https://lrec2022.lrec-conf.org/en/), 2022
