+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.12"
+VERSION="0.93.13"
 SAMPLING_RATE = 16000
