+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Version 0.93.13 (24-01-27)
+--------------------------
+* plot: replaced "class_label" by target name for continuous distributions
+
 Version 0.93.12
 ---------------
 * bugfix: map_continuous_to_cat crashed on empty data
