+++ nkululeko/utils/stats.py
@@ -1,7 +1,8 @@
-import math
 from itertools import combinations
+import math
 
 import numpy as np
+import pandas as pd
 
 
 def check_na(a):
@@ -14,9 +15,8 @@ def check_na(a):
         return a
 
 
-def cohen_d(d1, d2):
-    """
-    Compute Cohen's d from two distributions of real valued arrays.
+def cohen_d(d1: np.array, d2: np.array) -> float:
+    """Compute Cohen's d from two distributions of real valued arrays.
 
     Args:
         d1: one array
@@ -50,7 +50,9 @@ def all_combinations(items_list):
     return result
 
 
-def get_effect_size(df, target, variable):
+def get_effect_size(
+    df: pd.DataFrame, target: str, variable: str
+) -> tuple[str, str, dict]:
     """Get the effect size as Cohen's D.
 
     Effect size is computed  from a real numbered variable on a categorical target.
@@ -68,10 +70,10 @@ def get_effect_size(df, target, variable):
     for c in categories:
         cats[c] = df[df[target] == c][variable].values
     combos = all_combinations(categories)
-    results = {}
+    results = {categories[0]: 0}
     if len(categories) == 1:
         cat_s = cohens_D_to_string(0)
-        return categories[0], cat_s, 0
+        return categories[0], cat_s, results
     else:
         for combo in combos:
             one = combo[0]
@@ -79,10 +81,10 @@ def get_effect_size(df, target, variable):
             results[f"{one}-{other}"] = cohen_d(cats[one], cats[other])
         max_cat = max(results, key=results.get)
         cat_s = cohens_D_to_string(float(results[max_cat]))
-    return max_cat, cat_s, results[max_cat]
+    return max_cat, cat_s, results
 
 
-def cohens_D_to_string(val):
+def cohens_D_to_string(val: float) -> str:
     if val < 0.2:
         rval = "no effect"
     elif val < 0.2:
