+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.14"
+VERSION="0.93.15"
 SAMPLING_RATE = 16000
