+++ data/jvnv/exp.ini
@@ -1,6 +1,6 @@
 [EXP]
-root = /tmp/results/
-name = exp_jvnv_hubert_all
+root = ./results/
+name = exp_jvnv_hubert
 ; runs = 1
 ; epochs = 1
 ; save = True
@@ -19,14 +19,15 @@ test.type = csv
 test.absolute_path = False
 test.split_strategy = test
 target = emotion
-; labels = ['anger', 'fear', 'sad', 'happy']
+labels = ['anger', 'fear', 'sad', 'happy']
 ; get the number of classes from the target column automatically
 [FEATS]
 type = ['hubert-xlarge-ll60k']
 ; no_reuse = False
 scale = standard
 [MODEL]
-type = knn
+type = svm
 ; save = True
+C_val = 10
 [RESAMPLE]
 replace = True
