### 2025-02-10 #7430 @faroit
> Sure, since i develop this framework i try to publish at conferences as often as possible to get more attention. Nonetheless, we constantly add features and change focus, based on our own research interests.
>
> New features since last publications would be, amongst others
>
> * ensemble learning
> * Wav2vec2 finetuning
> * adding speaker idenitification
> * augmentation and segmentation

@felixbur @bagustris, it took me some time to understand the changes, if possible I suggest to add more detailed release notes in the future. Wrt to the submission, I would also like you to add a comparison to the the previous papers and highlight the new additions in a few sentences in the paper 🙏
