### 2025-02-14 #7430 @bagustris
> [@felixbur](https://github.com/felixbur) [@bagustris](https://github.com/bagustris), it took me some time to understand the changes, if possible I suggest to add more detailed release notes in the future. Wrt to the submission, I would also like you to add a comparison to the previous papers and highlight the new additions in a few sentences in the paper 🙏

@faroit We have [CHANGELOG.md](https://github.com/felixbur/nkululeko/blob/main/CHANGELOG.md) for each release and it serves similarly as release notes (in the style of scikit-learn, release history (`whats_new`) is same as changelog).

@felixbur, I think we should add the Nkululeko version for (future) each publication and suggest anyone who uses nkululeko to mention the Nkululeko version in their paper.
