### 2025-02-14 #7430 @felixbur
>anyone who uses nkululeko to mention the Nkululeko version in their paper.
Makes sense.

The problem with our Changelog, as i realized too late, is that i didn't put dates to the versions (i somehow assumed that would be automatic), I'll do this from now on, but agree that makes it hard to keep track what change happened when.
