### 2025-03-09 #7430 @faroit
You added the following to the paper:

> Besides many minor enhancements; ensemble learning, Wav2vec2 model
finetuning, adding automatic speaker identification, extending augmentation and segmentation.

I think this is great but given the new contributions, I do think that now you have to get a bit more specific in the remaining paper about the novel contributions and the software aspect of it to differentiate from the previous publications. This shouldn't stop us from proceeding with the reviews, though but I just wanted to flag this already here.
