+++ test_runs.sh
@@ -1,123 +1,0 @@
-#!/bin/bash
-
-# Display help message
-function Help {
-    echo "Usage: test_runs.sh [options]"
-    echo "Options:"
-    echo "  --Explore: test explore module"
-    echo "  --Nkulu: test basic nkululeko"
-    echo "  --Feat: test selected acoustic features"
-    echo "  --Aug: test augmentation"
-    echo "  --Pred: test prediction"
-    echo "  --Demo: test demo"
-    echo "  --Test: test test module"
-    echo "  --Multi: test multidb"
-    echo "  --Spot: test spotlight"
-    echo "  --all: test all modules"
-    echo "  --help: display this help message"
-}
-
-# test explore module
-function Explore {
-    python -m nkululeko.explore --config tests/exp_emodb_explore_data.ini
-    python -m nkululeko.explore --config tests/exp_emodb_explore_featimportance.ini
-    python -m nkululeko.explore --config tests/exp_emodb_explore_scatter.ini
-    python -m nkululeko.explore --config tests/exp_emodb_explore_features.ini
-    python -m nkululeko.explore --config tests/exp_agedb_explore_data.ini
-}
-# test basic nkululeko
-function Nkulu {
-    python -m nkululeko.nkululeko --config tests/exp_emodb_os_praat_xgb.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_featimport_xgb.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_cnn.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_balancing.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_audmodel_xgb.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_split.ini
-    python -m nkululeko.nkululeko --config tests/exp_ravdess_os_xgb.ini
-    python -m nkululeko.nkululeko --config tests/exp_agedb_class_os_xgb.ini 
-}
-# test features
-function Feat {
-    python -m nkululeko.nkululeko --config tests/exp_emodb_hubert_xgb.ini 
-    python -m nkululeko.nkululeko --config tests/exp_emodb_audmodel_xgb.ini 
-    python -m nkululeko.nkululeko --config tests/exp_emodb_wavlm_xgb.ini 
-    python -m nkululeko.nkululeko --config tests/exp_emodb_whisper_xgb.ini 
-}
-# test augmentation
-function Aug {
-    python -m nkululeko.augment --config tests/exp_emodb_augment_os_xgb.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb-aug_os_xgb.ini
-    python -m nkululeko.augment --config tests/exp_emodb_random_splice_os_xgb.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_rs_os_xgb.ini
-    python -m nkululeko.aug_train --config tests/emodb_aug_train.ini
-}
-# test prediction
-function Pred {
-    python -m nkululeko.predict --config tests/exp_emodb_predict.ini
-    python -m nkululeko.nkululeko --config tests/emodb_demo.ini
-}
-# test demo
-function Demo {
-    python -m nkululeko.nkululeko --config tests/exp_emodb_os_xgb.ini 
-    python -m nkululeko.nkululeko --config tests/exp_emodb_os_svm.ini
-    python -m nkululeko.nkululeko --config tests/exp_emodb_os_knn.ini
[…]
