+++ requirements_wav2vec.txt
@@ -1,6 +1,0 @@
-audiofile
-git+https://github.com/huggingface/datasets.git
-git+https://github.com/huggingface/transformers.git
-jiwer
-torchaudio
-librosa
