+++ nkululeko/reporting/reporter.py
@@ -138,7 +138,7 @@ class Reporter:
             self.util.error(f"unknown metric: {self.metric}")
         return test_result, upper, lower
 
-    def print_probabilities(self):
+    def print_probabilities(self, file_name = None):
         """Print the probabilities per class to a file in the store."""
         if (
             self.util.exp_is_classification()
@@ -168,11 +168,11 @@ class Reporter:
             )
             probas["uncertainty"] = uncertainty
             probas["correct"] = probas.predicted == probas.truth
-            sp = self.util.get_pred_name()
-
+            if file_name is None:
+                file_name = self.util.get_pred_name()+".csv"
             self.probas = probas
-            probas.to_csv(sp)
-            self.util.debug(f"Saved probabilities to {sp}")
+            probas.to_csv(file_name)
+            self.util.debug(f"Saved probabilities to {file_name}")
             plots = Plots()
             ax, caption = plots.plotcatcont(
                 probas, "correct", "uncertainty", "uncertainty", "correct"
@@ -182,7 +182,7 @@ class Reporter:
                 caption,
                 "Uncertainty",
                 "uncertainty_samples",
-                self.util.get_exp_name(),
+                file_name,
             )
 
     def set_id(self, run, epoch):
@@ -368,7 +368,7 @@ class Reporter:
 
         res_dir = self.util.get_path("res_dir")
         rpt = (
-            f"Best score at epoch: {epoch}, UAR: {uar_str}"
+            f"Confusion matrix result for epoch: {epoch}, UAR: {uar_str}"
             + f", (+-{up_str}/{low_str}), ACC: {acc_str}"
         )
         # print(rpt)
@@ -392,13 +392,16 @@ class Reporter:
             text_file.write(result_str)
         self.util.debug(result_str)
 
-    def print_results(self, epoch=None):
+    def print_results(self, epoch=None, file_name = None):
         if epoch is None:
             epoch = self.epoch
         """Print all evaluation values to text file."""
         res_dir = self.util.get_path("res_dir")
-        fig_dir = self.util.get_path("fig_dir")
-        file_name = f"{res_dir}{self.util.get_exp_name()}_{epoch}{self.filenameadd}.txt"
+        if file_name is None:
+            file_name = f"{res_dir}{self.util.get_exp_name()}_{epoch}{self.filenameadd}.txt"
+        else:
+            self.util.debug(f"####->{file_name}<-####")
+            file_name = f"{res_dir}{file_name}{self.filenameadd}.txt"
         if self.util.exp_is_classification():
             labels = glob_conf.labels
             try:
@@ -427,25 +430,6 @@ class Reporter:
                 f1_per_class = (
                     f"result per class (F1 score): {c_ress} from epoch: {epoch}"
                 )
-                # the following auc is buggy, preds should be probabilities
-                # if len(np.unique(self.truths)) == 2:
-                #     fpr, tpr, _ = roc_curve(self.truths, self.preds)
-                #     auc_score = auc(fpr, tpr)
-                #     plot_path = f"{fig_dir}{self.util.get_exp_name()}_{epoch}{self.filenameadd}_roc.{self.format}"
-                #     plt.figure()
-                #     display = RocCurveDisplay(
-                #         fpr=fpr,
-                #         tpr=tpr,
-                #         roc_auc=auc_score,
-                #         estimator_name=f"{self.model_type} estimator",
-                #     )
-                #     display.plot(ax=None)
-                #     plt.savefig(plot_path)
-                #     plt.close()
-                #     self.util.debug(f"Saved ROC curve to {plot_path}")
-                #     pauc_score = roc_auc_score(self.truths, self.preds, max_fpr=0.1)
[…]
