+++ nkululeko/runmanager.py
@@ -22,7 +22,8 @@ class Runmanager:
     )  # The dataframes
     reports = []
 
-    def __init__(self, df_train, df_test, feats_train, feats_test):
+    def __init__(self, df_train, df_test, feats_train,
+                 feats_test, dev_x=None, dev_y=None):
         """Constructor setting up the dataframes.
 
         Args:
@@ -38,11 +39,10 @@ class Runmanager:
             feats_train,
             feats_test,
         )
+        self.df_dev, self.feats_dev = dev_x, dev_y
         self.util = Util("runmanager")
         self.target = glob_conf.config["DATA"]["target"]
-        # intialize a new model
-        # model_type = glob_conf.config['MODEL']['type']
-        # self._select_model(model_type)
+        self.split3 = eval(self.util.config_val("EXP", "traindevtest", "False"))
 
     def do_runs(self):
         """Start the runs."""
@@ -55,14 +55,34 @@ class Runmanager:
             )
             # set the run index as global variable for reporting
             self.util.set_config_val("EXP", "run", run)
-            self.modelrunner = Modelrunner(
-                self.df_train,
-                self.df_test,
-                self.feats_train,
-                self.feats_test,
-                run,
+            if self.df_dev is not None:
+                self.modelrunner = Modelrunner(
+                    self.df_train,
+                    self.df_dev,
+                    self.feats_train,
+                    self.feats_dev,
+                    run,
+                )
+                self.reports, last_epoch = self.modelrunner.do_epochs()
+            else:
+                self.modelrunner = Modelrunner(
+                    self.df_train,
+                    self.df_test,
+                    self.feats_train,
+                    self.feats_test,
+                    run,
+                )
+                self.reports, last_epoch = self.modelrunner.do_epochs()
+
+            last_report = self.reports[-1]
+            plot_name_suggest = self.util.get_exp_name()
+            plot_name = (
+                self.util.config_val("PLOT", "name", plot_name_suggest)
+                + f"_last_{last_report.run}_{last_report.epoch:03d}"
             )
-            self.reports, last_epoch = self.modelrunner.do_epochs()
+            # finally, print out the numbers for this run
+            self.print_report(last_report, plot_name)
+
             # wrap up the run
             plot_anim_progression = self.util.config_val("PLOT", "anim_progression", 0)
             if plot_anim_progression:
@@ -82,7 +102,6 @@ class Runmanager:
                 # possibly this value has not been set
                 epoch_num = 1
             if epoch_num > 1 and plot_epoch_progression:
-                plot_name_suggest = self.util.get_exp_name()
                 plot_name = (
                     self.util.config_val("PLOT", "name", plot_name_suggest)
                     + "_epoch_progression"
@@ -91,25 +110,24 @@ class Runmanager:
                 self.reports[-1].plot_epoch_progression(self.reports, plot_name)
             # remember the best run
             best_report = self.get_best_result(self.reports)
-            plot_best_model = self.util.config_val("PLOT", "best_model", False)
-
-            if epoch_num > 1 and plot_best_model:
-                plot_name_suggest = self.util.get_exp_name()
-                plot_name = (
-                    self.util.config_val("PLOT", "name", plot_name_suggest)
-                    + f"_BEST_{best_report.run}_{best_report.epoch:03d}_BEST_cnf"
-                )
-                self.util.debug(
-                    f"best result with run {best_report.run} and epoch"
-                    f" {best_report.epoch}:"
-                    f" {best_report.result.get_test_result()}"
-                )
-                self.print_model(best_report, plot_name)
+            plot_name = (
+                self.util.config_val("PLOT", "name", plot_name_suggest)
[…]
