+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.93.16"
+VERSION="0.94.0"
 SAMPLING_RATE = 16000
