+++ requirements.txt
@@ -7,7 +7,7 @@ audmetric
 audiomentations==0.31.0
 audonnx
 confidence-intervals
-#cylimiter
+cylimiter
 audtorch
 datasets
 imageio
