+++ paper/paper.md
@@ -174,6 +174,126 @@ from epoch: 0
 DEBUG: experiment: Done, used 7.439 seconds
 DONE
 ```
+
+# What has been added since the last publication
+
+Besides many small changes, mainly three big additions extended nkululeko's functionality since the last published publications. We introduce them in the next subsections.
+
+
+## Finetune transformer models
+
+With [nkululeko](https://github.com/felixbur/nkululeko) since version 0.85.0 you can finetune a transformer model with [huggingface](https://huggingface.co/docs/transformers/training) (and even  [publish it there if you like](https://huggingface.co/docs/hub/models-uploading)).
+
+If you like to have your model published, set:
+```
+[MODEL]
+push_to_hub = True
+```
+
+Finetuning in this context means to train the (pre-trained) transformer layers with your new training data labels, as opposed to only using the last layer as embeddings.
+
+The only thing you need to do is to set your MODEL type to *finetune*:
+```
+[FEATS]
+type = []
+[MODEL]
+type = finetune
+```
+The acoustic features can/should be empty, because the transformer model starts with CNN layers to model the acoustics frame-wise.  The frames are then getting pooled by the model for the whole utterance (max. duration the first 8 seconds, the rest is ignored).
+
+The default base model is the one from [facebook](https://huggingface.co/facebook/wav2vec2-large-robust-ft-swbd-300h), but you can specify a different one like this:
+```
+[MODEL]
+type = finetune
+pretrained_model = microsoft/wavlm-base
+
+duration = 10.5
+```
+The parameter *max_duration* is also optional (default=8) and means the maximum duration of your samples / segments (in seconds) that will be used, starting from 0. The rest is disregarded. 
+
+You can use the usual deep learning parameters:
+```
+[MODEL]
+learning_rate = .001
+batch_size = 16
+device = cuda:3
+measure = mse
+loss = mse
+```
+but all of them have defaults.
+
+The loss function is fixed to 
+* weighted cross entropy for classification
+* concordance correlation coefficient for regression
+
+The resulting best model and the huggingface logs (which can be read by [tensorboard](https://www.tensorflow.org/tensorboard)) are stored in the project folder.
+
+
+## Ensemble classification
+
+With [nkululeko](https://github.com/felixbur/nkululeko) since version 0.88.0  you can combine experiment results and report on the outcome, by using the **ensemble** module.
+
+For example, you would like to know if the combination of expert features and learned embeddings works better than one of those. You could then do
+```
+python -m nkululeko.ensemble \
+--method max_class \
+tests/exp_emodb_praat_xgb.ini \
+tests/exp_emodb_ast_xgb.ini \
+tests/exp_emodb_wav2vec_xgb.in
+```
+(all in one line)
+and would then get the results for a majority voting of the three results for Praat, AST and Wav2vec2 features.
+
+Other  methods are *mean*, *max*, *sum*, *max_class*, *uncertainty_threshold*, *uncertainty_weighted*, *confidence_weighted*:
+* **majority_voting**: The modality function for classification: predict the category that most classifiers agree on.
+* **mean**: For classification: compute the arithmetic mean of probabilities from all predictors for each labels, use highest probability to infer the label.
+* **max**: For classification: use the maximum value of probabilities from all predictors for each labels, use highest probability to infer the label.
+* **sum**: For classification: use the sum of probabilities from all predictors for each labels, use highest probability to infer the label.
+* **max_class**: For classification: compare the highest probabilities of all models across classes (instead of same class as in max_ensemble) and return the highest probability and the class
[…]
