### 2025-04-08 #7430 @felixbur
> [@felixbur](https://github.com/felixbur) would you be able to turn the changelog into text for the paper so that its more clear how this publication differs from previous ones?

ok, @faroit , I added a section *What has been added since the last publication* , i hope the paper is not too long now.
