### 2025-04-14 #8049 @editorialbot
[REVIEW]: Nkululeko 1.0: A Python package to predict speaker characteristics with a high-level interface
**Submitting author:** @felixbur (<a href="http://orcid.org/0000-0002-2689-0545">Felix Burkhardt</a>)
**Repository:** https://github.com/felixbur/nkululeko
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
