### 2025-04-28 1526657
Update README.md
removed slack link
