+++ paper/paper.md
@@ -20,9 +20,9 @@ affiliations:
    index: 1
  - name: TU Berlin, Germany
    index: 2
- - name: National Institute of Advanced Industrial Science and Technology (AIST), Japan
+ - name: Nara Institute of Science and Technology (NAIST), Japan
    index: 3
-date: 24 December 2024
+date: 15 April 2025  # Date of JOSS Review, submitted 2024-12-24
 bibliography: paper.bib
 
 ---
@@ -196,7 +196,8 @@ type = finetune
 The acoustic features can/should be empty, because the transformer model starts with CNN layers to model the acoustics frame-wise.  The frames are then getting pooled by the model for the whole utterance.
 
 The default base model is the one from [facebook](https://huggingface.co/facebook/wav2vec2-large-robust-ft-swbd-300h), but you can specify a different one like this:
-```
+
+```ini
 [MODEL]
 type = finetune
 pretrained_model = microsoft/wavlm-base
@@ -206,7 +207,8 @@ duration = 10.5
 The parameter *max_duration* is also optional (default=8) and means the maximum duration of your samples / segments (in seconds) that will be used, starting from 0. The rest is disregarded. 
 
 You can use the usual deep learning parameters:
-```
+
+```ini
 [MODEL]
 learning_rate = .001
 batch_size = 16
@@ -224,7 +226,8 @@ The loss function is fixed to
 The resulting best model and the huggingface logs (which can be read by [tensorboard](https://www.tensorflow.org/tensorboard)) are stored in the project folder.
 
 If you like to have your model published, set:
-```
+
+```ini
 [MODEL]
 push_to_hub = True
 ```
@@ -232,10 +235,11 @@ push_to_hub = True
 
 ## Ensemble classification
 
-With [nkululeko](https://github.com/felixbur/nkululeko) since version 0.88.0  you can combine experiment results and report on the outcome, by using the **ensemble** module.
+With [nkululeko](https://github.com/felixbur/nkululeko) since version 0.88.0 you can combine experiment results and report on the outcome, by using the **ensemble** module.
 
-For example, you would like to know if the combination of expert features and learned embeddings works better than one of those. You could then do
-```
+For example, you would like to know if the combination of expert features and learned embeddings works better than one of those. You could then do: 
+
+```bash
 python -m nkululeko.ensemble \
 --method max_class \
 tests/exp_emodb_praat_xgb.ini \
@@ -248,9 +252,9 @@ and would then get the results for a majority voting of the three results for Pr
 Other  methods to combine the different predictors, are *mean*, *max*, *sum*, *max_class*, *uncertainty_threshold*, *uncertainty_weighted*, *confidence_weighted*:
 
 * **majority_voting**: The modality function for classification: predict the category that most classifiers agree on.
-* **mean**: For classification: compute the arithmetic mean of probabilities from all predictors for each labels, use highest probability to infer the label.
-* **max**: For classification: use the maximum value of probabilities from all predictors for each labels, use highest probability to infer the label.
-* **sum**: For classification: use the sum of probabilities from all predictors for each labels, use highest probability to infer the label.
+* **mean**: For classification: compute the arithmetic mean of probabilities from all predictors for each label, use the highest probability to infer the label.
+* **max**: For classification: use the maximum value of probabilities from all predictors for each label, use the highest probability to infer the label.
+* **sum**: For classification: use the sum of probabilities from all predictors for each label, use the highest probability to infer the label.
 * **max_class**: For classification: compare the highest probabilities of all models across classes (instead of same class as in max_ensemble) and return the highest probability and the class
[…]
