### 2025-05-05 felixbur/nkululeko#188 @isdanni
[FOSS] Nkululeko 1.0: A Python package to predict speaker characteristics with a high-level interface
Hi @felixbur, thanks for submitting your project to FOSS! I'm one of the reviewers assigned to your submission [Nkululeko 1.0: A Python package to predict speaker characteristics with a high-level interface](https://github.com/openjournals/joss-reviews/issues/8049#issuecomment-2838871113). This thread will be used for my review and feedback.

## Documentation

I highly suggest using one of the README.md templates to highlight your project features and move detailed plots to the documentation website. For example, choosing one of the simple templates from [makeareadme.com](https://www.makeareadme.com/).
- Move the **Installation** and **Documentation** sections to the top of the README, before any visualizations or plots.
- Make optional dependencies clearer. For example:
  - Use `pip install nkululeko[spotlight]` for Spotlight support
  - Or specify versions directly in [`requirements.txt`](https://github.com/felixbur/nkululeko/blob/main/requirements.txt), e.g., `audformat==1.3.1`
- Add a brief “**Who is this for?**” section under the Project title, such as:
  *"Nkululeko is for speech researchers and ML practitioners focused on speaker traits and emotion recognition."*
- Move detailed plots (e.g., **Confusion matrix**, **t-SNE**, **Bias checking**) to the docs site to keep the README simple and focused on setup and usage.
