### 2025-05-05 felixbur/nkululeko#188 @isdanni
## Paper

The draft paper successfully identifies the statement of the need and includes a clear summary. There are a few suggestions:
- Minor typos:
    - In line 265, there appears to be a duplicated word: "and and bias analysis".
    - In line 190: "attachched".
 - References:
     - In line 283, "classificatio" missing "n" in the end.
 - To emphasize package novelty compared to the existing solutions: highly unique contributions in the summary section. For example, mention that it has config-based experimentation or a user-friendly interface.
 - If possible, consider adding a simple diagram illustrating Nkululeko’s internal workflow. For example, data input(in which format) → feature extraction → ... → output.
