### 2025-05-06 felixbur/nkululeko#188 @bagustris
@felixbur

I'll work on the docs-related issues:

- Move detailed plots (e.g., Confusion matrix, t-SNE, Bias checking) to the docs site to keep the README simple and focused on setup and usage.
