### 2025-05-07 c7b5c2e
update README and docs
