+++ README.md
@@ -1,89 +1,14 @@
 
-- [Overview](#overview)
-  - [Confusion matrix](#confusion-matrix)
-  - [Epoch progression](#epoch-progression)
-  - [Feature importance](#feature-importance)
-  - [Feature distribution](#feature-distribution)
-  - [t-SNE plots](#t-sne-plots)
-  - [Data distribution](#data-distribution)
-  - [Bias checking](#bias-checking)
-  - [Uncertainty](#uncertainty)
-- [Documentation](#documentation)
-- [Installation](#installation)
-- [Usage](#usage)
-  - [ini-file values](#ini-file-values)
-  - [Hello World example](#hello-world-example)
-  - [Features](#features)
-- [License](#license)
-- [Contributing](#contributing)
-- [Citing](#citing)
+# Nkululeko
 
- 
-## Overview
-A project to detect speaker characteristics by machine learning experiments with a high-level interface.
+Nkululeko is a project to detect speaker characteristics by machine learning experiments with a high-level interface.
 
 The idea is to have a framework (based on e.g. sklearn and torch) that can be used to rapidly and automatically analyse audio data and explore machine learning models based on that data. 
 
-* NEW with nkululeko: [Ensemble learning](http://blog.syntheticspeech.de/2024/06/25/nkululeko-ensemble-classifiers-with-late-fusion/)
-* NEW: [Finetune transformer-models](http://blog.syntheticspeech.de/2024/05/29/nkululeko-how-to-finetune-a-transformer-model/)
-* The latest features can be seen in [the ini-file](./ini_file.md) options that are used to control Nkululeko
-* Below is a [Hello World example](#helloworld) that should set you up fastly, also on [Google Colab](https://colab.research.google.com/drive/1GYNBd5cdZQ1QC3Jm58qoeMaJg3UuPhjw?usp=sharing#scrollTo=4G_SjuF9xeQf), and [with Kaggle](https://www.kaggle.com/felixburk/nkululeko-hello-world-example)
-* [Thanks to deepwiki, here's an analysis of the source code](https://deepwiki.com/felixbur/nkululeko)
-* [Here's a blog post on how to set up nkululeko on your computer.](http://blog.syntheticspeech.de/2021/08/30/how-to-set-up-your-first-nkululeko-project/)
-* [Here's a slide presentation about nkululeko](docs/nkululeko.pdf)
-* [Here's a video presentation about nkululeko](https://www.youtube.com/playlist?list=PLRceVavtxLg0y2jiLmpnUfiMtfvkK912D)
-* [Here's the 2022 LREC article on nkululeko](http://felix.syntheticspeech.de/publications/Nkululeko_LREC.pdf)
+Some abilities that Nkululeko provides: combines acoustic features and machine learning models (including feature selection and features concatenation); performs data exploration, selection and visualizes the results; finetuning; ensemble learning models; soft labeling (predicting labels with pre-trained model); and inference the model on a test set.
 
-Here are some examples of typical output:
-
-### Confusion matrix
-Per default, Nkululeko displays results as a confusion matrix using binning with regression.
-
-<img src="meta/images/conf_mat.png" width="500px"/>
-
-### Epoch progression
-The point when overfitting starts can sometimes be seen by looking at the results per epoch:
-
-<img src="meta/images/epoch_progression.png" width="500px"/>
-
-### Feature importance
-Using the *explore* interface, Nkululeko analyses the importance of acoustic features:
- 
-<img src="meta/images/feat_importance.png" width="500px"/>
-
-### Feature distribution
-And can show the distribution of specific features per category:
-
[…]
