### 2025-05-09 199d4a4
simplify readme
