### 2025-05-12 7ec217b
Update installation process, add version info, and create basic unit test
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
