+++ docs/source/installation.rst
@@ -1,42 +1,70 @@
 Installation
 ----------------
 
-The easiest and safest way to try Nkululeko is using virtualenv, either using venv or mini (conda).  Here an example using Conda. ::
+Nkululeko requires Python 3.9 or higher. The easiest and safest way to install Nkululeko is using a virtual environment, either with venv or conda.
 
+Using Conda
+^^^^^^^^^^^
 
-    $ conda create -n nkululeko python=3.8  
-    $ conda activate nkululeko  
-    $ pip install nkululeko  
+.. code-block:: bash
 
-And for venv: ::
-
-    $ python3 -m venv nkululeko  
-    $ source nkululeko/bin/activate  
+    $ conda create -n nkululeko python=3.9
+    $ conda activate nkululeko
     $ pip install nkululeko
 
-For development version, you install directly from the source: ::
+Using venv
+^^^^^^^^^^
 
-    # clone the repository  
-    $ git clone https://github.com/felixbur/nkululeko.git  
-    $ cd nkululeko  
-    # install in editable mode  
-    $ pip install -e .  
-    # or add the path to your PYTHONPATH
-    $ export PYTHONPATH=$PYTHONPATH:/path/to/nkululeko
-    # You can also work from Nkululeko root directory to enable importing 
-    # Nkululeko without installing it
+.. code-block:: bash
+
+    $ python3 -m venv nkululeko
+    $ source nkululeko/bin/activate  # On Windows: nkululeko\Scripts\activate
+    $ pip install nkululeko
+
+Current version: ``0.94.1``
+
+Optional Dependencies
+^^^^^^^^^^^^^^^^^^^^
+
+Some functionalities require additional packages:
+
+.. code-block:: bash
+
+    # For SQUIM model
+    $ pip install --pre torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/nightly/cpu
+
+    # For spotlight adapter
+    $ pip install renumics-spotlight sliceguard
+
+    # For CPU-only PyTorch installation
+    $ pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
+
+Development Installation
+^^^^^^^^^^^^^^^^^^^^^^^
+
+For the development version, install directly from the source:
+
+.. code-block:: bash
+
+    # Clone the repository
+    $ git clone https://github.com/felixbur/nkululeko.git
     $ cd nkululeko
-    # pwd
-    # /path/to/nkululeko
+    # Install in editable mode
+    $ pip install -e .
 
-Check the installation: ::
+Verify Installation
+^^^^^^^^^^^^^^^^^^
 
-    # via pip
-    pip list | grep nkululeko
-    # inside python
+Check that Nkululeko is installed correctly:
+
+.. code-block:: bash
+
+    # Via pip
+    $ pip list | grep nkululeko
+
+    # Inside Python
     >>> import nkululeko
     >>> nkululeko.__version__
-    # to differentiate between the installed and the development version
-    >>> nkululeko.__file__
+    >>> nkululeko.__file__  # Shows installation path
 
-If you see the version of Nkululeko from one of the first two command above, e.g. ``0.55.0``, you are ready to go.
+If you see the version of Nkululeko (e.g., ``0.94.1``), you are ready to go.
