+++ pyproject.toml
@@ -4,3 +4,7 @@ requires = [
   "wheel"
 ]
 build-backend = "setuptools.build_meta"
+
+[tool.pytest]
+testpaths = ["tests"]
+python_files = "test_*.py"
