### 2025-05-12 4450f01
Update optional dependencies with extras_require and clearer documentation
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
