+++ docs/source/installation.rst
@@ -26,18 +26,36 @@ Current version: ``0.94.1``
 Optional Dependencies
 ^^^^^^^^^^^^^^^^^^^^
 
-Some functionalities require additional packages:
+Nkululeko supports optional dependencies through extras. You can install them using the following syntax:
 
 .. code-block:: bash
 
-    # For SQUIM model
-    $ pip install --pre torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/nightly/cpu
+    # Install with PyTorch support (default PyTorch)
+    $ pip install nkululeko[torch]
 
-    # For spotlight adapter
-    $ pip install renumics-spotlight sliceguard
-
-    # For CPU-only PyTorch installation
+    # Install with CPU-only PyTorch
+    $ pip install nkululeko[torch-cpu]
+    # Or manually:
     $ pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
+
+    # Install with Spotlight support
+    $ pip install nkululeko[spotlight]
+
+    # Install with TensorFlow support
+    $ pip install nkululeko[tensorflow]
+
+    # Install all optional dependencies
+    $ pip install nkululeko[all]
+
+For specific model support:
+
+.. code-block:: bash
+
+    # For SQUIM model (requires nightly PyTorch)
+    $ pip install nkululeko[torch-nightly]
+    # Or manually:
+    $ pip uninstall -y torch torchvision torchaudio
+    $ pip install --pre torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/nightly/cpu
 
 Development Installation
 ^^^^^^^^^^^^^^^^^^^^^^^
@@ -49,8 +67,10 @@ For the development version, install directly from the source:
     # Clone the repository
     $ git clone https://github.com/felixbur/nkululeko.git
     $ cd nkululeko
-    # Install in editable mode
-    $ pip install -e .
+    # Install in editable mode with all dependencies
+    $ pip install -e ".[all]"
+    # Or with specific extras
+    $ pip install -e ".[torch,spotlight]"
 
 Verify Installation
 ^^^^^^^^^^^^^^^^^^
