### 2025-05-12 670cbb0
Update requirements.txt and setup.cfg with version specifications and extras
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
