+++ requirements.txt
@@ -1,36 +1,45 @@
-audeer
-audobject
-audformat
-audinterface
-audiofile
-audmetric
+# Core dependencies
+audeer>=1.0.0
+audformat>=1.3.1
+audinterface>=1.0.0
+audiofile>=1.0.0
 audiomentations==0.31.0
-audonnx
-confidence-intervals
-cylimiter
-audtorch
-datasets
-imageio
-imblearn
-matplotlib
-numpy
-opensmile
-pandas
-praat-parselmouth
-pylatex
-renumics-spotlight 
-sliceguard
-scikit_learn
-scipy
-seaborn
-shap
-sounddevice
-splitutils
-# tensorflow
-# tensorflow_hub
-torch
-torchaudio
-torchvision
-transformers
-umap-learn
-xgboost
+audmetric>=1.0.0
+audonnx>=1.0.0
+confidence-intervals>=0.1.0
+datasets>=2.0.0
+imageio>=2.0.0
+matplotlib>=3.0.0
+numpy>=1.20.0
+opensmile>=2.0.0
+pandas>=1.0.0
+praat-parselmouth>=0.4.0
+scikit_learn>=1.0.0
+scipy>=1.0.0
+seaborn>=0.11.0
+sounddevice>=0.4.0
+transformers>=4.0.0
+umap-learn>=0.5.0
+xgboost>=1.0.0
+pylatex>=1.0.0
+
+# Optional dependencies - PyTorch
+torch>=1.0.0
+torchaudio>=0.10.0
+torchvision>=0.10.0
+
+# Optional dependencies - Spotlight
+renumics-spotlight>=0.1.0
+sliceguard>=0.1.0
+
+# Optional dependencies - TensorFlow (commented by default)
+# tensorflow>=2.0.0
+# tensorflow_hub>=0.12.0
+
+# Other optional dependencies
+audobject>=1.0.0
+cylimiter>=0.0.1
+audtorch>=0.0.1
+imblearn>=0.0.0
+shap>=0.40.0
+splitutils>=0.0.1
