+++ setup.cfg
@@ -17,32 +17,61 @@ classifiers =
 [options]
 packages = find:
 install_requires =
-    audeer
-    audformat
-    audinterface
-    audiofile
-    audiomentations
-    audmetric
-    audonnx
-    confidence_intervals
-    datasets
-    imageio
-    matplotlib
-    numpy
-    opensmile
-    pandas
-    praat-parselmouth
-    scikit_learn
-    scipy
-    seaborn
-    sounddevice
-    # tensorflow
-    # tensorflow_hub
-    torch
-    torchvision
-    transformers
-    umap-learn
-    xgboost
-    pylatex
+    audeer>=1.0.0
+    audformat>=1.3.1
+    audinterface>=1.0.0
+    audiofile>=1.0.0
+    audiomentations==0.31.0
+    audmetric>=1.0.0
+    audonnx>=1.0.0
+    confidence_intervals>=0.1.0
+    datasets>=2.0.0
+    imageio>=2.0.0
+    matplotlib>=3.0.0
+    numpy>=1.20.0
+    opensmile>=2.0.0
+    pandas>=1.0.0
+    praat-parselmouth>=0.4.0
+    scikit_learn>=1.0.0
+    scipy>=1.0.0
+    seaborn>=0.11.0
+    sounddevice>=0.4.0
+    transformers>=4.0.0
+    umap-learn>=0.5.0
+    xgboost>=1.0.0
+    pylatex>=1.0.0
 python_requires = >=3.9
 include_package_data = False
+
+[options.extras_require]
+torch =
+    torch>=1.0.0
+    torchvision>=0.10.0
+    torchaudio>=0.10.0
+torch-cpu =
+    torch>=1.0.0
+    torchvision>=0.10.0
+    torchaudio>=0.10.0
+torch-nightly =
+    torch
+    torchvision
+    torchaudio
+spotlight =
+    renumics-spotlight>=0.1.0
+    sliceguard>=0.1.0
+tensorflow =
+    tensorflow>=2.0.0
+    tensorflow_hub>=0.12.0
+all =
+    torch>=1.0.0
+    torchvision>=0.10.0
+    torchaudio>=0.10.0
+    renumics-spotlight>=0.1.0
+    sliceguard>=0.1.0
+    tensorflow>=2.0.0
+    tensorflow_hub>=0.12.0
+    shap>=0.40.0
+    imblearn>=0.0.0
+    cylimiter>=0.0.1
+    audtorch>=0.0.1
+    splitutils>=0.0.1
