### 2025-05-12 36db40e
Fix circular import in setup.py
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
