### 2025-05-12 7cf95a3
Fix audonnx version requirement to match available packages
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
