+++ requirements.txt
@@ -5,7 +5,7 @@ audinterface>=1.0.0
 audiofile>=1.0.0
 audiomentations==0.31.0
 audmetric>=1.0.0
-audonnx>=1.0.0
+audonnx>=0.7.0
 confidence-intervals>=0.1.0
 datasets>=2.0.0
 imageio>=2.0.0
