### 2025-05-12 78d0a1f
Fix confidence-intervals version requirement to match available packages
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
