+++ setup.cfg
@@ -24,7 +24,7 @@ install_requires =
     audiomentations==0.31.0
     audmetric>=1.0.0
     audonnx>=0.7.0
-    confidence_intervals>=0.1.0
+    confidence-intervals>=0.0.2
     datasets>=2.0.0
     imageio>=2.0.0
     matplotlib>=3.0.0
