### 2025-05-12 47007ba
Update unit tests to handle optional dependencies gracefully
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
