### 2025-05-12 9603d36
Add installation test script and GitHub Actions workflows for Python 3.10, 3.11, and 3.12
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
