+++ .github/workflows/test-install-python310.yml
@@ -1,0 +1,20 @@
+name: Test Installation Python 3.10
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  test-install:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
+    - name: Test installation
+      run: |
+        python scripts/test_install.py
