+++ .github/workflows/test-install-python311.yml
@@ -1,0 +1,20 @@
+name: Test Installation Python 3.11
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  test-install:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Test installation
+      run: |
+        python scripts/test_install.py
