+++ .github/workflows/basic_tests_for_aud_csv.yaml
@@ -10,7 +10,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
-        python-version: '3.9'
+        python-version: '3.10'
         # cache: 'pip' # caching pip dependencies, causing error due to too large
     - name: Install dependencies
       run: |
