+++ requirements.txt
@@ -37,9 +37,9 @@ sliceguard>=0.0.35
 # tensorflow_hub>=0.12.0
 
 # Other optional dependencies
-audobject>=1.0.0
-cylimiter>=0.0.1
-audtorch>=0.0.1
+# audobject>=1.0.0
+# cylimiter>=0.0.1
+# audtorch>=0.0.1
 imblearn>=0.0.0
 shap>=0.40.0
 splitutils>=0.0.1
