### 2025-05-12 6668c52
Fix GitHub Actions workflow by updating actions versions and pinning audobject version
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
