+++ .github/workflows/basic_tests_for_aud_csv.yaml
@@ -6,16 +6,17 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v5
       with:
-        python-version: '3.9'
-        # cache: 'pip' # caching pip dependencies, causing error due to too large
+        python-version: '3.10'
+        cache: 'pip'
+        cache-dependency-path: 'requirements.txt'
     - name: Install dependencies
       run: |
         sudo apt-get -y install sox libportaudio2
-        # python3 -m pip install --upgrade pip
+        python3 -m pip install --upgrade pip
         pip install -r requirements.txt
         # pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
         pip install --editable .
