### 2025-05-12 bc627e6
Update requirements.txt
disable renumics on default
