### 2025-05-12 c8d62c1
Update README.md
create venv before install nkululeko
