+++ tests/test_basic.py
@@ -1,52 +1,0 @@
-import unittest
-import os
-import sys
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-import nkululeko
-from nkululeko.constants import VERSION
-
-
-class TestBasic(unittest.TestCase):
-    """Basic tests for nkululeko package."""
-
-    def test_version(self):
-        """Test that version is correctly defined."""
-        self.assertEqual(nkululeko.__version__, VERSION)
-        self.assertTrue(isinstance(VERSION, str))
-        self.assertTrue(len(VERSION) > 0)
-
-    def test_import(self):
-        """Test that main modules can be imported."""
-        import nkululeko.nkululeko
-        import nkululeko.explore
-        import nkululeko.augment
-        import nkululeko.demo
-        import nkululeko.predict
-        
-        try:
-            import nkululeko.resample
-        except ImportError:
-            print("Skipping resample module import (requires torchaudio)")
-            
-        try:
-            import nkululeko.segment
-        except ImportError:
-            print("Skipping segment module import (may require optional dependencies)")
-            
-        try:
-            import nkululeko.test
-        except ImportError:
-            print("Skipping test module import (may require optional dependencies)")
-            
-        try:
-            import nkululeko.ensemble
-        except ImportError:
-            print("Skipping ensemble module import (may require optional dependencies)")
-            
-        self.assertTrue(True)  # If imports succeed, test passes
-
-
-if __name__ == '__main__':
-    unittest.main()
