### 2025-05-13 fe1b978
update readme, test install
