+++ run_test2.sh
@@ -1,195 +1,0 @@
-#!/bin/bash
-
-# save output to log.out
-exec > >(tee -i log.out)
-
-# save error to log.err
-exec 2> >(tee -i log.err >&2)
-
-
-# Display help message
-function Help {
-    echo "Usage: test_runs.sh [options]"
-    echo "Example: ./test_runs.sh nkululeko"
-    echo "Options:"
-    echo "  nkululeko: test basic nkululeko"
-    echo "  augment: test augmentation"
-    echo "  predict: test prediction"
-    echo "  demo: test demo"
-    echo "  test: test test module"
-    echo "  multidb: test multidb"
-    echo "  explore: test explore module (must be run last)"
-    echo "  all: test all modules"
-    echo "  -spotlight: test all modules except spotlight (useful in SSH)"
-    echo "  -overwrite: remove (old) results directory and create if not exist"
-    echo "  --help: display this help message"
-}
-
-# rm results dir if argument is "nkululeko" or "all"
-# TODO: move root to /tmp so no need to do this
-if [ "$1" == "nkululeko" ] || [ "$1" == "all" ];  [ "$1" == "-spotlight" ]; then 
-    # add overwrite argument
-    if [ "$2" == "-overwrite" ]; then
-        echo "Removing (old) results directory and create if not exist"
-        rm -rf tests/results/*
-        mkdir -p tests/results
-    fi
-fi
-
-# Run a test and check for errors
-function RunTest {
-    "$@"
-    if [ $? -ne 0 ]; then
-        echo "Error: Test failed - $@"
-        return 1   # exit after error message
-    # else
-    #     return 0   # continue after error message
-    fi
-}
-
-# resample before performing other tests
-resample_ini_files=(
-    exp_polish_gmm.ini
-)
-# test basic nkululeko
-nkululeko_ini_files=(
-    exp_emodb_os_praat_xgb.ini
-    exp_emodb_limit_size.ini
-    exp_emodb_featimport_xgb.ini
-    exp_emodb_cnn.ini
-    exp_emodb_balancing.ini
-    exp_emodb_split.ini
-    exp_ravdess_os_xgb.ini
-    exp_agedb_class_os_xgb.ini
-    exp_emodb_hubert_xgb.ini
-    exp_emodb_wavlm_xgb.ini
-    exp_emodb_whisper_xgb.ini
-    emodb_demo.ini
-    exp_emodb_os_xgb_test.ini
-    exp_emodb_wav2vec2_test.ini
-    exp_emodb_os_xgb.ini
-    exp_emodb_os_svm.ini
-    exp_emodb_os_knn.ini
-    exp_emodb_os_mlp.ini
-    exp_agedb_os_xgr.ini
-    exp_agedb_os_mlp.ini
-    exp_polish_gmm.ini
-)
-
-# test augmentation
-augment_ini_files=(
-    exp_emodb_augment_os_xgb.ini
-    exp_emodb-aug_os_xgb.ini
-    exp_emodb_random_splice_os_xgb.ini
-    exp_emodb_rs_os_xgb.ini
-    emodb_aug_train.ini
-)
-
-# test prediction
-predict_ini_files=(
-    exp_emodb_predict.ini
-) 
-# test demo
-demo_ini_files=(
-    exp_emodb_os_xgb.ini
-    exp_emodb_os_svm.ini
-    exp_emodb_os_knn.ini
-    exp_emodb_os_mlp.ini
-    exp_agedb_os_xgr.ini
-    exp_agedb_os_mlp.ini
-)
-
-# test test module
-test_ini_files=(
-    exp_emodb_os_xgb_test.ini
-    exp_emodb_wav2vec2_test.ini
-)
-
[…]
