+++ tests/exp_androids_resample.ini
@@ -1,21 +1,0 @@
-[EXP]
-root = ./tests/results/
-name = exp_androids_explore
-runs = 1
-epochs = 1
-save = True
-[DATA]
-databases = ['data']
-data = ./data/androids/androids.csv
-data.type = csv
-data.absolute_path = False
-data.split_strategy = random
-target = depression
-labels = ['depressed', 'control']
-filter = [['task', 'interview']]
-[FEATS]
-type = ['os']
-scale = standard
-[RESAMPLE]
-replace = True
-target = androids_resampled.csv

[…]
