### 2025-05-13 244d12b
rename old tests to examples
