+++ examples/exp_emodb_whisper_xgb.ini
@@ -1,0 +1,18 @@
+[EXP]
+root = ./tests/results/
+name = exp_emodb_feats
+runs = 1
+epochs = 1
+save = True
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
+labels = ['angry', 'happy']
+target = emotion
+[FEATS]
+type = ['whisper']
+[MODEL]
+type = xgb
