+++ examples/exp_emodb_wav2vec_mlp.ini
@@ -1,0 +1,27 @@
+[EXP]
+root = ./tests/results/
+name = exp_emodb
+runs = 1
+epochs = 10
+save = True
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+labels = ['angry', 'happy', 'neutral', 'sad']
+emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
+emodb.limit = 350
+target = emotion
+[FEATS]
+type = ['wav2vec']
+scale = standard
+[MODEL]
+type = mlp
+layers = {'l1':14, 'l2':8}
+drop = .4
+save = True
+[PLOT]
+best_model = True
+epoch_progression = True
