+++ examples/exp_agedb_explore_data.ini
@@ -1,0 +1,26 @@
+[EXP]
+root = ./tests/results/
+name = exp_agedb_explore
+type = regression
+runs = 1
+epochs = 1
+save = True
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = random
+emodb.limit_samples = 20
+target = age
+labels = ['u29', 'ü29']
+bins  = [-100000, 29, 100000]
+[FEATS]
+type = ['os']
+[MODEL]
+type = xgr
+[EXPL]
+sample_selection = all
+value_counts = [['gender', 'duration']]
+#age.bin_reals = True
+#, ['age'], ['duration'], ['age', 'duration']]
+#dist_type = hist
+[PLOT]
