+++ examples/exp_emodb_finetune.ini
@@ -1,0 +1,21 @@
+[EXP]
+root = ./tests/results/
+name = finetuned
+runs = 1
+epochs = 5
+save = True
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+; labels = ['anger', 'sadness']
+target = emotion
+[FEATS]
+type = []
+[MODEL]
+type = finetune
+; device = cpu
+batch_size = 4
+pretrained_model = microsoft/wavlm-base
