+++ examples/exp_emodb_clap_xgb.ini
@@ -1,0 +1,22 @@
+[EXP]
+root = ./tests/results/
+name = exp_emodb
+runs = 1
+epochs = 1
+save = True
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
+labels = ['angry', 'happy', 'neutral', 'sad']
+target = emotion
+[FEATS]
+type = ['clap']
+scale = standard
+trill.model = /home/audeering.local/fburkhardt/research/trill_model/
+wav2vec.model = /home/audeering.local/fburkhardt/research/wav2vec2-embeddings/wav2vec2-large-robust-ft-swbd-300h 
+[MODEL]
+type = xgb
