+++ examples/exp_emodb_clap_mlp.ini
@@ -1,0 +1,30 @@
+[EXP]
+root = ./tests/results/
+name = exp_emodb
+runs = 1
+epochs = 10
+save = True
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
+labels = ['angry', 'happy', 'neutral', 'sad']
+target = emotion
+[FEATS]
+#type = ['trill']
+type = ['clap']
+#type = ['wav2vec']
+trill.model = /home/audeering.local/fburkhardt/research/trill_model/
+wav2vec.model = /home/audeering.local/fburkhardt/research/wav2vec2-embeddings/wav2vec2-large-robust-ft-swbd-300h 
+scale = standard
+[MODEL]
+type = mlp
+layers = {'l1':1024, 'l2':128}
+#drop = .4
+save = True
+[PLOT]
+best_model = True
+epoch_progression = True
