+++ examples/exp_emodb_rs_os_xgb.ini
@@ -1,0 +1,25 @@
+[EXP]
+root = ./tests/results/
+name = exp_emodb_add_random_spliced
+runs = 1
+epochs = 1
+save = True
+[DATA]
+databases = ['emodb', 'augmented']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+augmented = ./tests/results/exp_emodb/random_spliced.csv
+augmented.type = csv
+augmented.absolute_path = True
+augmented.split_strategy = train
+emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
+labels = ['angry', 'happy', 'neutral', 'sad']
+target = emotion
+[FEATS]
+type = ['os']
+scale = standard
+[MODEL]
+type = xgb
+save = True

[…]
