### 2025-05-13 24ebe18
readd test_install
