+++ .github/workflows/test-install-python313.yml
@@ -1,4 +1,4 @@
-name: Test Installation Python 3.10
+name: Test Installation Python 3.13
 
 on:
   push:
@@ -11,10 +11,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
-    - name: Set up Python 3.10
+    - name: Set up Python 3.13
       uses: actions/setup-python@v4
       with:
-        python-version: '3.10'
+        python-version: '3.13'
     - name: Test installation
       run: |
-        python scripts/test_install.py
+        python tests/test_install.py
