+++ .github/workflows/test-install-python311.yml
@@ -1,10 +1,6 @@
 name: Test Installation Python 3.11
 
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
+on: [push, pull_request]
 
 jobs:
   test-install:
