+++ .github/workflows/test-install-python312.yml
@@ -1,10 +1,6 @@
 name: Test Installation Python 3.12
 
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
+on: [push, pull_request]
 
 jobs:
   test-install:
