### 2025-05-13 felixbur/nkululeko#188 @bagustris
@isdanni

> * Move the **Installation** and **Documentation** sections to the top of the README, before any visualizations or plots.
> * Make optional dependencies clearer. For example:
>
>   * Use `pip install nkululeko[spotlight]` for Spotlight support
>   * Or specify versions directly in [`requirements.txt`](https://github.com/felixbur/nkululeko/blob/main/requirements.txt), e.g., `audformat==1.3.1`
> * Add a brief “**Who is this for?**” section under the Project title, such as:
>   _"Nkululeko is for speech researchers and ML practitioners focused on speaker traits and emotion recognition."_
> * Move detailed plots (e.g., **Confusion matrix**, **t-SNE**, **Bias checking**) to the docs site to keep the README simple and focused on setup and usage.

I updated README per requested. The only exception here is that two plots (bias and uncertainty estimation) were kept and moved ton `Feature` because those are unique features of Nkululeko that may not present in other toolkit.
