### 2025-05-13 felixbur/nkululeko#188 @bagustris
@isdanni done in #192

> ## Paper
> The draft paper successfully identifies the statement of the need and includes a clear summary. There are a few suggestions:
>
> * Minor typos:
>
>   * In line 265, there appears to be a duplicated word: "and and bias analysis".
>   * In line 190: "attachched".
> * References:
>
>   * In line 283, "classificatio" missing "n" in the end.
> * Highlight unique contributions in the summary section to emphasize package novelty compared to the existing solutions. For example, mention that it has config-based experimentation or a user-friendly interface.
> * If possible, consider adding a simple diagram illustrating Nkululeko’s internal workflow. For example, data input(in which format) → feature extraction → ... → output.

The only exception is the last suggestion. Instead of showing Nkululeko internal workflow, which is huge, I plotted a simple workflow experimentation with Nkululeko in the updated paper. The Nkululeko internal workflow can be seen in the README or DeepWiki page: https://deepwiki.org/felixbur/nkululeko (already stated in the README News).
