### 2025-05-13 felixbur/nkululeko#188 @bagustris
> ## Installation
> * I suggest adding a code block on creating the virtual environment:
>
> # specify python version if needed
> python -m venv venv
> source venv/bin/activate  # specify OS versions, add a separate line for Windows users
> pip install nkululeko
> * Clarify which dependencies are optional. Also, if some are CPU-only/GPU-only, what would be the changes? Or simply define including the optional packages using `pip install nkululeko[all]`.
> * Replace casual phrasing like "if you don't have a GPU (which is probably true...)" with more direct guidelines such as "If you are using a CPU-only setup..."

@isdanni

I accommodate them in PR #192.
