### 2025-05-14 f2e0c26
fix image in the paper
