+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-`Nkululeko` [@nkululeko:2022] is a framework for audio-based machine learning explorations using command line interface and a configuration file, based on machine learning packages like sklearn [@scikit-learn:2011] and pytorch [@torch:2020]. It is written in Python without the need to write Python code by the users. The main features are: training and evaluation of labelled speech databases with state-of-the-art machine learning approach and acoustic feature extractors, a live demonstration interface, and the possibility to store databases with predicted labels. Based on this, the framework can also be used to check on bias in databases by exploring correlations of target labels, like, e.g. depression or diagnosis, with predicted, or additionally given, labels like age, gender, signal distortion ratio or mean opinion score in.
+`Nkululeko` [@nkululeko:2022] is a toolkit for audio-based machine learning explorations using command line interface and a configuration file, based on machine learning packages like sklearn [@scikit-learn:2011] and pytorch [@torch:2020]. It is written in Python without the need to write Python code by the users. The main features are: training and evaluation of labelled speech databases with state-of-the-art machine learning approach and acoustic feature extractors, a live demonstration interface, and the possibility to store databases with predicted labels. Based on this, the framework can also be used to check on bias in databases by exploring correlations of target labels, like, e.g. depression or diagnosis, with predicted, or additionally given, labels like age, gender, signal distortion ratio or mean opinion score in.
 
 # Design choices 
 
@@ -42,7 +42,7 @@ The program is intended for **novice** people interested in speaker characterist
 - Inference of speaker characteristics from a given audio file or streaming audio (can be said also as “weak” labeling for semi-supervised learning).
 
 <!-- module + INI file -->
-Hence, one should be able to use Nkululeko after installing and preparing/downloading their data in the correct format in a single line. 
+Hence, one should be able to use Nkululeko after installing and preparing/downloading their data in the correct format in a single line (add `python -m` if working in development environment without installing it). 
 
 ```bash
 $ nkululeko.MODULE_NAME --config CONFIG_FILE.ini
@@ -106,10 +106,10 @@ value_counts = [['gender'], ['age'], ['est_sdr'], ['est_pesq'], ['est_mos']]
 latex = androids-report
 ```
 
-As can be seen, some of the values simply contain Python data structures like arrays or dictionaries.
+As can be seen, several values simply contain Python data structures like arrays or dictionaries.
 Within this example, an experiment is specified with the name *explore-androids*, and a **result** folder with this name will be created, containing all figures and textual results, including an automatically generated Latex and PDF report on the findings. The overall flow of basic Nkululeko experiments can be shown in \autoref{fig:nkulu_flow}. 
 
-![Nkululeko’s workflow: from a raw dataset into experiment results]\label{fig:nkulu_flow}(./assets/nkulu_flow.png)
+![Nkululeko’s workflow: from a raw dataset into experiment results \label{fig:nkulu_flow}](./assets/nkulu_flow.png)
 
 The *DATA* section sets the location of the database and specifies filters on the sample, in this case limiting the data to 20 samples per speaker at most and at least 2 seconds long.
 In this section, the split sets (training, development, and test) are also specified. There is a special feature named *balance splits* that lets the user specify criteria that should be used to stratify the splits, for example, based on signal distortion ratio.
@@ -125,7 +125,7 @@ The `nkululeko` configuration can specify further sections:
 # Example of usage
[…]
