### 2025-05-15 d8a06ad
Update documentation with Devin Wiki content and fix Spotlight installation instructions
Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
