### 2025-05-23 29dd806
add emotion2vec
