+++ examples/exp_polish_emotion2vec.ini
@@ -1,0 +1,27 @@
+[EXP]
+root = results/
+name = exp_polish_emotion2vec
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/polish/polish_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/polish/polish_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/polish/polish_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+labels = ['anger', 'neutral', 'fear']
+[FEATS]
+type = ['emotion2vec-base']
+needs_feature_extraction = True
+no_reuse = True
+[MODEL]
+type = xgb
+[RESAMPLE]
+replace = True
