+++ nkululeko/feat_extract/feats_emotion2vec.py
@@ -74,7 +74,7 @@ class Emotion2vec(Featureset):
         store = self.util.get_path("store")
         storage = f"{store}{self.name}.pkl"
         extract = self.util.config_val("FEATS", "needs_feature_extraction", False)
-        no_reuse = eval(self.util.config_val("FEATS", "no_reuse", "False"))
+        no_reuse = self.util.config_val("FEATS", "no_reuse", "False") == "True"
         if extract or no_reuse or not os.path.isfile(storage):
             if not self.model_initialized:
                 self.init_model()
