+++ nkululeko/feat_extract/feats_emotion2vec.py
@@ -1,6 +1,6 @@
 # feats_emotion2vec.py
 # emotion2vec feature extractor for Nkululeko
-# example feat_type = "emotion2vec-large", "emotion2vec-base", "emotion2vec-seed"
+# choices for feat_type = "emotion2vec", "emotion2vec-large", "emotion2vec-base", "emotion2vec-seed"
 
 # requirements:
 # pip install "modelscope>=1.9.5,<2.0.0"
@@ -49,7 +49,7 @@ class Emotion2vec(Featureset):
         # Map feat_type to model names
         model_mapping = {
             "emotion2vec": "iic/emotion2vec_base",
-            "emotion2vec-base": "iic/emotion2vec_plus_base",
+            "emotion2vec-base": "iic/emotion2vec_base_finetuned",
             "emotion2vec-seed": "iic/emotion2vec_plus_seed",
             "emotion2vec-large": "iic/emotion2vec_plus_large",
         }
@@ -86,33 +86,7 @@ class Emotion2vec(Featureset):
             for idx, (file, start, end) in enumerate(
                 tqdm(self.data_df.index.to_list())
             ):
-                # First, load to get sampling_rate
-                _, sampling_rate = torchaudio.load(file, frame_offset=0, num_frames=-1)
-
-                # Calculate offsets in samples
-                start_sample = (
-                    int(start.total_seconds() * sampling_rate)
-                    if hasattr(start, "total_seconds")
-                    else 0
-                )
-                num_samples = (
-                    int((end - start).total_seconds() * sampling_rate)
-                    if hasattr(start, "total_seconds") and hasattr(end, "total_seconds")
-                    else -1
-                )
-
-                # Now load the segment
-                signal, sampling_rate = torchaudio.load(
-                    file, frame_offset=start_sample, num_frames=num_samples
-                )
-
-                # Resample to 16kHz if needed (emotion2vec expects 16kHz)
-                if sampling_rate != 16000:
-                    resampler = torchaudio.transforms.Resample(sampling_rate, 16000)
-                    signal = resampler(signal)
-                    sampling_rate = 16000
-
-                emb = self.get_embeddings(signal, sampling_rate, file)
+                emb = self.extract_embedding(file, start, end)
                 emb_series.iloc[idx] = emb
             self.df = pd.DataFrame(emb_series.values.tolist(), index=self.data_df.index)
             self.df.to_pickle(storage)
@@ -130,32 +104,44 @@ class Emotion2vec(Featureset):
                     f"got nan: {self.df.shape} {self.df.isnull().sum().sum()}"
                 )
 
-    def get_embeddings(self, signal, sampling_rate, file):
-        """Extract embeddings from raw audio signal."""
+    def extract_embedding(self, file, start, end):
+        """Extract embeddings directly from audio file."""
         try:
-            # Convert tensor to numpy if needed
-            if torch.is_tensor(signal):
+            # Handle segment extraction if needed
+            if hasattr(start, "total_seconds") and hasattr(end, "total_seconds"):
+                # Load audio segment
+                _, sampling_rate = torchaudio.load(file, frame_offset=0, num_frames=-1)
+                start_sample = int(start.total_seconds() * sampling_rate)
+                num_samples = int((end - start).total_seconds() * sampling_rate)
+                signal, sampling_rate = torchaudio.load(
+                    file, frame_offset=start_sample, num_frames=num_samples
+                )
+
+                # Resample to 16kHz if needed
+                if sampling_rate != 16000:
+                    resampler = torchaudio.transforms.Resample(sampling_rate, 16000)
+                    signal = resampler(signal)
+                    sampling_rate = 16000
+
+                # Convert to numpy and save as temporary file
[…]
