+++ examples/exp_polish_emotion2vec.ini
@@ -18,7 +18,7 @@ test.split_strategy = test
 target = emotion
 labels = ['anger', 'neutral', 'fear']
 [FEATS]
-type = ['emotion2vec-base']
+type = ['emotion2vec']
 needs_feature_extraction = True
 no_reuse = True
 [MODEL]
