### 2025-05-23 72d41bb
Update nkululeko/feat_extract/feats_emotion2vec.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
