### 2025-05-23 2a070f1
Update nkululeko/feat_extract/feats_emotion2vec.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
