### 2025-05-26 #8049 @faroit
@isdanni 👋 thanks again for accepting to review. Can you give us an update on how the review is going? If you have issues that you'd like the authors to address, please create issues  in the [target repository](https://github.com/felixbur/nkululeko) and link to those issues (especially acceptance-blockers) by leaving comments in in this thread.
