### 2025-05-27 #8049 @isdanni
> [@isdanni](https://github.com/isdanni) 👋 thanks again for accepting to review. Can you give us an update on how the review is going? If you have issues that you'd like the authors to address, please create issues in the [target repository](https://github.com/felixbur/nkululeko) and link to those issues (especially acceptance-blockers) by leaving comments in in this thread.

Thanks for checking in! I have an [active issue](https://github.com/felixbur/nkululeko/issues/188) in the target repo. Authors have responded to my first round of review. I will do a second round by the end of this week.
