+++ unzip.py
@@ -8,5 +8,5 @@ import sys
 from zipfile import PyZipFile
 
 for zip_file in sys.argv[1:]:
-    pzf = PyZipFile(zip_file)
-    pzf.extractall()
+    with PyZipFile(zip_file) as pzf:
+        pzf.extractall()
