### 2025-05-28 felixbur/nkululeko#201 @fakufaku
[FOSS, reviewer2] Nkululeko 1.0: A Python package to predict speaker characteristics with a high-level interface
Hi @felixbur , I'll be the infamous reviewer 2 for your submission to FOSS. I will list my comments below.

The paper is overall well written and easy to follow. It introduces clearly the goal of the package and how to use it.

## Repo

* In the README: `do machine learning experiments combining features and learners` is ambiguous. Does this mean training a new model or applying an existing model to a dataset? I see that this is defined in the doce as `Central class orchestrating data loading, feature extraction, and model training`. Please include this in the paper and the relevant section of the README.
* It may be useful to include the dataset licence information in the READMEs of the datasets under `data`.
* The link to the doc in `https://github.com/felixbur/nkululeko/tree/main/docs/README.md` is broken.

## Paper

* Line 115: `all of them have defaults.` If the defaults are the values just above, a quick note would be nice to make this explicit.

## Doc

* In `https://nkululeko.readthedocs.io/latest/nkuluflag.html`, I think `python -m nkululeko.nkuluflag.py` should be `python -m nkululeko.nkuluflag`.
* In `https://nkululeko.readthedocs.io/latest/ensemble.html` might there be a collision between the double hyphens of the options and RST syntax? Or it may be that `--` is rendered as an em-dash. Also seen in other pages of the doc.
