### 2025-05-29 cc390f4
Enhance dataset documentation and licensing information
- Updated README.md to include a new section for dataset license information and added a "License" column to the dataset table.
- Modified descr.yml to include license details for various datasets.
- Adjusted make_readme.py to accommodate the new "License" column in the generated README.
- Corrected the documentation URL in docs/README.md.
- Improved formatting in demo.md and ensemble.md for better readability.
- Updated nkuluflag.md to correct the usage example.
- Enhanced paper.md to clarify module functionalities and added a table for 
[…]
