+++ README.md
@@ -1,289 +1,83 @@
-## Nkululeko
 
-Nkululeko is a project to detect speaker characteristics by machine learning experiments with a high-level interface.
+Nkululeko database repository
+=============================
 
-The idea is to have a framework (based on e.g. sklearn and torch) that can be used to rapidly and automatically analyse audio data and explore machine learning models based on that data. 
-
-Some abilities that Nkululeko provides: combines acoustic features and machine learning models (including feature selection and features concatenation); performs data exploration, selection and visualization the results; finetuning; ensemble learning models; soft labeling (predicting labels with pre-trained model); and inference the model on a test set.
-
-## Who is this for?
-Nkululeko is for speech processing learners, researchers and ML practitioners focused on speaker characteristics, e.g., emotion, age, gender, or disorder detection.
-
-## Installation
-
-Nkululeko requires Python 3.9 or higher with the following build status:  
-
-![Python 3.10](https://github.com/bagustris/nkululeko/actions/workflows/py310-aud-csv.yml/badge.svg)  
-![Python 3.11](https://github.com/bagustris/nkululeko/actions/workflows/py311.yml/badge.svg)    
-![Python 3.12](https://github.com/bagustris/nkululeko/actions/workflows/py312.yml/badge.svg)  
-![Python 3.13](https://github.com/bagustris/nkululeko/actions/workflows/py313.yml/badge.svg)    
-
-Create and activate a virtual Python environment and simply install Nkululeko:
-
-```bash
-python -m venv .env
-source .env/bin/activate  # specify OS versions, add a separate line for Windows users 
-pip install nkululeko
-```
-
-Current version: **0.94.1**
-
-### Optional Dependencies
-
-Nkululeko supports optional dependencies through extras:
-
-```bash
-# Install with PyTorch support
-pip install nkululeko[torch]
-
-# Install with CPU-only PyTorch
-pip install nkululeko[torch-cpu]
-
-# Install with TensorFlow support
-pip install nkululeko[tensorflow]
-
-# Install all optional dependencies
-pip install nkululeko[all]
-```
-
-#### Manual Installation Options
-
-You can also install dependencies manually:
-
-##### PyTorch Installation
-
-For CPU-only installation (recommended for most users):
-```bash
-pip install torch==1.13.1 torchvision==0.14.1 torchaudio==0.13.1 --index-url https://download.pytorch.org/whl/cpu
-```
-
-For GPU support (cuda 12.6):
-```bash
-pip install torch torchvision torchaudio
-```
-
-Some functionalities require extra packages to be installed, which we didn't include automatically:
-
-* For spotlight adapter:
-  ```bash
-  pip install PyYAML  # Install PyYAML first to avoid dependency issues
-  pip install nkululeko[spotlight]
-  ```
-
-Some examples for *ini*-files (which you use to control nkululeko) are in the [examples folder](https://github.com/felixbur/nkululeko/tree/main/examples).
+# Data
 
 
+This is the default top directory for Nkululeko data import. Each database should be in its own subfolder (you can also use `ln -sf` to soft link original database path to these subfolders) and contain a README how to import the data to Nkululeko CSV or audformat.
 
-## Documentation
-The documentation, along with extensions of installation, usage, INI file format, and examples, can be found [nkululeko.readthedocs.io](https://nkululeko.readthedocs.io).
+## Accessibility
[…]
