### 2025-05-29 felixbur/nkululeko#201 @bagustris
Hi @fakufaku,

Commit [cc390f4](https://github.com/felixbur/nkululeko/commit/cc390f4f287e20ac7623c7b29c8ebc0961647964) was targeted to fix this issue.

Let me know your comments.
