+++ data/make_readme.py
@@ -5,15 +5,20 @@ from mdutils import MdUtils
 
 root = "./data"
 
-mdFile = MdUtils(file_name="README.md", title="Nkululeko database repository")
+mdFile = MdUtils(file_name=os.path.join(root, "README.md"), title="Nkululeko Data Directory")
 mdFile.create_md_file()
-mdFile.new_header(level=1, title="Data")
+mdFile.new_header(level=1, title="README Creation")
 mdFile.new_paragraph(
     "This is the default top directory for Nkululeko data import. "
     "Each database should be in its own subfolder "
     "(you can also use `ln -sf` to soft link original database "
     "path to these subfolders) "
     "and contain a README how to import the data to Nkululeko CSV or audformat."
+    "This README file is automatically generated by "
+    "`data/make_readme.py` script (needs `mdutils` package). "
+    "To update this file, run the script again after adding new datasets "
+    "and adding the necessary information to the `descr.yml` file. "
+
 )
 mdFile.new_header(level=2, title="Accessibility")
 mdFile.new_paragraph(
@@ -51,7 +56,7 @@ for dataset_name in descr.keys():
             descr[dataset_name][1]["target"],
             descr[dataset_name][2]["descr"],
             descr[dataset_name][3]["access"],
-            descr[dataset_name][4]["license"] if len(descr[dataset_name]) > 4 and "license" in descr[dataset_name][4] else "unknown",
+            descr[dataset_name][4]["license"],
         ]
     )
     db_num += 1
@@ -69,3 +74,7 @@ mdFile.new_line(
 )
 mdFile.new_line()
 mdFile.create_md_file()
+
+print(
+    f"README.md file created with {db_num} datasets in {root} directory."
+)
