+++ README.md
@@ -1,83 +1,291 @@
+## Nkululeko
 
-Nkululeko database repository
-=============================
+Nkululeko is a project to detect speaker characteristics by machine learning experiments with a high-level interface.
 
-# Data
+The idea is to have a framework (based on e.g. sklearn and torch) that can be used to rapidly and automatically analyse audio data and explore machine learning models based on that data. 
 
+Some abilities that Nkululeko provides: combines acoustic features and machine learning models (including feature selection and features concatenation); performs data exploration, selection and visualization the results; finetuning; ensemble learning models; soft labeling (predicting labels with pre-trained model); and inference the model on a test set.
 
-This is the default top directory for Nkululeko data import. Each database should be in its own subfolder (you can also use `ln -sf` to soft link original database path to these subfolders) and contain a README how to import the data to Nkululeko CSV or audformat.
+Nkululeko orchestrates data loading, feature extraction, and model training, allowing you to specify your experiment in a configuration file. The framework handles the process from raw data to trained model and evaluation, making it easy to run machine learning experiments without directly coding in Python.
 
-## Accessibility
+## Who is this for?
+Nkululeko is for speech processing learners, researchers and ML practitioners focused on speaker characteristics, e.g., emotion, age, gender, or disorder detection.
 
+## Installation
 
-The column `access` in the table below indicates the database's accessability. The following values are used:
-- `public`: the database is publicly available in the internet and can be downloaded directly without any restrictions.
-- `restricted`: the database is publicly available on the internet but requires registration or other restrictions to download.
-- `private`: the database is not publicly available on the internet and requires the private information of the owner of the dataset.
+Nkululeko requires Python 3.9 or higher with the following build status:  
 
+![Python 3.10](https://github.com/bagustris/nkululeko/actions/workflows/py310-aud-csv.yml/badge.svg)  
+![Python 3.11](https://github.com/bagustris/nkululeko/actions/workflows/py311.yml/badge.svg)    
+![Python 3.12](https://github.com/bagustris/nkululeko/actions/workflows/py312.yml/badge.svg)  
+![Python 3.13](https://github.com/bagustris/nkululeko/actions/workflows/py313.yml/badge.svg)    
 
-To support open science and reproducible research, we encourage to submit PR and recipes for public dataset for now on.
-|Name|Target|Description|Access|License|
-| :--- | :--- | :--- | :--- | :--- |
-|emorynlp|emotion|English Emotion Dataset from Friends TV Show|public|unknown|
-|emns|emotion,intensity|British, singles peaker, UAR=.479|public|CC BY 4.0|
-|test|none|Test data for nkululeko|public|unknown|
-|catsvsdogs|cats_dogs|kaggle test set|public|unknown|
-|clac|healthyspeech,age,gender|English|public|Custom academic/research use|
-|vivae|emotion|English vocal bursts|public|unknown|
-|emofilm|emotion|English, Spanish, Italian|restricted|unknown|
-|emozionalmente|emotion|Italian|public|unknown|
-|laughter-types|laughter|Master Thesis from TUBerlin|public|unknown|
[…]
