+++ data/README.md
@@ -1,15 +1,13 @@
-Nkululeko database repository
-=============================
 
-# Data
+Nkululeko Data Directory
+========================
 
-This is the default top directory for Nkululeko data import. Each database should be in its own subfolder (you can also use `ln -sf` to soft link original database path to these subfolders) and contain a README how to import the data to Nkululeko CSV or audformat.
+# README Creation
 
-## Dataset License Information
 
-Each dataset may have its own license terms. Please refer to the README or license file within each dataset subfolder for specific license details. If not present, consult the official source of the dataset for license and usage information. If you add a new dataset, please include a README with license and citation information.
-
+This is the default top directory for Nkululeko data import. Each database should be in its own subfolder (you can also use `ln -sf` to soft link original database path to these subfolders) and contain a README how to import the data to Nkululeko CSV or audformat.This README file is automatically generated by `data/make_readme.py` script (needs `mdutils` package). To update this file, run the script again after adding new datasets and adding the necessary information to the `descr.yml` file. 
 ## Accessibility
+
 
 The column `access` in the table below indicates the database's accessability. The following values are used:
 - `public`: the database is publicly available in the internet and can be downloaded directly without any restrictions.
@@ -21,37 +19,38 @@ To support open science and reproducible research, we encourage to submit PR and
 |Name|Target|Description|Access|License|
 | :--- | :--- | :--- | :--- | :--- |
 |emorynlp|emotion|English Emotion Dataset from Friends TV Show|public|unknown|
-|emns|emotion,intensity|British, singles peaker, UAR=.479|public|unknown|
+|emns|emotion,intensity|British, singles speaker, UAR=.479|public|CC BY 4.0|
 |test|none|Test data for nkululeko|public|unknown|
-|clac|healthyspeech,age,gender|English|public|unknown|
+|catsvsdogs|cats_dogs|kaggle test set|public|unknown|
+|clac|healthyspeech,age,gender|English|public|Custom academic/research use|
 |vivae|emotion|English vocal bursts|public|unknown|
 |emofilm|emotion|English, Spanish, Italian|restricted|unknown|
 |emozionalmente|emotion|Italian|public|unknown|
 |laughter-types|laughter|Master Thesis from TUBerlin|public|unknown|
-|savee|emotion|English, from tfds|restricted|unknown|
-|emovo|emotion|Italian|restricted|unknown|
-|subesco|emotion|Bangla|public|unknown|
+|savee|emotion|English, from tfds|restricted|Custom academic/research use|
+|emovo|emotion|Italian|restricted|CC BY-NC-SA 4.0|
+|subesco|emotion|Bangla|public|Custom academic/research use|
 |oreau2|emotion|French|public|unknown|
 |mess|emotion|English|public|unknown|
-|emov-db|emotion|English|public|unknown|
+|emov-db|emotion|English|public|CC BY 4.0|
 |odyssey-cat-2024|emotion|Data for Odyssey 2024 challenge, needs MSPPodcast|restricted|unknown|
-|ravdess|emotion,speaker|English|public|unknown|
+|ravdess|emotion,speaker|English|public|CC BY 4.0|
 |erysac|emotion|Russian, children|public|unknown|
-|demos|emotion|Italian|restricted|unknown|
-|emodb|emotion|German|public|unknown|
+|demos|emotion|Italian|restricted|Custom academic/research use|
[…]
