+++ data/descr.yml
@@ -10,7 +10,7 @@ EMORYLP:
 EMNS:
 - name: emns
 - target: emotion,intensity
-- descr: British, singles peaker, UAR=.479
+- descr: British, singles speaker, UAR=.479
 - access: public
 - license: CC BY 4.0
 
