+++ examples/exp_emodb_os_xgb_test.ini
@@ -4,7 +4,7 @@ name = exp_testmodule
 save = True
 [DATA]
 databases = ['emodb']
-root_folders = tests/data_roots.ini
+root_folders = examples/data_roots.ini
 target = emotion
 tests = ['crema-d']
 labels = ['angry', 'happy']
