### 2025-06-03 a5e89be
update examples root dir
