+++ examples/age_tu.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = age_tu
 type = regression
 runs = 1
