+++ examples/exp_emodb_praat_xgb.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_emodb_feats
 runs = 1
 epochs = 1
