+++ examples/exp_agedb_os_mlp.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_agedb_mlp
 type = regression
 runs = 1
