+++ examples/exp_agedb_explore_data.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_agedb_explore
 type = regression
 runs = 1
