+++ examples/exp_ravdess_speaker.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_ravdess_speaker
 runs = 1
 epochs = 1
