+++ examples/exp_emodb_audmodel_xgb.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_emodb_audmodel_xgb
 runs = 1
 epochs = 1
