+++ examples/exp_ravdess_os_xgb.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_ravdess
 runs = 1
 epochs = 1
