+++ examples/exp_androids_cnn.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_androids_cnn
 runs = 1
 epochs = 50
