+++ examples/exp_emodb_os_knn.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_emodb_classifiers
 runs = 1
 epochs = 10
