+++ examples/exp_emodb_wav2vec_mlp.ini
@@ -1,5 +1,5 @@
 [EXP]
-root = ./tests/results/
+root = ./examples/results/
 name = exp_emodb
 runs = 1
 epochs = 10
