### 2025-06-03 21a0843
comment personal path in examples
