+++ examples/exp_emodb_clap_mlp.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb
+name = exp_emodb_clap
 runs = 1
 epochs = 10
 save = True
@@ -17,8 +17,8 @@ target = emotion
 #type = ['trill']
 type = ['clap']
 #type = ['wav2vec']
-trill.model = /home/audeering.local/fburkhardt/research/trill_model/
-wav2vec.model = /home/audeering.local/fburkhardt/research/wav2vec2-embeddings/wav2vec2-large-robust-ft-swbd-300h 
+; trill.model = /home/audeering.local/fburkhardt/research/trill_model/
+; wav2vec.model = /home/audeering.local/fburkhardt/research/wav2vec2-embeddings/wav2vec2-large-robust-ft-swbd-300h 
 scale = standard
 [MODEL]
 type = mlp
