+++ examples/exp_emodb_clap_xgb.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb
+name = exp_emodb_clap
 runs = 1
 epochs = 1
 save = True
@@ -16,7 +16,7 @@ target = emotion
 [FEATS]
 type = ['clap']
 scale = standard
-trill.model = /home/audeering.local/fburkhardt/research/trill_model/
-wav2vec.model = /home/audeering.local/fburkhardt/research/wav2vec2-embeddings/wav2vec2-large-robust-ft-swbd-300h 
+; trill.model = /home/audeering.local/fburkhardt/research/trill_model/
+; wav2vec.model = /home/audeering.local/fburkhardt/research/wav2vec2-embeddings/wav2vec2-large-robust-ft-swbd-300h 
 [MODEL]
 type = xgb
