+++ examples/exp_emodb_audmodel_xgb.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb_audmodel_xgb
+name = exp_emodb_audmodel
 runs = 1
 epochs = 1
 save = True
