+++ examples/exp_emodb_os_xgb.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb_classifiers
+name = exp_emodb_os
 runs = 1
 epochs = 10
 save = True
