+++ examples/exp_emodb_audmodel_mlp.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb
+name = exp_emodb_audmodel
 runs = 1
 epochs = 10
 save = True
