+++ examples/exp_emodb_praat_xgb.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb_feats
+name = exp_emodb_praat
 runs = 1
 epochs = 1
 save = True
