+++ examples/exp_emodb_finetune.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = finetuned
+name = wavlm_finetuned
 runs = 1
 epochs = 5
 save = True
