+++ examples/exp_ravdess_ast_xgb.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_ravdess
+name = exp_ravdess_ast
 runs = 1
 epochs = 1
 save = True
