+++ examples/exp_emodb_wav2vec_mlp.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb
+name = exp_emodb_wav2vec
 runs = 1
 epochs = 10
 save = True
