+++ examples/exp_emodb_os_svm.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb_classifiers
+name = exp_emodb_praat
 save = True
 [DATA]
 databases = ['emodb']
