+++ examples/exp_ravdess_os_xgb.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_ravdess
+name = exp_ravdess_os
 runs = 1
 epochs = 1
 save = True
