### 2025-06-04 770845e
add exp.ini for dementianet
