+++ data/dementianet/exp.ini
@@ -1,0 +1,40 @@
+[EXP]
+root = ./results
+name = exp_dementianet_praat
+runs = 1
+epochs = 1
+save = True
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/dementianet/dementianet_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/dementianet/dementianet_val.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/dementianet/dementianet_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = dementia
+[FEATS]
+type = ['praat']
+; features = ['duration', 'meanF0Hz', 'stdevF0Hz', 'HNR', 'localJitter',
+;        'localabsoluteJitter', 'rapJitter', 'ppq5Jitter', 'ddpJitter',
+;        'localShimmer', 'localdbShimmer', 'apq3Shimmer', 'apq5Shimmer',
+;        'apq11Shimmer', 'ddaShimmer', 'f1_mean', 'f2_mean', 'f3_mean',
+;        'f4_mean', 'f1_median', 'f2_median', 'f3_median', 'f4_median',
+;        'JitterPCA', 'ShimmerPCA', 'pF', 'fdisp', 'avgFormant', 'mff',
+;        'fitch_vtl', 'delta_f', 'vtl_delta_f', 'nsyll', 'npause', 'dur_s',
+;        'phonationtime_s', 'speechrate_nsyll_dur',
+;        'articulation_rate_nsyll_phonationtime', 'ASD_speakingtime_nsyll']
+scale = standard
+; balancing = smote
+[MODEL]
+type = xgb
+; save = True
+; C_val = 10
+[RESAMPLE]
+replace = True
