+++ data/dementianet/exp.ini
@@ -30,8 +30,15 @@ type = ['praat']
 ;        'fitch_vtl', 'delta_f', 'vtl_delta_f', 'nsyll', 'npause', 'dur_s',
 ;        'phonationtime_s', 'speechrate_nsyll_dur',
 ;        'articulation_rate_nsyll_phonationtime', 'ASD_speakingtime_nsyll']
-scale = standard
+; features = ['speechrate_nsyll_dur', 'f3_mean', 'localJitter', 'localdbShimmer', 'fitch_vtl', 'JitterPCA', 'ShimmerPCA', 'pF', 'fdisp', 'avgFormant', 'mff', 'duration', 'delta_f', 'f3_median', 'vtl_delta_f', 'nsyll', 'npause', 'dur_s', 'phonationtime_s', 'articulation_rate_nsyll_phonationtime']
+scale = speaker
 ; balancing = smote
+[EXPL]
+feature_distributions = True
+model = ['tree']
+permutation = True
+max_feats = 20
+sample_selection = all
 [MODEL]
 type = xgb
 ; save = True
