+++ nkululeko/nkululeko.py
@@ -54,6 +54,13 @@ def doit(config_file):
     reports, last_epochs = expr.run()
     result = expr.get_best_report(reports).result.test
     expr.store_report()
+
+    # check if we want to export the model
+    o_path = util.config_val("EXP", "export_onnx", "False")
+    if eval (o_path):
+        print(f"Exporting ONNX model to {o_path}")
+        o_path = o_path.replace('"', '')
+        expr.runmgr.get_best_model().export_onnx(str(o_path))
     print("DONE")
     return result, int(np.asarray(last_epochs).min())
 
