+++ data/dementiabank/process_database.py
@@ -1,0 +1,276 @@
+# process_database.py for DementiaBank dataset
+# Following TensorFlow datasets split strategy with speaker-aware grouping
+
+import argparse
+import collections
+import os
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+
+
+def _compute_split_boundaries(split_probs, n_items):
+    """Computes boundary indices for each of the splits in split_probs.
+
+    Args:
+        split_probs: List of (split_name, prob), e.g. [('train', 0.7), ('validation', 0.1), ('test', 0.2)]
+        n_items: Number of items we want to split.
+
+    Returns:
+        The item indices of boundaries between different splits. For the above
+        example and n_items=100, these will be
+        [('train', 0, 70), ('validation', 70, 80), ('test', 80, 100)].
+    """
+    if len(split_probs) > n_items:
+        raise ValueError(
+            f"Not enough items for the splits. There are {len(split_probs)} "
+            f"splits while there are only {n_items} items"
+        )
+    total_probs = sum(p for name, p in split_probs)
+    if abs(1 - total_probs) > 1e-8:
+        raise ValueError(f"Probs should sum up to 1. probs={split_probs}")
+
+    split_boundaries = []
+    sum_p = 0.0
+    for name, p in split_probs:
+        prev = sum_p
+        sum_p += p
+        split_boundaries.append((name, int(prev * n_items), int(sum_p * n_items)))
+
+    # Guard against rounding errors.
+    split_boundaries[-1] = (
+        split_boundaries[-1][0],
+        split_boundaries[-1][1],
+        n_items,
+    )
+
+    return split_boundaries
+
+
+def _get_inter_splits_by_group(items_and_groups, split_probs, split_number):
+    """Split items to train/dev/test, so all items in group go into same split.
+
+    Each group contains all the samples from the same speaker ID. The samples are
+    splitted so that all each speaker belongs to exactly one split.
+
+    Args:
+        items_and_groups: Sequence of (item_id, group_id) pairs.
+        split_probs: List of (split_name, prob), e.g. [('train', 0.7), ('validation', 0.1), ('test', 0.2)]
+        split_number: Generated splits should change with split_number.
+
+    Returns:
+        Dictionary that looks like {split name -> list(ids)}.
+    """
+    groups = sorted(set(group_id for item_id, group_id in items_and_groups))
+    rng = np.random.RandomState(split_number)
+    rng.shuffle(groups)
+
+    split_boundaries = _compute_split_boundaries(split_probs, len(groups))
+    group_id_to_split = {}
+    for split_name, i_start, i_end in split_boundaries:
+        for i in range(i_start, i_end):
+            group_id_to_split[groups[i]] = split_name
+
+    split_to_ids = collections.defaultdict(list)
+    for item_id, group_id in items_and_groups:
+        split = group_id_to_split[group_id]
+        split_to_ids[split].append(item_id)
+
+    return split_to_ids
+
+
+def process_database(data_dir=None, output_dir=None):
+    """
+    Process the DementiaBank dataset to create train/val/test splits using TensorFlow datasets strategy.
+
+    This follows the same speaker-aware splitting strategy as TensorFlow datasets:
+    - 70% train, 10% validation, 20% test
+    - Speaker-aware: all files from the same speaker go to the same split
+    - Uses random state 0 for reproducibility (same as TF datasets)
+
+    Args:
+        data_dir: Path to the dementiabank directory (optional, defaults to current directory)
[…]
