### 2025-06-05 5a0ce78
feat: modify xgb to add early_stopping (see new example)
