### 2025-06-05 39df358
Update nkululeko/models/model_xgb.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
