### 2025-06-05 f4b9dd7
Update data/dementiabank/process_database.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
