+++ nkululeko/models/model_xgb.py
@@ -14,31 +14,35 @@ class XGB_model(Model):
         super().__init__(df_train, df_test, feats_train, feats_test)
         self.name = "xgb"
         self.is_classifier = True
-        
+
         # Configure XGBoost parameters
         xgb_params = {}
-        
+
         # Get early stopping configuration
-        self.early_stopping_rounds = self.util.config_val("MODEL", "early_stopping_rounds", False)
+        self.early_stopping_rounds = self.util.config_val(
+            "MODEL", "early_stopping_rounds", False
+        )
         self.eval_metric = self.util.config_val("MODEL", "eval_metric", "logloss")
-        
+
         # Set up other XGBoost parameters that can be configured
         n_estimators = self.util.config_val("MODEL", "n_estimators", 100)
         max_depth = self.util.config_val("MODEL", "max_depth", 6)
         learning_rate = self.util.config_val("MODEL", "learning_rate", 0.3)
-        
-        xgb_params['n_estimators'] = int(n_estimators)
-        xgb_params['max_depth'] = int(max_depth)
-        xgb_params['learning_rate'] = float(learning_rate)
-        
+        subsample = self.util.config_val("MODEL", "subsample", 1.0)
+
+        xgb_params["n_estimators"] = int(n_estimators)
+        xgb_params["max_depth"] = int(max_depth)
+        xgb_params["learning_rate"] = float(learning_rate)
+        xgb_params["subsample"] = float(subsample)
+
         # Set random state for reproducibility
-        xgb_params['random_state'] = 42
-        
+        xgb_params["random_state"] = 42
+
         # Add early stopping parameters to model initialization if configured
         if self.early_stopping_rounds:
-            xgb_params['early_stopping_rounds'] = int(self.early_stopping_rounds)
-            xgb_params['eval_metric'] = self.eval_metric
-        
+            xgb_params["early_stopping_rounds"] = int(self.early_stopping_rounds)
+            xgb_params["eval_metric"] = self.eval_metric
+
         # Initialize classifier with parameters
         self.clf = XGBClassifier(**xgb_params)
 
@@ -54,21 +58,24 @@ class XGB_model(Model):
 
         feats = self.feats_train.to_numpy()
         labels = self.df_train[self.target]
-        
+
         # Configure fitting parameters
         fit_params = {}
-        
+
         # Check if early stopping is configured
         if self.early_stopping_rounds:
             # Check if we're in split3 mode (train/dev/test) where validation data is available
             import ast
-            split3 = ast.literal_eval(self.util.config_val("EXP", "traindevtest", "False"))
-            
+
+            split3 = ast.literal_eval(
+                self.util.config_val("EXP", "traindevtest", "False")
+            )
+
             if split3 and self.feats_test is not None and self.df_test is not None:
                 # In split3 mode, self.feats_test and self.df_test are actually the dev set
                 feats_dev = self.feats_test.to_numpy()
                 labels_dev = self.df_test[self.target]
-                
+
                 # Handle NANs in dev features
                 if self.feats_test.isna().to_numpy().any():
                     self.util.debug(
@@ -76,61 +83,81 @@ class XGB_model(Model):
                         f" {self.feats_test.isna().sum().sum()} NANs with 0"
                     )
                     feats_dev = self.feats_test.fillna(0).to_numpy()
-                
+
                 # Set up early stopping with validation data
                 eval_set = [(feats, labels), (feats_dev, labels_dev)]
-                fit_params['eval_set'] = eval_set
-                fit_params['verbose'] = True
-                
-                self.util.debug(f"Training XGBoost with early stopping (using dev set):")
-                self.util.debug(f"  - early_stopping_rounds: {self.early_stopping_rounds}")
[…]
