### 2025-06-10 53931ba
feats: update shap, make pylatex optional
