+++ requirements.txt
@@ -19,7 +19,7 @@ scipy>=1.0.0
 seaborn>=0.11.0
 sounddevice>=0.4.0
 transformers>=4.0.0
-umap-learn>=0.5.0
+umap-learn==0.5.7
 xgboost>=1.0.0
 pylatex>=1.0.0
 
