+++ requirements.txt
@@ -10,7 +10,7 @@ confidence-intervals>=0.0.2
 datasets>=2.0.0
 imageio>=2.0.0
 matplotlib>=3.0.0
-numpy>=1.20.0
+numpy==2.3.0
 opensmile>=2.0.0
 pandas>=1.0.0
 praat-parselmouth>=0.4.0
@@ -41,5 +41,5 @@ torchvision>=0.10.0
 # cylimiter>=0.0.1
 # audtorch>=0.0.1
 imblearn>=0.0.0
-shap>=0.40.0
+shap==0.47.2
 splitutils>=0.0.1
