### 2025-06-12 5e85a06
update requirement and readme (add uv)
