+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.94.2"
+VERSION="0.94.3"
 SAMPLING_RATE = 16000
