### 2025-06-13 2a7d4aa
fix: shap in explore module
