+++ requirements.txt
@@ -14,7 +14,7 @@ numpy==2.3.0
 opensmile>=2.0.0
 pandas>=1.0.0
 praat-parselmouth>=0.4.0
-scikit_learn>=1.0.0
+scikit_learn>=1.4.2
 scipy>=1.0.0
 seaborn>=0.11.0
 sounddevice>=0.4.0
