### 2025-06-13 59bc3b1
fix: error when balancing in explore module
