+++ nkululeko/plots.py
@@ -87,9 +87,10 @@ class Plots:
 
     def plot_distributions(self, df, type_s="samples"):
         class_label, df = self._check_binning("class_label", df)
-        attributes = ast.literal_eval(
-            self.util.config_val("EXPL", "value_counts", False)
-        )
+        value_counts_conf = self.util.config_val("EXPL", "value_counts", False)
+        if not isinstance(value_counts_conf, str):
+            value_counts_conf = str(value_counts_conf)
+        attributes = ast.literal_eval(value_counts_conf)
         # always plot the distribution of the main attribute
         filename = f"{class_label}_distribution"
         if self.util.is_categorical(df[class_label]):
@@ -216,7 +217,7 @@ class Plots:
 
     def save_plot(self, ax, caption, header, filename, type_s):
         # one up because of the runs
-        fig_dir = self.util.get_path("fig_dir")  # + "../"
+        fig_dir = self.util.get_path("fig_dir") + "../"
         fig_plots = ax.figure
         # avoid warning
         # plt.tight_layout()
@@ -359,7 +360,7 @@ class Plots:
 
     def plot_durations(self, df, filename, sample_selection, caption=""):
         # one up because of the runs
-        fig_dir = self.util.get_path("fig_dir")  # + "../"
+        fig_dir = self.util.get_path("fig_dir") + "../"
         try:
             ax = sns.histplot(df, x="duration", hue="class_label", kde=True)
         except AttributeError as ae:
@@ -393,7 +394,7 @@ class Plots:
         filename = "speakers"
         caption = "speakers"
         # one up because of the runs
-        fig_dir = self.util.get_path("fig_dir")  # + "../"
+        fig_dir = self.util.get_path("fig_dir") + "../"
         sns.set_style("whitegrid")  # Set style for chart
         ax = df["speaker"].value_counts().plot(kind="pie", autopct="%1.1f%%")
         title = f"Speaker distr. for {sample_selection} {df.shape[0]}."
@@ -462,10 +463,11 @@ class Plots:
 
     def scatter_plot(self, feats, label_df, label, dimred_type):
         dim_num = int(self.util.config_val("EXPL", "scatter.dim", 2))
-        # one up because of the runs
-        fig_dir = self.util.get_path("fig_dir")  # + "../"
+        # one up because of the runs (for explore module)
+        fig_dir = self.util.get_path("fig_dir") + "../"
         sample_selection = self.util.config_val("EXPL", "sample_selection", "all")
-        filename = f"{label}_{self.util.get_feattype_name()}_{sample_selection}_{dimred_type}_{str(dim_num)}d"
+        exp_name = self.util.get_name()
+        filename = f"{label}_{exp_name}_{self.util.get_feattype_name()}_{sample_selection}_{dimred_type}_{str(dim_num)}d"
         filename = f"{fig_dir}{filename}.{self.format}"
         self.util.debug(f"computing {dimred_type}, this might take a while...")
         data = None
@@ -573,6 +575,7 @@ class Plots:
             self.util.error(f"wrong dimension number: {dim_num}")
         fig = ax.figure
         plt.savefig(filename)
+        self.util.debug(f"plotted {dimred_type} scatter plot to {filename}")
         fig.clear()
         plt.close(fig)
         glob_conf.report.add_item(
@@ -599,7 +602,7 @@ class Plots:
         # remove fullstops in the name
         feature_name = feature.replace(".", "-")
         # one up because of the runs
-        fig_dir = self.util.get_path("fig_dir")  # + "../"
+        fig_dir = self.util.get_path("fig_dir") + "../"
[…]

[…]
