### 2025-06-19 c68ce95
updates plots and model_tuned
