### 2025-06-24 a124721
Update nkululeko/experiment.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
