### 2025-06-24 c8ac250
Update nkululeko/models/model_tuned.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
