### 2025-06-25 6397c03
Apply black formatting to emotion2vec implementation
- Format code style for consistency with project standards
- No functional changes, only formatting improvements

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
