+++ nkululeko/models/model_tuned.py
@@ -84,6 +84,17 @@ class TunedModel(BaseModel):
         # print finetuning information via debug
         self.util.debug(f"Finetuning from model: {pretrained_model}")
 
+        if any(
+            emotion_model in pretrained_model
+            for emotion_model in ["emotion2vec", "iic/emotion2vec"]
+        ):
+            self._init_emotion2vec_model(pretrained_model)
+            return
+
+        self._init_huggingface_model(pretrained_model)
+
+    def _init_huggingface_model(self, pretrained_model):
+        """Initialize HuggingFace transformer model for finetuning."""
         # create dataset
         dataset = {}
         target_name = glob_conf.target
@@ -99,43 +110,9 @@ class TunedModel(BaseModel):
             df = y.reset_index()
             df.start = df.start.dt.total_seconds()
             df.end = df.end.dt.total_seconds()
-            #     ds = datasets.Dataset.from_pandas(df)
-            #     dataset[split] = ds
 
-            # self.dataset = datasets.DatasetDict(dataset)
             if split == "train" and self.balancing:
-                if self.balancing == "ros":
-                    from imblearn.over_sampling import RandomOverSampler
-
-                    sampler = RandomOverSampler(random_state=42)
-                elif self.balancing == "smote":
-                    from imblearn.over_sampling import SMOTE
-
-                    sampler = SMOTE(random_state=42)
-                elif self.balancing == "adasyn":
-                    from imblearn.over_sampling import ADASYN
-
-                    sampler = ADASYN(random_state=42)
-                else:
-                    self.util.error(f"Unknown balancing algorithm: {self.balancing}")
-
-                # The `fit_resample` method from `imblearn` may lack complete type annotations,
-                # leading to type-checking errors. Suppressing these errors as the method is used correctly.
-                X_resampled, y_resampled = sampler.fit_resample(  # type: ignore
-                    df[["start", "end"]], df["targets"]
-                )
-                df = pd.DataFrame(
-                    {
-                        "start": X_resampled["start"],
-                        "end": X_resampled["end"],
-                        "targets": y_resampled,
-                    }
-                )
-
-                # print the before and after class distribution
-                self.util.debug(
-                    f"balanced with: {self.balancing}, new size: {len(df)}, was {len(data_sources[split])}"
-                )
+                df = self._apply_balancing(df, data_sources[split])
 
             ds = datasets.Dataset.from_pandas(df)
             dataset[split] = ds
@@ -208,6 +185,121 @@ class TunedModel(BaseModel):
         self.model.train()  # type: ignore
         self.model_initialized = True
 
+    def _init_emotion2vec_model(self, pretrained_model):
+        """Initialize emotion2vec model for finetuning."""
+        try:
+            from funasr import AutoModel
+        except ImportError:
+            self.util.error(
+                "FunASR is required for emotion2vec finetuning. "
+                "Please install with: pip install funasr modelscope"
+            )
+            return
+
+        model_mapping = {
+            "emotion2vec": "iic/emotion2vec_base",
+            "emotion2vec-base": "iic/emotion2vec_base_finetuned",
+            "emotion2vec-seed": "iic/emotion2vec_plus_seed",
+            "emotion2vec-large": "iic/emotion2vec_plus_large",
+        }
+
+        if pretrained_model in model_mapping:
+            model_path = model_mapping[pretrained_model]
+        else:
+            model_path = pretrained_model
+
+        self._create_emotion2vec_dataset()
+
+        self.emotion2vec_backbone = AutoModel(model=model_path)
+
+        if self.is_classifier:
+            le = glob_conf.label_encoder
+            if le is None:
[…]
