### 2025-06-25 90dd814
Update nkululeko/models/model_tuned.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
