### 2025-06-25 13d64ee
Update nkululeko/plots.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
