### 2025-06-25 e9537da
Fix emotion2vec finetuning prediction pipeline
- Add missing predict() method to Emotion2vecModel class
- Fix ModelOutput.__getitem__ slice handling for HuggingFace Trainer compatibility
- Resolve 'Empty predictions tuple received' error during evaluation
- Successfully tested with Polish dataset: UAR 0.256, ACC 0.256 on 90 test files
- Complete end-to-end emotion2vec finetuning pipeline now functional

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
