### 2025-06-25 319a377
Add Polish emotion2vec prediction example and documentation tutorial
- Create exp_polish_emotion2vec_predict.ini example configuration
- Add comprehensive emotion_prediction.md tutorial with usage examples
- Integrate tutorial into documentation index structure
- Enable users to easily test emotion prediction with Polish dataset

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
