### 2025-06-25 e41733f
Fix dataset paths in Polish emotion2vec prediction example
- Change from ../data/polish/ to ./data/polish/ for correct relative paths
- Paths should be relative to nkululeko root directory, not examples directory
- Tested and verified configuration works correctly from root directory

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
