+++ examples/exp_polish_emotion2vec_predict.ini
@@ -6,15 +6,15 @@ runs = 1
 
 [DATA]
 databases = ['train', 'dev', 'test']
-train = ../data/polish/polish_train.csv
+train = ./data/polish/polish_train.csv
 train.type = csv
 train.absolute_path = False
 train.split_strategy = train
-dev = ../data/polish/polish_dev.csv
+dev = ./data/polish/polish_dev.csv
 dev.type = csv
 dev.absolute_path = False
 dev.split_strategy = train
-test = ../data/polish/polish_test.csv
+test = ./data/polish/polish_test.csv
 test.type = csv
 test.absolute_path = False
 test.split_strategy = test
