### 2025-06-26 a25b5dc
update emotion2vec to use hf
