+++ nkululeko/models/model_tuned.py
@@ -192,15 +192,15 @@ class TunedModel(BaseModel):
         except ImportError:
             self.util.error(
                 "FunASR is required for emotion2vec finetuning. "
-                "Please install with: pip install funasr modelscope"
+                "Please install with: pip install funasr"
             )
             return
 
         model_mapping = {
-            "emotion2vec": "iic/emotion2vec_base",
-            "emotion2vec-base": "iic/emotion2vec_base_finetuned",
-            "emotion2vec-seed": "iic/emotion2vec_plus_seed",
-            "emotion2vec-large": "iic/emotion2vec_plus_large",
+            "emotion2vec": "emotion2vec/emotion2vec_base",
+            "emotion2vec-base": "emotion2vec/emotion2vec_base", 
+            "emotion2vec-seed": "emotion2vec/emotion2vec_plus_seed",
+            "emotion2vec-large": "emotion2vec/emotion2vec_plus_large",
         }
 
         if pretrained_model in model_mapping:
@@ -210,7 +210,10 @@ class TunedModel(BaseModel):
 
         self._create_emotion2vec_dataset()
 
-        self.emotion2vec_backbone = AutoModel(model=model_path)
+        self.emotion2vec_backbone = AutoModel(
+            model=model_path,
+            hub="hf"  # Use HuggingFace Hub instead of ModelScope
+        )
 
         if self.is_classifier:
             le = glob_conf.label_encoder
@@ -226,6 +229,7 @@ class TunedModel(BaseModel):
                 is_classifier=True,
                 sampling_rate=self.sampling_rate,
                 final_dropout=self.drop,
+                model_name=pretrained_model,
             )
         else:
             self.config = EmotionVecConfig(
@@ -233,6 +237,7 @@ class TunedModel(BaseModel):
                 is_classifier=False,
                 sampling_rate=self.sampling_rate,
                 final_dropout=self.drop,
+                model_name=pretrained_model,
             )
 
         self.model = Emotion2vecModel(self.emotion2vec_backbone, self.config)
@@ -839,12 +844,14 @@ class EmotionVecConfig:
         is_classifier=True,
         sampling_rate=16000,
         final_dropout=0.1,
+        model_name=None,
         **kwargs,
     ):
         self.num_labels = num_labels
         self.is_classifier = is_classifier
         self.sampling_rate = sampling_rate
         self.final_dropout = final_dropout
+        self.model_name = model_name
         for key, value in kwargs.items():
             setattr(self, key, value)
 
@@ -872,11 +879,23 @@ class Emotion2vecModel(torch.nn.Module):
         self.config = config
         self.is_classifier = config.is_classifier
 
-        embedding_dim = 768
+        # Determine embedding dimension based on model variant (hardcoded)
+        embedding_dim = self._get_embedding_dim_by_model()
         self.head = torch.nn.Sequential(
             torch.nn.Dropout(config.final_dropout),
             torch.nn.Linear(embedding_dim, config.num_labels),
         )
+
+    def _get_embedding_dim_by_model(self):
+        """Get embedding dimension based on model variant."""
+        model_name = getattr(self.config, 'model_name', '')
+        
+        # Large models have 1024 dimensions
+        if 'large' in model_name.lower():
+            return 1024
+        # Base, seed, and other models have 768 dimensions
+        else:
+            return 768
 
     def forward(self, input_values, labels=None, **kwargs):
         embeddings = self._extract_embeddings(input_values)
@@ -893,8 +912,11 @@ class Emotion2vecModel(torch.nn.Module):
 
     def _extract_embeddings(self, input_values):
         batch_embeddings = []
+        device = next(self.parameters()).device  # Get the device of the model
[…]
