+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.94.4"
+VERSION="0.95.0"
 SAMPLING_RATE = 16000
