+++ nkululeko/balance.py
@@ -1,0 +1,221 @@
+# balance.py
+"""
+Data and feature balancing module for imbalanced datasets.
+
+This module provides a unified interface for various balancing techniques
+including over-sampling, under-sampling, and combination methods.
+"""
+
+import pandas as pd
+import numpy as np
+from nkululeko.utils.util import Util
+import nkululeko.glob_conf as glob_conf
+
+
+class DataBalancer:
+    """Class to handle data and feature balancing operations."""
+    
+    def __init__(self, random_state=42):
+        """
+        Initialize the DataBalancer.
+        
+        Args:
+            random_state (int): Random state for reproducible results
+        """
+        self.util = Util("data_balancer")
+        self.random_state = random_state
+        
+        # Supported balancing algorithms
+        self.oversampling_methods = [
+            'ros',           # RandomOverSampler
+            'smote',         # SMOTE
+            'adasyn',        # ADASYN
+            'borderlinesmote',  # BorderlineSMOTE
+            'svmsmote'       # SVMSMOTE
+        ]
+        
+        self.undersampling_methods = [
+            'clustercentroids',   # ClusterCentroids
+            'randomundersampler', # RandomUnderSampler
+            'editednearestneighbours',  # EditedNearestNeighbours
+            'tomeklinks'          # TomekLinks
+        ]
+        
+        self.combination_methods = [
+            'smoteenn',      # SMOTEENN
+            'smotetomek'     # SMOTETomek
+        ]
+    
+    def get_supported_methods(self):
+        """Get all supported balancing methods."""
+        return {
+            'oversampling': self.oversampling_methods,
+            'undersampling': self.undersampling_methods,
+            'combination': self.combination_methods
+        }
+    
+    def is_valid_method(self, method):
+        """Check if a balancing method is supported."""
+        all_methods = (self.oversampling_methods + 
+                      self.undersampling_methods + 
+                      self.combination_methods)
+        return method.lower() in all_methods
+    
+    def balance_features(self, df_train, feats_train, target_column, method):
+        """
+        Balance features using the specified method.
+        
+        Args:
+            df_train (pd.DataFrame): Training dataframe with target labels
+            feats_train (np.ndarray or pd.DataFrame): Training features
+            target_column (str): Name of the target column
+            method (str): Balancing method to use
+            
+        Returns:
+            tuple: (balanced_df, balanced_features)
+        """
+        if not self.is_valid_method(method):
+            available_methods = (self.oversampling_methods + 
+                               self.undersampling_methods + 
+                               self.combination_methods)
+            self.util.error(
+                f"Unknown balancing algorithm: {method}. "
+                f"Available methods: {available_methods}"
+            )
+            return df_train, feats_train
+        
+        orig_size = len(df_train)
+        self.util.debug(f"Balancing features with: {method}")
+        self.util.debug(f"Original dataset size: {orig_size}")
+        
+        # Get original class distribution
+        orig_dist = df_train[target_column].value_counts().to_dict()
+        self.util.debug(f"Original class distribution: {orig_dist}")
+        
+        try:
+            # Apply the specified balancing method
+            X_res, y_res = self._apply_balancing_method(
+                feats_train, df_train[target_column], method
+            )
+            
+            # Create new balanced dataframe
+            balanced_df = pd.DataFrame({target_column: y_res})
+            
+            # If original dataframe has an index, try to preserve it
+            if hasattr(X_res, 'index'):
+                balanced_df.index = X_res.index
+            
+            new_size = len(balanced_df)
+            new_dist = balanced_df[target_column].value_counts().to_dict()
+            
+            self.util.debug(f"Balanced dataset size: {new_size} (was {orig_size})")
[…]
