+++ tmp.ini
@@ -1,51 +1,0 @@
-[EXP]
-root = ./results
-databases = ['dementianet', 'dementiabank']
-name = dementianet
-runs = 1
-epochs = 1
-save = True
-
-[DATA]
-dementianet = /home/bagus/github/nkululeko/data/dementianet/
-dementianet.type = csv
-dementianet.files = ['dementianet_train.csv', 'dementianet_val.csv', 'dementianet_test.csv']
-dementianet.mapping = {'nodementia': 'control', 'dementia': 'dementia'}
-dementianet.split_strategy = database
-dementiabank = /home/bagus/github/nkululeko/data/dementiabank/
-dementiabank.type = csv
-dementiabank.files = ['dementiabank_train.csv', 'dementiabank_val.csv', 'dementiabank_test.csv']
-dementiabank.mapping = {'control': 'control', 'dementia': 'dementia'}
-dementiabank.split_strategy = database
-target = dementia
-databases = ['dementianet']
-
-[FEATS]
-type = ['praat']
-scale = speaker
-balancing = ros
-
-[EXPL]
-feature_distributions = True
-model = ['xgb']
-permutation = True
-max_feats = 15
-sample_selection = all
-shap = True
-
-[MODEL]
-type = xgb
-save = True
-n_estimators = 100
-max_depth = 6
-learning_rate = 0.3
-subsample = 1.0
-
-[RESAMPLE]
-replace = True
-
-[PLOT]
-format = png
-
-[CROSSDB]
-
