### 2025-06-27 7e9e671
Update nkululeko/predict.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
