### 2025-06-27 7dbbdd6
Update nkululeko/reporting/reporter.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
