### 2025-06-27 5e91158
Remove .pyc files and add to .gitignore
