+++ nkululeko/tests/test_balancing.py
@@ -1,256 +1,0 @@
-#!/usr/bin/env python3
-"""
-Pytest test suite for comprehensive balancing methods testing
-using the new DataBalancer class
-
-Tests all balancing methods including:
-- ClusterCentroids (undersampling)
-- SMOTE (oversampling) 
-- ADASYN (oversampling)
-- RandomUnderSampler (undersampling)
-- SMOTEENN (combination)
-- And more...
-
-Run with:
-    pytest nkululeko/tests/test_balancing.py -v
-    or
-    pytest nkululeko/tests -v
-
-From the nkululeko root directory.
-"""
-
-import os
-import sys
-import numpy as np
-import pandas as pd
-import pytest
-
-# Import the modules we need to test
-from nkululeko.balance import DataBalancer
-from nkululeko import glob_conf
-from nkululeko.utils.util import Util
-
-
-@pytest.fixture
-def imbalanced_data():
-    """Create sample imbalanced data for testing"""
-    np.random.seed(42)
-    
-    # Create a synthetic dataset with class imbalance
-    # Majority class (label 0): 100 samples
-    # Minority class (label 1): 20 samples
-    majority_features = np.random.randn(100, 10)
-    minority_features = np.random.randn(20, 10) + 2  # Shift to make them distinguishable
-    
-    features = np.vstack([majority_features, minority_features])
-    labels = np.array([0] * 100 + [1] * 20)
-    
-    # Create DataFrames
-    df_train = pd.DataFrame({'target': labels})
-    feats_train = features
-    
-    return df_train, feats_train
-
-
-@pytest.fixture
-def mock_config():
-    """Mock the global configuration for testing"""
-    original_config = getattr(glob_conf, 'config', None)
-    
-    # Set up mock configuration
-    glob_conf.config = {
-        'FEATS': {'balancing': 'clustercentroids'},
-        'DATA': {'target': 'target'},
-        'MODEL': {'type': 'mlp'},
-        'EXP': {'epochs': '1'}
-    }
-    
-    yield glob_conf.config
-    
-    # Restore original configuration
-    if original_config is not None:
-        glob_conf.config = original_config
-
-
-class TestDataBalancer:
-    """Test suite for DataBalancer class"""
-    
-    def test_data_balancer_initialization(self):
-        """Test DataBalancer can be initialized"""
-        balancer = DataBalancer(random_state=42)
-        assert balancer is not None
-        assert hasattr(balancer, 'random_state')
-        assert balancer.random_state == 42
-    
-    def test_supported_methods(self):
-        """Test that supported methods are correctly reported"""
-        balancer = DataBalancer(random_state=42)
-        supported = balancer.get_supported_methods()
-        
-        assert 'oversampling' in supported
-        assert 'undersampling' in supported
-        assert 'combination' in supported
-        
-        # Check that ClusterCentroids is in undersampling methods
-        assert 'clustercentroids' in supported['undersampling']
-    
-    def test_is_valid_method(self):
-        """Test method validation"""
-        balancer = DataBalancer(random_state=42)
-        
-        # Test valid methods
-        assert balancer.is_valid_method('clustercentroids') == True
-        assert balancer.is_valid_method('smote') == True
-        assert balancer.is_valid_method('adasyn') == True
-        
-        # Test invalid method
-        assert balancer.is_valid_method('invalid_method') == False
-    
-    def test_clustercentroids_balancing(self, imbalanced_data, mock_config):
-        """Test ClusterCentroids balancing functionality"""
-        df_train, feats_train = imbalanced_data
-        
-        # Create balancer
-        balancer = DataBalancer(random_state=42)
-        
-        # Get original statistics
-        orig_train_size = feats_train.shape[0]
[…]
