### 2025-06-27 b79aea3
add balancing test
