+++ nkululeko/test_pretrain.py
@@ -1,0 +1,304 @@
+# test_pretrain.py
+import argparse
+import configparser
+import json
+import os.path
+
+import audeer
+import audiofile
+import audmetric
+import datasets
+import numpy as np
+import pandas as pd
+import torch
+import transformers
+
+import nkululeko.experiment as exp
+import nkululeko.glob_conf as glob_conf
+import nkululeko.models.finetune_model as fm
+from nkululeko.constants import VERSION
+from nkululeko.utils.util import Util
+
+
+def doit(config_file):
+    # test if the configuration file exists
+    if not os.path.isfile(config_file):
+        print(f"ERROR: no such file: {config_file}")
+        exit()
+
+    # load one configuration per experiment
+    config = configparser.ConfigParser()
+    config.read(config_file)
+
+    # create a new experiment
+    expr = exp.Experiment(config)
+    module = "test_pretrain"
+    expr.set_module(module)
+    util = Util(module)
+    util.debug(
+        f"running {expr.name} from config {config_file}, nkululeko version"
+        f" {VERSION}"
+    )
+
+    if util.config_val("EXP", "no_warnings", False):
+        import warnings
+
+        warnings.filterwarnings("ignore")
+
+    # load the data
+    expr.load_datasets()
+
+    # split into train and test
+    expr.fill_train_and_tests()
+    util.debug(f"train shape : {expr.df_train.shape}, test shape:{expr.df_test.shape}")
+
+    model_root = util.get_path("model_dir")
+    log_root = audeer.mkdir("log")
+    torch_root = audeer.path(model_root, "torch")
+
+    metrics_gender = {
+        "UAR": audmetric.unweighted_average_recall,
+        "ACC": audmetric.accuracy,
+    }
+
+    sampling_rate = 16000
+    max_duration_sec = 8.0
+
+    model_path = "facebook/wav2vec2-large-robust-ft-swbd-300h"
+    num_layers = None
+
+    os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
+    os.environ["CUDA_VISIBLE_DEVICES"] = "0"
+
+    batch_size = 16
+    accumulation_steps = 4
+    # create dataset
+
+    dataset = {}
+    target_name = glob_conf.target
+    data_sources = {
+        "train": pd.DataFrame(expr.df_train[target_name]),
+        "dev": pd.DataFrame(expr.df_test[target_name]),
+    }
+
+    for split in ["train", "dev"]:
+        df = data_sources[split]
+        df[target_name] = df[target_name].astype("float")
+
+        y = pd.Series(
+            data=df.itertuples(index=False, name=None),
+            index=df.index,
+            dtype=object,
+            name="labels",
+        )
+
+        y.name = "targets"
+        df = y.reset_index()
+        df.start = df.start.dt.total_seconds()
+        df.end = df.end.dt.total_seconds()
+
+        print(f"{split}: {len(df)}")
+
+        ds = datasets.Dataset.from_pandas(df)
+        dataset[split] = ds
+
+    dataset = datasets.DatasetDict(dataset)
+
+    # load pre-trained model
+    le = glob_conf.label_encoder
+    mapping = dict(zip(le.classes_, range(len(le.classes_))))
+    target_mapping = {k: int(v) for k, v in mapping.items()}
+    target_mapping_reverse = {value: key for key, value in target_mapping.items()}
+
+    config = transformers.AutoConfig.from_pretrained(
+        model_path,
+        num_labels=len(target_mapping),
+        label2id=target_mapping,
+        id2label=target_mapping_reverse,
+        finetuning_task=target_name,
+    )
+    if num_layers is not None:
[…]
