### 2025-06-28 edb2ce3
add balancing examples
