+++ examples/exp_emodb_balancing.ini
@@ -15,6 +15,8 @@ labels = ['angry', 'happy', 'neutral', 'sad']
 target = emotion
 [FEATS]
 type = ['os']
+; balancing = ros, smote, adasynn, smoteenc, borderlinsmote, clustercentroids, 
+; randomundersampler, editednearestneighbours, tomeklinks, smoteenc, smotetomek
 balancing = adasyn
 [MODEL]
 type = xgb
