+++ examples/exp_balancing_smote.ini
@@ -1,0 +1,33 @@
+[EXP]
+root = ./examples/results/
+name = exp_polish_smote_balancing
+runs = 1
+epochs = 1
+save = True
+
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/polish/polish_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/polish/polish_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/polish/polish_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+
+[FEATS]
+type = ['os']
+balancing = smote
+
+[MODEL]
+type = xgb
+save = True
+
+[PLOT]
+name = smote_balancing_results
