### 2025-06-28 0551f50
fix run script
