+++ nkululeko/tests/test_balancing.py
@@ -1,51 +1,36 @@
 #!/usr/bin/env python3
 """
-Pytest test suite for comprehensive balancing methods testing
-using the new DataBalancer class
+Simple and comprehensive test suite for all balancing methods in DataBalancer.
 
-Tests all balancing methods including:
-- ClusterCentroids (undersampling)
-- SMOTE (oversampling) 
-- ADASYN (oversampling)
-- RandomUnderSampler (undersampling)
-- SMOTEENN (combination)
-- And more...
+Tests all 11 balancing methods from balance.py:
 
-Run with:
-    pytest nkululeko/tests/test_balancing.py -v
-    or
-    pytest nkululeko/tests -v
+Oversampling (5): ros, smote, adasyn, borderlinesmote, svmsmote
+Undersampling (4): clustercentroids, randomundersampler, editednearestneighbours, tomeklinks  
+Combination (2): smoteenn, smotetomek
 
-From the nkululeko root directory.
+Run with: pytest nkululeko/tests/test_balancing.py -v
 """
 
-import os
-import sys
 import numpy as np
 import pandas as pd
 import pytest
-
-# Import the modules we need to test
 from nkululeko.balance import DataBalancer
-from nkululeko import glob_conf
-from nkululeko.utils.util import Util
+import nkululeko.glob_conf as glob_conf
 
 
 @pytest.fixture
-def imbalanced_data():
-    """Create sample imbalanced data for testing"""
+def sample_data():
+    """Create sample imbalanced data that works with all methods"""
     np.random.seed(42)
     
-    # Create a synthetic dataset with class imbalance
-    # Majority class (label 0): 100 samples
-    # Minority class (label 1): 20 samples
-    majority_features = np.random.randn(100, 10)
-    minority_features = np.random.randn(20, 10) + 2  # Shift to make them distinguishable
+    # Majority class: 100 samples, Minority class: 25 samples
+    # Well-separated for better algorithm performance
+    majority_features = np.random.randn(100, 10) 
+    minority_features = np.random.randn(25, 10) + 3  # Good separation
     
     features = np.vstack([majority_features, minority_features])
-    labels = np.array([0] * 100 + [1] * 20)
+    labels = np.array([0] * 100 + [1] * 25)
     
-    # Create DataFrames
     df_train = pd.DataFrame({'target': labels})
     feats_train = features
     
@@ -54,104 +39,153 @@ def imbalanced_data():
 
 @pytest.fixture
 def mock_config():
-    """Mock the global configuration for testing"""
+    """Mock configuration for testing"""
     original_config = getattr(glob_conf, 'config', None)
     
-    # Set up mock configuration
     glob_conf.config = {
-        'FEATS': {'balancing': 'clustercentroids'},
+        'FEATS': {'balancing': 'smote'},
         'DATA': {'target': 'target'},
-        'MODEL': {'type': 'mlp'},
-        'EXP': {'epochs': '1'}
+        'MODEL': {'type': 'mlp'}
     }
     
     yield glob_conf.config
     
-    # Restore original configuration
     if original_config is not None:
         glob_conf.config = original_config
 
 
 class TestDataBalancer:
-    """Test suite for DataBalancer class"""
+    """Simple test suite for DataBalancer - tests all 11 methods"""
     
-    def test_data_balancer_initialization(self):
-        """Test DataBalancer can be initialized"""
+    def test_initialization(self):
+        """Test 1: DataBalancer can be initialized"""
         balancer = DataBalancer(random_state=42)
         assert balancer is not None
-        assert hasattr(balancer, 'random_state')
         assert balancer.random_state == 42
     
-    def test_supported_methods(self):
-        """Test that supported methods are correctly reported"""
-        balancer = DataBalancer(random_state=42)
[…]
