+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.95.0"
+VERSION="0.95.1"
 SAMPLING_RATE = 16000
