### 2025-06-30 f79a471
Add hyperparameter optimization module
- Create nkululeko/optim.py with OptimizationRunner class
- Support parameter ranges, steps, and discrete choices for MLP models
- Add example configuration files for Polish dataset testing
- Update ini_file.md documentation with OPTIM section
- Implement grid search over hyperparameter combinations
- Support both traditional ML and neural network model types

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
