+++ nkululeko/optim.py
@@ -1,0 +1,253 @@
+#!/usr/bin/env python3
+
+import argparse
+import ast
+import configparser
+import itertools
+import os
+
+from nkululeko.constants import VERSION
+from nkululeko.utils.util import Util
+
+
+class OptimizationRunner:
+    """Hyperparameter optimization runner for nkululeko experiments."""
+
+    def __init__(self, config):
+        self.config = config
+        self.util = Util("optim")
+        self.results = []
+
+    def parse_optim_params(self):
+        """Parse OPTIM section parameters into search spaces."""
+        if "OPTIM" not in self.config:
+            self.util.error("No [OPTIM] section found in configuration")
+
+        optim_config = self.config["OPTIM"]
+        self.model_type = optim_config.get("model", "mlp")
+
+        param_specs = {}
+        for key, value in optim_config.items():
+            if key == "model":
+                continue
+            param_specs[key] = self._parse_param_spec(key, value)
+
+        return param_specs
+
+    def _parse_param_spec(self, param_name, param_value):
+        """Parse individual parameter specification."""
+        try:
+            parsed = ast.literal_eval(param_value)
+        except (ValueError, SyntaxError):
+            if isinstance(param_value, str):
+                return [param_value]
+            return param_value
+
+        if isinstance(parsed, tuple):
+            if len(parsed) == 2:
+                return self._generate_range(parsed[0], parsed[1], param_name)
+            elif len(parsed) == 3:
+                return self._generate_range_with_step(
+                    parsed[0], parsed[1], parsed[2], param_name
+                )
+        elif isinstance(parsed, list):
+            return parsed
+        else:
+            return [parsed]
+
+    def _generate_range(self, min_val, max_val, param_name):
+        """Generate parameter range based on parameter type."""
+        if param_name in ["nlayers"]:
+            return list(range(min_val, max_val + 1))
+        elif param_name in ["nnodes", "bs"]:
+            result = []
+            current = min_val
+            while current <= max_val:
+                result.append(current)
+                current *= 2
+            return result
+        elif param_name in ["lr", "do"]:
+            result = []
+            current = min_val
+            while current <= max_val + 0.0001:
+                result.append(round(current, 4))
+                current += 0.0001
+            return result
+        else:
+            return list(range(min_val, max_val + 1))
+
+    def _generate_range_with_step(self, min_val, max_val, step, param_name):
+        """Generate parameter range with explicit step."""
+        if isinstance(step, float) or isinstance(min_val, float):
+            result = []
+            current = min_val
+            while current <= max_val + step/2:
+                result.append(round(current, 4))
+                current += step
+            return result
+        else:
+            return list(range(min_val, max_val + 1, step))
+
+    def generate_param_combinations(self, param_specs):
+        """Generate all parameter combinations for grid search."""
+        param_names = list(param_specs.keys())
+        param_values = list(param_specs.values())
+
+        combinations = []
+        for combo in itertools.product(*param_values):
+            param_dict = dict(zip(param_names, combo))
+            combinations.append(param_dict)
+
+        return combinations
+
+    def run_optimization(self):
+        """Run hyperparameter optimization."""
+        param_specs = self.parse_optim_params()
+        combinations = self.generate_param_combinations(param_specs)
+
+        self.util.debug(
+            f"Starting optimization with {len(combinations)} parameter combinations"
+        )
+
+        best_result = None
+        best_params = None
+        best_score = -float('inf') if self.util.high_is_good() else float('inf')
+
+        for i, params in enumerate(combinations):
[…]
