+++ examples/exp_polish_optim_mlp.ini
@@ -1,0 +1,37 @@
+[EXP]
+root = ./examples/results/
+name = exp_polish_optim_mlp
+runs = 1
+epochs = 5
+
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/polish/polish_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/polish/polish_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/polish/polish_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+labels = ['anger', 'neutral', 'fear']
+
+[FEATS]
+type = ['os']
+
+[OPTIM]
+model = mlp
+nlayers = (1, 3)
+nnodes = (16, 256)
+lr = (0.0001, 0.1, 0.0001)
+bs = (2, 256)
+loss = ["cross", "f1"]
+do = (0.1, 0.5, 0.1)
+
+[RESAMPLE]
+replace = True
