### 2025-06-30 965c02c
Add example configurations for SVM, XGB, and KNN optimization
- exp_polish_optim_svm.ini: SVM with C_val and kernel parameters
- exp_polish_optim_xgb.ini: XGBoost with n_estimators, max_depth, learning_rate, subsample
- exp_polish_optim_knn.ini: KNN with K_val and KNN_weights parameters
- Demonstrates traditional ML model support through tuning_params mechanism

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
