+++ examples/exp_polish_optim_xgb.ini
@@ -1,0 +1,35 @@
+[EXP]
+root = ./examples/results/
+name = exp_polish_optim_xgb
+runs = 1
+epochs = 1
+
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/polish/polish_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/polish/polish_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/polish/polish_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+labels = ['anger', 'neutral', 'fear']
+
+[FEATS]
+type = ['os']
+
+[OPTIM]
+model = xgb
+n_estimators = [50, 100, 200]
+max_depth = [3, 6, 9]
+learning_rate = [0.01, 0.1, 0.2]
+subsample = [0.8, 0.9, 1.0]
+
+[RESAMPLE]
+replace = True
