+++ examples/exp_polish_optim_knn.ini
@@ -1,0 +1,33 @@
+[EXP]
+root = ./examples/results/
+name = exp_polish_optim_knn
+runs = 1
+epochs = 1
+
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/polish/polish_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/polish/polish_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/polish/polish_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+labels = ['anger', 'neutral', 'fear']
+
+[FEATS]
+type = ['os']
+
+[OPTIM]
+model = knn
+K_val = [3, 5, 7, 9, 11]
+KNN_weights = ["uniform", "distance"]
+
+[RESAMPLE]
+replace = True
