### 2025-06-30 4da6b31
Fix MODEL section initialization bug
- Add _ensure_model_section() method to initialize MODEL section early
- Call _ensure_model_section() at start of run_optimization()
- Prevents KeyError: 'MODEL' when runmanager accesses config
- Fixes issue reported by user when running exp_polish_optim_mlp.ini

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
