### 2025-06-30 f12d158
Move MODEL section initialization to _run_single_experiment
- Move _ensure_model_section() call to happen right before Experiment() creation
- Remove early call from run_optimization()
- Attempt to fix timing issue with global config access

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
