### 2025-06-30 2f30a47
Fix KeyError: 'MODEL' by initializing MODEL section before Experiment creation
- Move MODEL section initialization to happen before creating Experiment object
- This prevents runmanager from accessing glob_conf.config['MODEL']['type'] before it exists
- Inline the initialization logic instead of using _ensure_model_section() method
- Fixes the persistent KeyError reported by user when running optimization module

Co-Authored-By: Bagus Tris Atmaja <bagus.tris@naist.ac.jp>
