+++ nkululeko/optim.py
@@ -82,7 +82,7 @@ class OptimizationRunner:
         if isinstance(step, float) or isinstance(min_val, float):
             result = []
             current = min_val
-            while current <= max_val + step/2:
+            while current <= max_val + step / 2:
                 result.append(round(current, 4))
                 current += step
             return result
@@ -112,7 +112,7 @@ class OptimizationRunner:
 
         best_result = None
         best_params = None
-        best_score = -float('inf') if self.util.high_is_good() else float('inf')
+        best_score = -float("inf") if self.util.high_is_good() else float("inf")
 
         for i, params in enumerate(combinations):
             self.util.debug(f"Testing combination {i+1}/{len(combinations)}: {params}")
@@ -163,12 +163,12 @@ class OptimizationRunner:
             self._update_mlp_params(params)
         else:
             self._update_traditional_ml_params(params)
-    
+
     def _ensure_model_section(self):
         """Ensure MODEL section exists with basic configuration."""
         if "MODEL" not in self.config:
             self.config.add_section("MODEL")
-        
+
         if "type" not in self.config["MODEL"]:
             self.config["MODEL"]["type"] = self.model_type
 
@@ -205,23 +205,26 @@ class OptimizationRunner:
     def _run_single_experiment(self):
         """Run a single experiment with current configuration."""
         import nkululeko.experiment as exp
-        
-        self._ensure_model_section()
-        
+
+        if "MODEL" not in self.config:
+            self.config.add_section("MODEL")
+        if "type" not in self.config["MODEL"]:
+            self.config["MODEL"]["type"] = self.model_type
+
         expr = exp.Experiment(self.config)
         expr.set_module("optim")
-        
+
         expr.load_datasets()
-        
+
         expr.fill_train_and_tests()
-        
+
         expr.extract_feats()
-        
+
         expr.init_runmanager()
-        
+
         reports, last_epochs = expr.run()
         result = expr.get_best_report(reports).result.test
-        
+
         return result, int(min(last_epochs))
 
 
@@ -233,15 +236,15 @@ def doit(config_file):
 
     config = configparser.ConfigParser()
     config.read(config_file)
-    
+
     optimizer = OptimizationRunner(config)
-    
+
     best_params, best_result, all_results = optimizer.run_optimization()
-    
+
     print("OPTIMIZATION COMPLETE")
     print(f"Best parameters: {best_params}")
     print(f"Best result: {best_result}")
-    
+
     return best_params, best_result
 
 
