+++ examples/exp_polish_optim_svm.ini
@@ -24,6 +24,10 @@ labels = ['anger', 'neutral', 'fear']
 [FEATS]
 type = ['os']
 
+[MODEL]
+type = svm
+C_val = 0.001
+
 [OPTIM]
 model = svm
 C_val = [0.1, 1.0, 10.0, 100.0]
