+++ examples/exp_polish_optim_xgb.ini
@@ -24,6 +24,9 @@ labels = ['anger', 'neutral', 'fear']
 [FEATS]
 type = ['os']
 
+[MODEL]
+type = xgb
+
 [OPTIM]
 model = xgb
 n_estimators = [50, 100, 200]
