### 2025-06-30 4e17767
rm untested optim files
