+++ examples/exp_polish_optim_minimal.ini
@@ -1,37 +1,0 @@
-[EXP]
-root = ./examples/results/
-name = exp_polish_optim_minimal
-runs = 1
-epochs = 1
-
-[DATA]
-databases = ['train', 'dev', 'test']
-train = ./data/polish/polish_train.csv
-train.type = csv
-train.absolute_path = False
-train.split_strategy = train
-dev = ./data/polish/polish_dev.csv
-dev.type = csv
-dev.absolute_path = False
-dev.split_strategy = train
-test = ./data/polish/polish_test.csv
-test.type = csv
-test.absolute_path = False
-test.split_strategy = test
-target = emotion
-labels = ['anger', 'neutral', 'fear']
-
-[FEATS]
-type = ['os']
-
-[OPTIM]
-model = mlp
-nlayers = [1]
-nnodes = [16]
-lr = [0.001]
-bs = [16]
-loss = ["cross"]
-do = [0.1]
-
-[RESAMPLE]
-replace = True
