### 2025-06-30 e93ec32
update optime to be intelligent
