+++ nkululeko/optim.py
@@ -7,6 +7,8 @@ import itertools
 import os
 import sys
 
+import numpy as np
+
 from nkululeko.constants import VERSION
 from nkululeko.utils.util import Util
 
@@ -19,6 +21,10 @@ class OptimizationRunner:
         self.util = Util("optim")
         self.results = []
         self.model_type = None  # Will be set when parsing OPTIM params
+        # New: Optimization strategy configuration
+        self.search_strategy = None
+        self.n_iter = None
+        self.cv_folds = None
 
     def parse_optim_params(self):
         """Parse OPTIM section parameters into search spaces."""
@@ -27,10 +33,15 @@ class OptimizationRunner:
 
         optim_config = self.config["OPTIM"]
         self.model_type = optim_config.get("model", "mlp")
+        
+        # Parse optimization strategy settings
+        self.search_strategy = optim_config.get("search_strategy", "grid")
+        self.n_iter = int(optim_config.get("n_iter", "50"))
+        self.cv_folds = int(optim_config.get("cv_folds", "3"))
 
         param_specs = {}
         for key, value in optim_config.items():
-            if key == "model":
+            if key in ["model", "search_strategy", "n_iter", "cv_folds"]:
                 continue
             param_specs[key] = self._parse_param_spec(key, value)
 
@@ -45,6 +56,13 @@ class OptimizationRunner:
             if isinstance(param_value, str):
                 return [param_value]
             return param_value
+
+        # Check for inefficient learning rate ranges and suggest better alternatives
+        if param_name == "lr" and isinstance(parsed, tuple) and len(parsed) == 3:
+            min_val, max_val, step = parsed
+            if step <= 0.0001 and (max_val - min_val) / step > 20:
+                self.util.debug(f"WARNING: Learning rate range {param_value} will generate {int((max_val - min_val) / step)} values!")
+                self.util.debug(f"Consider using discrete values like [0.0001, 0.001, 0.01, 0.1] or range (0.0001, 0.1) for log-scale sampling")
 
         if isinstance(parsed, tuple):
             if len(parsed) == 2:
@@ -72,14 +90,23 @@ class OptimizationRunner:
                 result.append(current)
                 current *= 2
             return result
-        elif param_name in ["lr", "do"]:
-            # For learning rate and dropout, generate 10-20 steps instead of 100
-            num_steps = 10
+        elif param_name in ["lr"]:
+            # For learning rate, use logarithmic scale sampling (more practical)
+            # Generate 5-8 values on log scale between min and max
+            num_samples = min(8, max(5, int(np.log10(max_val / min_val) * 2)))
+            log_min = np.log10(min_val)
+            log_max = np.log10(max_val)
+            log_values = np.linspace(log_min, log_max, num_samples)
+            result = [round(10**log_val, 6) for log_val in log_values]
+            return result
+        elif param_name in ["do"]:
+            # For dropout, generate reasonable steps
+            num_steps = 5
             step = (max_val - min_val) / num_steps
             result = []
             current = min_val
             while current <= max_val + step / 2:
-                result.append(round(current, 4))
+                result.append(round(current, 2))
                 current += step
             return result
         else:
@@ -176,6 +203,168 @@ class OptimizationRunner:
         self.save_results()
 
         return best_params, best_result, self.results
+
+    def run_intelligent_optimization(self):
+        """Run intelligent hyperparameter optimization using scikit-learn methods."""
+        param_specs = self.parse_optim_params()
+        
+        if not param_specs:
+            self.util.error("No optimization parameters found in [OPTIM] section")
+            return None, None, []
+        
+        # Use scikit-learn's built-in hyperparameter optimization
+        if self.model_type in ["svm", "svr", "xgb", "xgr", "knn", "tree"]:
[…]
