+++ examples/exp_polish_optim_svm.ini
@@ -29,9 +29,16 @@ type = svm
 C_val = 0.001
 
 [OPTIM]
+; model = svm
+; C_val = [0.1, 1.0, 10.0, 100.0]
+; kernel = ["linear", "rbf", "poly"]
 model = svm
-C_val = [0.1, 1.0, 10.0, 100.0]
+search_strategy = halving_random
+n_iter = 15
+cv_folds = 3
+C = [0.1, 1.0, 10.0, 100.0, 1000.0]
 kernel = ["linear", "rbf", "poly"]
+gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
 
 [RESAMPLE]
 replace = True
