+++ examples/exp_polish_optim_mlp.ini
@@ -26,12 +26,15 @@ type = ['os']
 
 [OPTIM]
 model = mlp
-nlayers = (1, 3)
-nnodes = (16, 256)
-lr = (0.0001, 0.1, 0.0001)
-bs = (2, 256)
+search_strategy = random
+n_iter = 20
+cv_folds = 3
+nlayers = [1, 2, 3]
+nnodes = [16, 32, 64, 128]
+lr = [0.0001, 0.001, 0.01, 0.1]
+bs = [4, 8, 16, 32]
 loss = ["cross", "f1"]
-do = (0.1, 0.5, 0.1)
+do = [0.1, 0.3, 0.5]
 
 [RESAMPLE]
 replace = True
