### 2025-06-30 afb5556
update optim example
