+++ examples/exp_polish_optim_xgb.ini
@@ -23,9 +23,15 @@ labels = ['anger', 'neutral', 'fear']
 
 [FEATS]
 type = ['os']
+scale = standard
+balancing = smoteenn
 
 [MODEL]
 type = xgb
+n_estimators = 50
+max_depth = 3
+learning_rate = 0.2
+subsample = 0.8
 
 [OPTIM]
 model = xgb
