+++ examples/exp_polish_optim_svm.ini
@@ -23,22 +23,27 @@ labels = ['anger', 'neutral', 'fear']
 
 [FEATS]
 type = ['os']
+scale = robust
+balancing = smoteenn
 
 [MODEL]
 type = svm
-C_val = 0.001
+C_val = 100.0
 
 [OPTIM]
-; model = svm
-; C_val = [0.1, 1.0, 10.0, 100.0]
-; kernel = ["linear", "rbf", "poly"]
+; the following three lines used grid search
 model = svm
-search_strategy = halving_random
-n_iter = 15
-cv_folds = 3
-C = [0.1, 1.0, 10.0, 100.0, 1000.0]
+C_val = [0.1, 1.0, 10.0, 100.0] 
 kernel = ["linear", "rbf", "poly"]
-gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
+
+; the following three lines used either: random, halving_grid, halving_random
+; model = svm
+; search_strategy = halving_grid
+; n_iter = 15
+; cv_folds = 3
+; C = [0.1, 1.0, 10.0, 100.0, 1000.0]
+; kernel = ["linear", "rbf", "poly"]
+; gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
 
 [RESAMPLE]
 replace = True
