### 2025-06-30 49846f9
update optim examples
