+++ examples/exp_polish_optim_xgb.ini
@@ -34,7 +34,17 @@ learning_rate = 0.2
 subsample = 0.8
 
 [OPTIM]
+; model = xgb
+; n_estimators = [50, 100, 200]
+; max_depth = [3, 6, 9]
+; learning_rate = [0.01, 0.1, 0.2]
+; subsample = [0.8, 0.9, 1.0]
+
+; the following three lines used either: random, halving_grid, halving_random
 model = xgb
+search_strategy = halving_grid
+n_iter = 15
+cv_folds = 3
 n_estimators = [50, 100, 200]
 max_depth = [3, 6, 9]
 learning_rate = [0.01, 0.1, 0.2]
