### 2025-06-30 bc74cc7
modified:   examples/exp_polish_optim_mlp.ini 	modified:   examples/exp_polish_optim_svm.ini 	new file:   examples/exp_polish_optim_svm_intelligent.ini 	modified:   examples/exp_polish_optim_xgb.ini 	modified:   ini_file.md 	modified:   nkululeko/optim.py 	new file:   nkululeko/tests/test_optim.py
