+++ nkululeko/optim.py
@@ -23,9 +23,10 @@ class OptimizationRunner:
         self.results = []
         self.model_type = None  # Will be set when parsing OPTIM params
         # New: Optimization strategy configuration
-        self.search_strategy = None
-        self.n_iter = None
-        self.cv_folds = None
+        self.search_strategy = "grid"  # Default values
+        self.n_iter = 50
+        self.cv_folds = 3
+        self.metric = "accuracy"
 
     def parse_optim_params(self):
         """Parse OPTIM section parameters into search spaces."""
@@ -34,15 +35,18 @@ class OptimizationRunner:
 
         optim_config = self.config["OPTIM"]
         self.model_type = optim_config.get("model", "mlp")
-        
+
         # Parse optimization strategy settings
         self.search_strategy = optim_config.get("search_strategy", "grid")
         self.n_iter = int(optim_config.get("n_iter", "50"))
         self.cv_folds = int(optim_config.get("cv_folds", "3"))
+        self.metric = optim_config.get("metric", "accuracy").lower()
+
+        self.util.debug(f"Parsed metric from config: '{self.metric}'")  # Debug line
 
         param_specs = {}
         for key, value in optim_config.items():
-            if key in ["model", "search_strategy", "n_iter", "cv_folds"]:
+            if key in ["model", "search_strategy", "n_iter", "cv_folds", "metric"]:
                 continue
             param_specs[key] = self._parse_param_spec(key, value)
 
@@ -53,7 +57,9 @@ class OptimizationRunner:
         try:
             parsed = ast.literal_eval(param_value)
         except (ValueError, SyntaxError) as e:
-            self.util.debug(f"Could not parse parameter {param_name}={param_value} as literal, treating as string: {e}")
+            self.util.debug(
+                f"Could not parse parameter {param_name}={param_value} as literal, treating as string: {e}"
+            )
             if isinstance(param_value, str):
                 return [param_value]
             return param_value
@@ -62,8 +68,12 @@ class OptimizationRunner:
         if param_name == "lr" and isinstance(parsed, tuple) and len(parsed) == 3:
             min_val, max_val, step = parsed
             if step <= 0.0001 and (max_val - min_val) / step > 20:
-                self.util.debug(f"WARNING: Learning rate range {param_value} will generate {int((max_val - min_val) / step)} values!")
-                self.util.debug(f"Consider using discrete values like [0.0001, 0.001, 0.01, 0.1] or range (0.0001, 0.1) for log-scale sampling")
+                self.util.debug(
+                    f"WARNING: Learning rate range {param_value} will generate {int((max_val - min_val) / step)} values!"
+                )
+                self.util.debug(
+                    "Consider using discrete values like [0.0001, 0.001, 0.01, 0.1] or range (0.0001, 0.1) for log-scale sampling"
+                )
 
         if isinstance(parsed, tuple):
             if len(parsed) == 2:
@@ -73,7 +83,9 @@ class OptimizationRunner:
                     parsed[0], parsed[1], parsed[2], param_name
                 )
             else:
-                self.util.error(f"Invalid tuple format for parameter {param_name}: {param_value}. Expected (min, max) or (min, max, step)")
+                self.util.error(
+                    f"Invalid tuple format for parameter {param_name}: {param_value}. Expected (min, max) or (min, max, step)"
+                )
                 return [parsed[0]]  # Fallback to first value
         elif isinstance(parsed, list):
             return parsed
@@ -115,7 +127,11 @@ class OptimizationRunner:
 
     def _generate_range_with_step(self, min_val, max_val, step, param_name):
         """Generate parameter range with explicit step."""
-        if isinstance(step, float) or isinstance(min_val, float) or isinstance(max_val, float):
+        if (
[…]
