+++ nkululeko/tests/test_optim.py
@@ -1,0 +1,176 @@
+import pytest
+from unittest.mock import MagicMock, patch
+from nkululeko.optim import OptimizationRunner
+
+@pytest.fixture
+def mock_config():
+    # Minimal configparser.ConfigParser mock
+    config = MagicMock()
+    config.__contains__.side_effect = lambda x: x in ["OPTIM", "MODEL", "DATA"]
+    config.__getitem__.side_effect = lambda x: {
+        "OPTIM": {"model": "svm", "search_strategy": "grid", "n_iter": "2", "cv_folds": "2"},
+        "MODEL": {"type": "svm"},
+        "DATA": {"target": "label"}
+    }[x]
+    config.get.side_effect = lambda section, option, fallback=None: {
+        ("MODEL", "tuning_params"): None,
+        ("DATA", "target"): "label"
+    }.get((section, option), fallback)
+    config.add_section = MagicMock()
+    config.remove_option = MagicMock()
+    config.set = MagicMock()
+    return config
+
+@pytest.fixture
+def runner(mock_config):
+    runner = OptimizationRunner(mock_config)
+    runner.util = MagicMock()
+    runner.util.high_is_good.return_value = True
+    runner.util.exp_is_classification.return_value = True
+    runner.util.debug = MagicMock()
+    runner.util.error = MagicMock()
+    runner.save_results = MagicMock()
+    runner.search_strategy = "grid"
+    runner.n_iter = 2
+    runner.cv_folds = 2
+    runner.model_type = "svm"
+    return runner
+
+@pytest.fixture
+def param_specs():
+    return {"C": [0.1, 1.0], "kernel": ["linear", "rbf"]}
+
+def test_run_sklearn_optimization_grid(runner, param_specs):
+    with patch("sklearn.model_selection.GridSearchCV") as mock_GridSearchCV, \
+         patch("nkululeko.models.model.Model") as mock_Model, \
+         patch("nkululeko.glob_conf.config", runner.config), \
+         patch("nkululeko.models.model_svm.SVM_model") as mock_SVM:
+        
+        # Mock the experiment module and its Experiment class
+        mock_exp_module = MagicMock()
+        mock_expr = MagicMock()
+        mock_expr.df_train = {"label": [0, 1, 0, 1]}
+        mock_expr.df_test = {}
+        mock_expr.feats_train = [[1, 2], [2, 3], [3, 4], [4, 5]]
+        mock_expr.feats_test = [[1, 2], [2, 3]]
+        mock_exp_module.Experiment.return_value = mock_expr
+        
+        # Mock sys.modules to return our mock when importing nkululeko.experiment
+        with patch.dict('sys.modules', {'nkululeko.experiment': mock_exp_module}):
+            mock_model_instance = MagicMock()
+            # Create a mock classifier that sklearn recognizes
+            mock_clf = MagicMock()
+            mock_clf.__sklearn_tags__ = MagicMock(return_value=MagicMock(estimator_type="classifier"))
+            mock_model_instance.clf = mock_clf
+            mock_Model.create.return_value = mock_model_instance
+            mock_SVM.return_value = mock_model_instance
+
+            # Mock GridSearchCV
+            mock_search = MagicMock()
+            mock_search.best_params_ = {"C": 1.0, "kernel": "linear"}
+            mock_search.best_score_ = 0.9
+            mock_search.cv_results_ = {
+                "params": [{"C": 0.1, "kernel": "linear"}, {"C": 1.0, "kernel": "linear"}],
+                "mean_test_score": [0.8, 0.9]
+            }
+            mock_GridSearchCV.return_value = mock_search
+
+            best_params, best_score, all_results = runner._run_sklearn_optimization(param_specs)
+
+            assert best_params == {"C": 1.0, "kernel": "linear"}
+            assert best_score == 0.9
[…]
