+++ examples/exp_polish_optim_svm_intelligent.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_optim_xgb
+name = exp_polish_optim_svm_intelligent
 runs = 1
 epochs = 1
 
@@ -23,32 +23,18 @@ labels = ['anger', 'neutral', 'fear']
 
 [FEATS]
 type = ['os']
-scale = standard
-balancing = smoteenn
 
 [MODEL]
-type = xgb
-n_estimators = 200
-max_depth = 9
-learning_rate = 0.01
-subsample = 0.8
+type = svm
 
 [OPTIM]
-; model = xgb
-; n_estimators = [50, 100, 200]
-; max_depth = [3, 6, 9]
-; learning_rate = [0.01, 0.1, 0.2]
-; subsample = [0.8, 0.9, 1.0]
-
-; the following three lines used either: random, halving_grid, halving_random
-model = xgb
-search_strategy = halving_grid
+model = svm
+search_strategy = halving_random
 n_iter = 15
 cv_folds = 3
-n_estimators = [50, 100, 200]
-max_depth = [3, 6, 9]
-learning_rate = [0.01, 0.1, 0.2]
-subsample = [0.8, 0.9, 1.0]
+C = [0.1, 1.0, 10.0, 100.0, 1000.0]
+kernel = ["linear", "rbf", "poly"]
+gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
 
 [RESAMPLE]
 replace = True

[…]
