+++ examples/exp_polish_optim_svm.ini
@@ -28,7 +28,8 @@ balancing = smoteenn
 
 [MODEL]
 type = svm
-C_val = 100.0
+C_val = 0.1
+kernel = linear
 
 [OPTIM]
 ; the following three lines used grid search
