+++ examples/exp_polish_optim_xgb.ini
@@ -3,6 +3,7 @@ root = ./examples/results/
 name = exp_polish_optim_xgb
 runs = 1
 epochs = 1
+random_seed = 42
 
 [DATA]
 databases = ['train', 'dev', 'test']
@@ -45,6 +46,7 @@ model = xgb
 search_strategy = halving_grid
 n_iter = 15
 cv_folds = 3
+random_state = 42
 n_estimators = [50, 100, 200]
 max_depth = [3, 6, 9]
 learning_rate = [0.01, 0.1, 0.2]
