### 2025-06-30 0e3261e
modified:   README.md 	modified:   docs/source/index.rst 	new file:   docs/source/optim.md 	new file:   docs/source/optim_quickstart.md 	modified:   examples/exp_polish_optim_mlp.ini 	modified:   examples/exp_polish_optim_svm.ini 	modified:   examples/exp_polish_optim_xgb.ini 	modified:   ini_file.md 	modified:   nkululeko/optim.py
