+++ nkululeko/optim.py
@@ -177,15 +177,13 @@ class OptimizationRunner:
             self.util.error("No optimization parameters found in [OPTIM] section")
             return None, None, []
 
-        # Use scikit-learn's optimization for compatible models, otherwise use manual grid search
-        if self.model_type in ["svm", "svr", "xgb", "xgr", "knn", "knn_reg", "tree", "tree_reg", "bayes", "lin_reg"]:
-            return self._run_sklearn_optimization(param_specs)
-        else:
-            # Use manual optimization for neural networks and other custom models
-            return self._run_manual_optimization(param_specs)
+        # Always use manual optimization to ensure consistent evaluation pipeline
+        # This prevents discrepancies between CV and final evaluation
+        self.util.debug("Using manual optimization for consistent evaluation pipeline")
+        return self._run_manual_optimization(param_specs)
 
     def _run_manual_optimization(self, param_specs):
-        """Run manual grid search optimization for models not compatible with sklearn."""
+        """Run manual grid search optimization with consistent evaluation pipeline."""
         combinations = self.generate_param_combinations(param_specs)
 
         if not combinations:
@@ -196,6 +194,18 @@ class OptimizationRunner:
             f"Starting manual optimization with {len(combinations)} parameter combinations"
         )
 
+        # Check if we should use cross-validation or train-test split
+        use_cv = self.search_strategy in ["grid_cv", "random_cv"] or (
+            hasattr(self, 'use_cv_in_manual') and self.use_cv_in_manual
+        )
+
+        if use_cv:
+            return self._run_manual_cv_optimization(combinations, param_specs)
+        else:
+            return self._run_manual_train_test_optimization(combinations)
+
+    def _run_manual_train_test_optimization(self, combinations):
+        """Run manual optimization using train-test split (matches final evaluation)."""
         best_result = None
         best_params = None
         best_score = -float("inf") if self.util.high_is_good() else float("inf")
@@ -245,9 +255,136 @@ class OptimizationRunner:
 
         return best_params, best_result, self.results
 
+    def _run_manual_cv_optimization(self, combinations, param_specs):
+        """Run manual optimization using cross-validation."""
+        import numpy as np
+        from sklearn.model_selection import StratifiedKFold
+
+        self.util.debug("Using cross-validation for optimization (may differ from final evaluation)")
+
+        # Set up the experiment once to get the data
+        import nkululeko.experiment as exp
+        expr = exp.Experiment(self.config)
+        expr.set_module("optim")
+        expr.load_datasets()
+        expr.fill_train_and_tests()
+        expr.extract_feats()
+
+        # Create stratified CV splits
+        cv_splitter = StratifiedKFold(
+            n_splits=self.cv_folds, 
+            shuffle=True, 
+            random_state=self.random_state
+        )
+
+        best_result = None
+        best_params = None
+        best_score = -float("inf") if self.util.high_is_good() else float("inf")
+
+        for i, params in enumerate(combinations):
+            self.util.debug(f"Testing combination {i+1}/{len(combinations)}: {params}")
+
+            # Run cross-validation for this parameter combination
+            cv_scores = []
+            
+            try:
+                for fold, (train_idx, val_idx) in enumerate(cv_splitter.split(
+                    expr.feats_train, expr.df_train[self.config["DATA"]["target"]]
+                )):
+                    self.util.debug(f"  Fold {fold+1}/{self.cv_folds}")
+                    
+                    # Create fold-specific data
+                    fold_train_feats = expr.feats_train.iloc[train_idx]
+                    fold_val_feats = expr.feats_train.iloc[val_idx]
+                    fold_train_df = expr.df_train.iloc[train_idx]
+                    fold_val_df = expr.df_train.iloc[val_idx]
+
+                    # Update config with current parameters
[…]
