+++ ini_file.md
@@ -239,6 +239,233 @@
     * **sdr**: estimated [SDR](https://arxiv.org/pdf/2304.01448.pdf) (Perceptual Evaluation of Speech Quality)
     * **spkrec**: speaker-id: [speechbrain embeddings](https://huggingface.co/speechbrain/spkrec-xvect-voxceleb)
     * **stoi**: estimated [STOI](https://arxiv.org/pdf/2304.01448.pdf) (Perceptual Evaluation of Speech Quality)
+    * **squim**: [TorchAudio SQUIM](https://pytorch.org/audio/stable/tutorials/squim_tutorial.html) (Speech Quality and Intelligibility Measures)
+    * **trill**: [Google Research TRILL features](https://ai.googleblog.com/2020/06/improving-speech-representations-and.html)
+    * **wav2vec2**: [Facebook's wav2vec2 models](https://huggingface.co/docs/transformers/model_doc/wav2vec2)
+    * **whisper**: [OpenAI's Whisper ASR model](https://openai.com/research/whisper)
+    * **xbow**: [openXBOW processed opensmile features](https://github.com/openXBOW/openXBOW)
+    * **audmodel**: [audEERING's models](https://github.com/audeering/audmodel)
+* **balancing**: [balance the data with respect to class distribution](https://imbalanced-learn.org/stable/)
+  * balancing = smote
+  * possible values:
+    * **ros**: Random Over Sampler
+    * **smote**: SMOTE
+    * **adasyn**: ADASYN
+    * **borderlinesmote**: Borderline SMOTE
+    * **svmsmote**: SVM SMOTE
+    * **smoteenn**: SMOTE + Edited Nearest Neighbours
+    * **smotetomek**: SMOTE + Tomek links
+    * **clustercentroids**: Cluster Centroids
+    * **randomundersampler**: Random Under Sampler
+    * **editednearestneighbours**: Edited Nearest Neighbours
+    * **tomeklinks**: Tomek Links
+* **scale**: [scale (standard normalize) the features](http://blog.syntheticspeech.de/2021/08/10/the-nkululeko-configuration-file/)
+  * scale = standard
+  * possible values:
+    * **standard**: z-transformation (mean of 0 and std of 1) based on the training set
+    * **robust**: robust scaler
+    * **speaker**: like *standard* but based on individual speaker sets (also for the test)
+    * **bins**: convert feature values into 0, .5 and 1 (for low, mid and high)
+* **set**: name of opensmile feature set, e.g. eGeMAPSv02, ComParE_2016, GeMAPSv01a, eGeMAPSv01a
+  * set = eGeMAPSv02
+* **level**: level of opensmile features
+  * level = functional
+  * possible values:
+    * **functional**: aggregated over the whole utterance
+    * **lld**: low-level descriptor: framewise
+* **balancing**: balance the features with respect to [class distribution](https://imbalanced-learn.org/stable/)
+  * balancing=smote
+  * possible values: see above under **balancing**
+* **no_reuse**: don't re-use any feature files, but start fresh
+  * no_reuse = False
+* **needs_feature_extraction**: force the features to be freshly extracted
+  * needs_feature_extraction = False
+
+### MODEL
+
+Model and training specifications. In general, default values should work for classification tasks.
+
+* **type**: select the model
+  * type = xgb
+  * possible values:
+    * **xgb**: [XGBoost](http://blog.syntheticspeech.de/2021/07/13/nkululeko-how-to-use-xgboost-for-speech-emotion-recognition/)
[…]
