+++ examples/exp_polish_optim_svm.ini
@@ -44,7 +44,7 @@ metric = uar
 ; cv_folds = 3
 ; C = [0.1, 1.0, 10.0, 100.0, 1000.0]
 ; kernel = ["linear", "rbf", "poly"]
-; gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
+gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
 
 [RESAMPLE]
 replace = True
