+++ examples/exp_polish_optim_mlp.ini
@@ -26,16 +26,17 @@ type = ['os']
 
 [OPTIM]
 model = mlp
-search_strategy = random
-metric = uar
-n_iter = 20
-cv_folds = 3
-nlayers = [1, 2, 3]
-nnodes = [16, 32, 64, 128]
-lr = [0.0001, 0.001, 0.01, 0.1]
+; search_strategy = random
+; metric = uar
+; n_iter = 20
+; cv_folds = 3
+nlayers = [1, 2]
+nnodes = [16, 32]
+lr = [0.0001, 0.001]
 bs = [4, 8, 16, 32]
 loss = ["cross", "f1"]
 do = [0.1, 0.3, 0.5]
+metric = uar
 
 [RESAMPLE]
 replace = True

[…]
