+++ examples/exp_polish_optim_svm_intelligent.ini
@@ -1,40 +1,0 @@
-[EXP]
-root = ./examples/results/
-name = exp_polish_optim_svm_intelligent
-runs = 1
-epochs = 1
-
-[DATA]
-databases = ['train', 'dev', 'test']
-train = ./data/polish/polish_train.csv
-train.type = csv
-train.absolute_path = False
-train.split_strategy = train
-dev = ./data/polish/polish_dev.csv
-dev.type = csv
-dev.absolute_path = False
-dev.split_strategy = train
-test = ./data/polish/polish_test.csv
-test.type = csv
-test.absolute_path = False
-test.split_strategy = test
-target = emotion
-labels = ['anger', 'neutral', 'fear']
-
-[FEATS]
-type = ['os']
-
-[MODEL]
-type = svm
-
-[OPTIM]
-model = svm
-search_strategy = halving_random
-n_iter = 15
-cv_folds = 3
-C = [0.1, 1.0, 10.0, 100.0, 1000.0]
-kernel = ["linear", "rbf", "poly"]
-gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
-
-[RESAMPLE]
-replace = True
