### 2025-07-01 678f396
Fix #186: Rename test files to avoid pytest conflicts
- Rename test.py to nkulu_test.py
- Rename test_pretrain.py to nkulu_test_pretrain.py
- Rename test_predictor.py to nkulu_test_predictor.py
- Update import statements in nkuluflag.py and experiment.py
- Update setup.py entry point for nkulu_test module
- Update internal module references and docstrings

This prevents pytest from automatically discovering these files as test cases,
resolving naming conflicts with the test discovery mechanism.
The existing unit test structure in tests/ and models/tests/ directories
remains unchanged and funct
[…]
