### 2025-07-03 2160ee6
fixed bug with classifying integers
