### 2025-07-03 20269f4
fixed bug with classifying integers
