+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.95.2"
+VERSION="0.95.3"
 SAMPLING_RATE = 16000
