### 2025-07-03 d3ab99f
modified:   nkululeko/experiment.py 	modified:   nkululeko/nkuluflag.py 	renamed:    nkululeko/nkulu_test.py -> nkululeko/testing.py 	renamed:    nkululeko/nkulu_test_predictor.py -> nkululeko/testing_predictor.py 	renamed:    nkululeko/nkulu_test_pretrain.py -> nkululeko/testing_pretrain.py 	modified:   setup.py
