### 2025-07-08 felixbur/nkululeko#188 @bagustris
@isdanni

I fixed them in PR #229. Let's see the generated PDF after it is merged.
