+++ nkululeko/data/dataset.py
@@ -6,7 +6,6 @@ from random import sample
 
 import numpy as np
 import pandas as pd
-
 import audformat
 
 from nkululeko.filter_data import DataFilter
@@ -905,6 +904,20 @@ class Dataset:
         mappings = self.util.config_val_data(self.name, "mapping", False)
         if mappings:
             mapping = ast.literal_eval(mappings)
+            # mapping should be a dictionary, the keys might encode lists.
+            keys = list(mapping.keys())
+            for key in keys:
+                # a comma in the key means that the key is a list of labels
+                if "," in key:
+                    # split the key and create a list
+                    key_list = [k.strip() for k in key.split(",")]
+                    # create a new mapping for each key
+                    for k in key_list:
+                        mapping[k] = mapping[key]
+                    # remove the old key
+                    del mapping[key]
+            if pd.api.types.is_numeric_dtype(df[target]):
+                df[target] = df[target].astype("string")
             df[target] = df[target].map(mapping)
             # remove any exiting nan values after mapping
             df = df.dropna()
