### 2025-07-09 #8049 @bagustris
@faroit
We fixed issues raised by @fakufaku in this issue: https://github.com/felixbur/nkululeko/issues/201
