+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "nkululeko"
-version = "0.95.5"
+version = "0.95.6"
 description = "Machine learning audio prediction experiments based on templates"
 authors = [
     {name = "Felix Burkhardt", email = "fxburk@gmail.com"},
