+++ ini_file.md
@@ -348,124 +348,7 @@ Model and training specifications. In general, default values should work for cl
 * **save**: save the trained model
   * save = False
 
-### OPTIM
 
-Hyperparameter optimization configuration. Used with the `nkululeko.optim` module to automatically find the best model parameters.
-
-* **model**: model type to optimize (must match MODEL.type)
-  * model = xgb
-  * possible values: same as MODEL.type
-* **search_strategy**: optimization search strategy
-  * search_strategy = grid
-  * possible values:
-    * **grid**: exhaustive grid search (conventional)
-    * **random**: random search
-    * **halving_grid**: successive halving grid search (intelligent)
-    * **halving_random**: successive halving random search (intelligent)
-* **metric**: optimization metric to maximize/minimize
-  * metric = uar
-  * possible values:
-    * **uar**: Unweighted Average Recall (good for imbalanced data)
-    * **accuracy**: Overall accuracy
-    * **f1**: F1-score (macro-averaged)
-    * **precision**: Precision (macro-averaged)  
-    * **recall**: Recall (macro-averaged)
-    * **specificity**: Specificity
-* **n_iter**: number of iterations for random search strategies
-  * n_iter = 50
-* **cv_folds**: number of cross-validation folds
-  * cv_folds = 3
-* **random_state**: random seed for reproducible optimization
-  * random_state = 42
-
-#### Model-Specific Parameters
-
-**XGBoost/XGRegressor parameters:**
-* **n_estimators**: number of trees to test
-  * n_estimators = [50, 100, 200]
-* **max_depth**: maximum tree depth values
-  * max_depth = [3, 6, 9]
-* **learning_rate**: learning rates to test
-  * learning_rate = [0.01, 0.1, 0.3]
-* **subsample**: subsampling ratios
-  * subsample = [0.6, 0.8, 1.0]
-* **colsample_bytree**: column subsampling ratios
-  * colsample_bytree = [0.6, 0.8, 1.0]
-
-**SVM/SVR parameters:**
-* **C_val**: regularization values
-  * C_val = [0.1, 1.0, 10.0, 100.0]
-* **kernel**: kernel types  
-  * kernel = ["linear", "rbf", "poly"]
-* **gamma**: gamma values
-  * gamma = ["scale", "auto", 0.001, 0.01, 0.1, 1.0]
-
-**KNN parameters:**
-* **K_val**: number of neighbors
-  * K_val = [3, 5, 7, 9, 11]
-* **weights**: weight functions
-  * weights = ["uniform", "distance"]
-
-**MLP parameters:**
-* **nlayers**: number of hidden layers
-  * nlayers = [1, 2, 3]
-* **nnodes**: number of nodes per layer
-  * nnodes = [16, 32, 64, 128]
-* **lr**: learning rates
-  * lr = [0.0001, 0.001, 0.01]
-* **bs**: batch sizes  
-  * bs = [8, 16, 32, 64]
-* **do**: dropout rates
-  * do = [0.1, 0.3, 0.5]
-* **loss**: loss functions
-  * loss = ["cross", "f1", "mse"]
-
-#### Parameter Format
-
-Parameters can be specified in three formats:
-
-1. **List format**: discrete values to test
-   ```ini
-   C_val = [0.1, 1.0, 10.0]
-   kernel = ["linear", "rbf"]
-   ```
-
-2. **Range format**: continuous range (auto-generates steps)
-   ```ini
-   learning_rate = (0.001, 0.1)  # log-spaced values
-   max_depth = (3, 10)           # integer range
[…]
