+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.95.6"
+VERSION="0.95.7"
 SAMPLING_RATE = 16000
