+++ examples/exp_emodb_os_xgb.ini
@@ -18,6 +18,11 @@ type = ['os']
 store_format = csv 
 scale = standard
 #no_reuse = True
+[EXPL]
+feature_distributions = True
+model = ['tree']
+value_counts = [['age'], ['gender']]
+max_feats = 5
 [MODEL]
 type = xgb
 [PLOT]
