### 2025-07-14 541fa07
fixed bug that boolean label was not recognized as categorical
