### 2025-07-14 4d09bcb
Update nkululeko/utils/util.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
