### 2025-07-14 e484ade
Update nkululeko/plots.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
