### 2025-07-14 ca3e144
Update nkululeko/plots.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
