+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.95.8"
+VERSION="0.95.9"
 SAMPLING_RATE = 16000
