+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Version 0.95.9 (25-07-14)
+--------------------------
+* added google translation
+
 Version 0.95.8 (25-07-14)
 --------------------------
 * fix bug that bool was not detected as categorical
