### 2025-07-14 95f62b4
added bert extractor
