+++ ini_file.md
@@ -222,6 +222,9 @@
       * "wav2vec2-large-robust-ft-swbd-300h"
       * **wav2vec2.model** = *path to the wav2vec2 model folder*
       * **wav2vec2.layer** = *which last hidden layer to use*
+    * **bert variants**: [Bert embeddings](https://huggingface.co/transformers/v3.0.2/model_doc/bert.html#bertmodel)
+      * **bert.model** = *path to the bert model folder (without the google-bert/)*
+      * **bert.layer** = *which last hidden layer to use*
     * **Hubert variants**: [facebook Hubert models](https://ai.meta.com/blog/hubert-self-supervised-representation-learning-for-speech-recognition-generation-and-compression/)
       * "hubert-base-ls960", "hubert-large-ll60k", "hubert-large-ls960-ft", hubert-xlarge-ll60k, "hubert-xlarge-ls960-ft"
     * **WavLM**:
