### 2025-07-15 3ec9c97
Update nkululeko/feat_extract/feats_bert.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
