### 2025-07-15 917f0f3
Update nkululeko/feat_extract/feats_bert.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
