### 2025-07-17 bcb7eee
fix scatter plot in regression
