### 2025-07-17 aa1265d
Update nkululeko/reporting/reporter.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
