### 2025-07-17 286cf3f
fixed bug and added output
