+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.96.1"
+VERSION="0.96.2"
 SAMPLING_RATE = 16000
