### 2025-07-17 55d5a2b
Update nkululeko/experiment.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
