+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.96.2"
+VERSION="0.96.3"
 SAMPLING_RATE = 16000
