### 2025-07-22 71eae66
Update nkululeko/modelrunner.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
