+++ nkululeko/modelrunner.py
@@ -91,10 +91,11 @@ class Modelrunner:
                 if plot_epochs:
                     self.util.debug(f"plotting conf matrix to {plot_name}")
                     report.plot_confmatrix(plot_name, epoch)
-                store_models = eval(
+                import ast
+                store_models = ast.literal_eval(
                     self.util.config_val("EXP", "save", "False")
-                ) or eval(self.util.config_val("EXP", "traindevtest", "False"))
-                plot_best_model = eval(
+                ) or ast.literal_eval(self.util.config_val("EXP", "traindevtest", "False"))
+                plot_best_model = ast.literal_eval(
                     self.util.config_val("PLOT", "best_model", "False")
                 )
                 if (store_models or plot_best_model) and (
