+++ nkululeko/reporting/reporter.py
@@ -343,22 +343,6 @@ class Reporter:
             alpha=5,
         )
         acc = accuracy(truths, preds)
-        # cm = confusion_matrix(
-        #     truths, preds, normalize=None
-        # )  # normalize must be one of {'true', 'pred', 'all', None}
-        # if cm.shape[0] != len(labels):
-        #     self.util.error(
-        #         f"mismatch between confmatrix dim ({cm.shape[0]}) and labels"
-        #         f" length ({len(labels)}: {labels})"
-        #     )
-
-        # try:
-        #     disp = ConfusionMatrixDisplay(
-        #         confusion_matrix=cm, display_labels=labels
-        #     ).plot(cmap="Blues")
-        # except ValueError:
-        #     disp = ConfusionMatrixDisplay(
-        #         confusion_matrix=cm,
         #         display_labels=list(labels).remove("neutral"),
         #     ).plot(cmap="Blues")
         le = glob_conf.label_encoder
