+++ examples/exp_emodb_audmodel_mlp.ini
@@ -15,12 +15,12 @@ emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral
 emodb.limit = 350
 target = emotion
 [FEATS]
-type = ['audmodel']
+type = ['audwav2vec2']
 scale = standard
 [MODEL]
 type = mlp
 n_jobs = 8
-layers = {'l1':14, 'l2':8}
+layers = [14, 8]
 drop = .4
 save = True
 [PLOT]
