+++ nkululeko/feat_extract/feats_audwav2vec2.py
@@ -74,7 +74,7 @@ class Audwav2vec2Set(Featureset):
             self.df = self.util.get_store(storage, store_format)
 
     def extract_sample(self, signal, sr):
-        if self.model == None:
+        if self.model is None:
             self.__init__("na", None)
         result = self.model(signal, sr)
         return np.asarray(result["hidden_states"].flatten())
