+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.96.5"
+VERSION="0.96.6"
 SAMPLING_RATE = 16000
