+++ nkululeko/feat_extract/feats_analyser.py
@@ -10,6 +10,8 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.tree import DecisionTreeClassifier
 from sklearn.tree import DecisionTreeRegressor
 
+import audeer
+
 import nkululeko.glob_conf as glob_conf
 from nkululeko.plots import Plots
 from nkululeko.reporting.defines import Header
@@ -105,7 +107,9 @@ class FeatureAnalyser:
             f"SHAP analysis, features = {feature_importance.index.tolist()}"
         )
         # Save to CSV (save all features, not just top ones)
-        csv_filename = os.path.join(fig_dir, f"SHAP_{feat_type}_importance_{model.name}.csv")
+        csv_filename = os.path.join(
+            fig_dir, f"SHAP_{feat_type}_importance_{model.name}.csv"
+        )
         feature_importance.to_csv(csv_filename)
         self.util.debug(f"Saved SHAP feature importance to {csv_filename}")
         self.util.debug(f"plotted SHAP feature importance to {filename}")
@@ -292,12 +296,14 @@ class FeatureAnalyser:
             title += "\n based on feature permutation"
         ax.set(title=title)
         plt.tight_layout()
-        fig_dir = self.util.get_path("fig_dir")
+        # one up because of the runs
+        fig_dir = audeer.path(self.util.get_path("fig_dir"), "..")
+        self.util.get_path("fig_dir")
         format = self.util.config_val("PLOT", "format", "png")
         filename = f"EXPL_{model_name}"
         if permutation:
             filename += "_perm"
-        filename = f"{fig_dir}{filename}.{format}"
+        filename = audeer.path(fig_dir, f"{filename}.{format}")
         plt.savefig(filename)
         fig = ax.figure
         fig.clear()
