### 2025-07-31 8da5bb3
Update nkululeko/plots.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
