### 2025-07-31 d0ebe3a
Update pyproject.toml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
