+++ nkululeko/plots.py
@@ -616,8 +616,14 @@ class Plots:
         filename = audeer.path(
             fig_dir, f"feat_dist_{title}_{feature_name}.{self.format}"
         )
+        ignore_gender = eval(self.util.config_val("EXPL", "ignore_gender", "False"))
         if self.util.is_categorical(df_labels[label]):
-            if "gender" in df_labels and df_labels["gender"].notna().any():
+            p_val = ""
+            if (
+                "gender" in df_labels
+                and df_labels["gender"].notna().any()
+                and not ignore_gender
+            ):
                 # plot distribution for each gender in parallel violin plots
                 df_plot = pd.DataFrame(
                     {
@@ -642,7 +648,6 @@ class Plots:
                 df_plot = pd.DataFrame(
                     {label: df_labels[label], feature: df_features[feature]}
                 )
-                p_val = ""
                 if df_labels[label].nunique() == 2:
                     label_1 = df_labels[label].unique()[0]
                     label_2 = df_labels[label].unique()[1]
