+++ examples/exp_emodb_explore_featimportance.ini
@@ -19,7 +19,9 @@ type = xgb
 [EXPL]
 model = ['tree', 'xgb']
 scatter = ['pca']
+scatter.target = ['emotion', 'gender', 'age']
 plot_tree = True
 sample_selection = all
 feature_distributions = True
+ignore_gender = True
 max_feats = 7
