### 2025-07-31 76439d9
added nan check for feature importance analyzer
