+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.96.7"
+VERSION="0.96.8"
 SAMPLING_RATE = 16000
