+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.96.8"
+VERSION="0.97.0"
 SAMPLING_RATE = 16000
