### 2025-08-04 4ac1d37
added binscale example
