+++ examples/exp_emodb_binscaled.ini
@@ -1,0 +1,18 @@
+[EXP]
+root = ./examples/results/
+name = exp_emodb_binscaled
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+labels = ['anger', 'happiness']
+target = emotion
+#no_reuse = True
+[FEATS]
+type = ['os']
+store_format = csv 
+scale = bins
+[MODEL]
+type = xgb
