+++ examples/exp_emodb_stress.ini
@@ -1,0 +1,24 @@
+[EXP]
+root = ./examples/results/
+name = emodb_stress
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.target_tables = ["emotion"]
+emodb.colnames = {"emotion": "stress"}
+emodb.filter = [["stress", ["anger", "neutral", "sadness", "disgust"]]]
+emodb.mapping = {"anger": "stress", "disgust": "stress", "neutral": "no stress", "sadness": "no stress"}
+emodb.split_strategy = speaker_split
+labels = ["stress", "no stress"]
+target = stress
+[FEATS]
+type = ['os']
+#type = ['wav2vec2']
+#wav2vec2.layer = 4
+#wav2vec2.model = audeering/wav2vec2-large-robust-12-ft-emotion-msp-dim
+#audmodel.id = 1c250fd7-1.1.1
+[MODEL]
+type = svm
+[PLOT]
+uncertainty_threshold = 0.3
+
