### 2025-08-06 98716c3
update readme and add examples for multi drop
