+++ examples/exp_polish_mlp_drops.ini
@@ -27,7 +27,7 @@ type = ['os']
 [MODEL]
 type = mlp
 layers = [64, 32]
-drop = 0.1
+drop = [0.2, 0.1]
 
 [RESAMPLE]
 replace = True
