+++ examples/exp_emodb_stress.ini
@@ -1,6 +1,8 @@
 [EXP]
 root = ./examples/results/
 name = emodb_stress
+save_test = ./examples/results/emodb_stress/test.csv
+epochs = 5
 [DATA]
 databases = ['emodb']
 emodb = ./data/emodb/emodb
@@ -18,7 +20,10 @@ type = ['os']
 #wav2vec2.model = audeering/wav2vec2-large-robust-12-ft-emotion-msp-dim
 #audmodel.id = 1c250fd7-1.1.1
 [MODEL]
-type = svm
+#type = svm
+type = mlp
+layers = [64, 12]
+drop = [.3, .4]
 [PLOT]
 uncertainty_threshold = 0.3
 
