+++ meta/demos/exp_bdtg_local.ini
@@ -1,36 +1,0 @@
-[EXP]
-root = ./
-type = regression
-name = exp_bdtg
-runs = 1
-epochs = 1
-[DATA]
-databases = ['bdtg']
-bdtg = /home/audeering.local/fburkhardt/audb/bundestag-age/3.1.0/135fc543/
-bdtg.split_strategy = specified
-bdtg.test_tables = ['age.test']
-bdtg.train_tables = ['age.train']
-bdtg.main_table = files
-bdtg.value_counts = True
-target = age
-# labels = ['20ies', '30ies', '40ies', '50ies', '60ies', '70ies', '80ies']
-# bins  = [-10, 30, 40, 50, 60, 70, 80, 100]
-labels = ['u40', '40ies', '50ies', '60ies', 'ü70']
-bins  = [-1000,  40, 50, 60, 70, 1000]
-[FEATS]
-type=os
-#type=mld
-scale=standard
-[MODEL]
-type = xgr
-#type = mlp_reg
-#layers = {'l1':8, 'l2':4}
-store = 1
-measure = ccc
-loss_function = ccc
-[PLOT]
-#plot_epochs = 1
-#plot_anim_progression = 1
-#fps = 3
-#plot_epoch_progression = 1
-#plot_best_model = 1
