### 2025-08-08 e488388
update flags module
