+++ examples/exp_polish.ini
@@ -1,0 +1,25 @@
+[FLAGS]
+classifiers = ['svm', 'xgb', 'mlp']
+features = ['os', 'praat', '[os, praat]', 'opensmile']
+balancing = ['smote', 'smoteenn', 'clustercentroid']
+scale = ['standard', 'robust']
+
+[DATA]
+databases = ['polish']
+target = 'emotion'
+labels = ['happy', 'sad', 'angry']
+
+[MODEL]
+type = 'xgb'
+layers = 'l1:128'
+tuning_params = ['learning_rate', 'drop']
+learning_rate = 0.01
+drop = 0.2
+
+[FEATS]
+type = 'os'
+set = 'default'
+
+[EXP]
+epochs = 10
+runs = 3
