+++ nkululeko/flags.py
@@ -1,5 +1,7 @@
 import argparse
+import ast
 import configparser
+import itertools
 import os
 import os.path
 import sys
@@ -8,8 +10,150 @@ from nkululeko.nkululeko import doit as nkulu
 from nkululeko.testing import do_it as test_mod
 
 
+def run_flags_experiments(config_file):
+    """Run multiple experiments based on FLAGS section combinations."""
+    config = configparser.ConfigParser()
+    config.read(config_file)
+    
+    # Check if FLAGS section exists
+    if "FLAGS" not in config:
+        print("ERROR: No [FLAGS] section found in configuration")
+        return []
+    
+    flags_config = config["FLAGS"]
+    
+    # Parse flag parameters
+    flag_params = {}
+    for key, value in flags_config.items():
+        try:
+            flag_params[key] = ast.literal_eval(value)
+        except (ValueError, SyntaxError):
+            flag_params[key] = [value]
+    
+    print(f"Flag parameters found: {flag_params}")
+    
+    # Generate all combinations
+    param_names = list(flag_params.keys())
+    param_values = list(flag_params.values())
+    
+    combinations = []
+    for combo in itertools.product(*param_values):
+        param_dict = dict(zip(param_names, combo))
+        combinations.append(param_dict)
+    
+    print(f"Running {len(combinations)} experiment combinations...")
+    
+    results = []
+    
+    for i, combo in enumerate(combinations, 1):
+        print(f"\n=== Experiment {i}/{len(combinations)} ===")
+        print(f"Parameters: {combo}")
+        
+        # Create a copy of the original config
+        experiment_config = configparser.ConfigParser()
+        experiment_config.read_dict(config._sections)
+        
+        # Ensure required sections exist
+        if "MODEL" not in experiment_config:
+            experiment_config.add_section("MODEL")
+        if "FEATS" not in experiment_config:
+            experiment_config.add_section("FEATS")
+            
+        # Copy no_reuse from DATA to FEATS section if it exists to ensure proper feature reuse
+        if "DATA" in experiment_config and "no_reuse" in experiment_config["DATA"]:
+            experiment_config["FEATS"]["no_reuse"] = experiment_config["DATA"]["no_reuse"]
+        
+        # Apply combination parameters
+        for param, value in combo.items():
+            if param == "models":
+                experiment_config["MODEL"]["type"] = value
+            elif param == "features":
+                experiment_config["FEATS"]["type"] = f"['{value}']"
+            elif param == "balancing":
+                experiment_config["FEATS"]["balancing"] = value
+            elif param == "scale":
+                experiment_config["FEATS"]["scale"] = value
+            else:
+                # Handle other custom parameters
+                if param.startswith("model_"):
+                    experiment_config["MODEL"][param[6:]] = str(value)
+                elif param.startswith("feats_"):
+                    experiment_config["FEATS"][param[6:]] = str(value)
+                elif param.startswith("exp_"):
+                    experiment_config["EXP"][param[4:]] = str(value)
+        
+        # Create experiment name with parameters
+        base_name = experiment_config["EXP"]["name"]
+        param_suffix = "_".join([f"{k}_{v}" for k, v in combo.items()])
+        experiment_config["EXP"]["name"] = f"{base_name}_{param_suffix}"
+        
+        # Write temporary config
+        tmp_config = f"tmp_flags_{i}.ini"
+        with open(tmp_config, "w") as tmp_file:
+            experiment_config.write(tmp_file)
+        
+        try:
+            # Run the experiment
+            result, last_epoch = nkulu(tmp_config)
+            results.append({
+                "parameters": combo,
+                "result": result,
[…]
