### 2025-08-08 e141b9e
add example for flags
