+++ examples/exp_polish_flags.ini
@@ -1,0 +1,34 @@
+[EXP]
+root = ./examples/results/
+name = exp_polish_flags3
+
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/polish/polish_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+; train.audio_path = ./POLISH
+dev = ./data/polish/polish_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+; dev.audio_path = ./POLISH
+test = ./data/polish/polish_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+
+[FEATS]
+type = []
+no_reuse = False
+
+[MODEL]
+type = []
+
+[FLAGS]
+models = ['xgb', 'svm']
+features = ['praat']   # can be multiple but best for one
+balancing = ['smote', 'clustercentroids']
+scale = ['standard', 'robust', 'minmax']
