+++ nkululeko/flags.py
@@ -12,16 +12,21 @@ from nkululeko.testing import do_it as test_mod
 
 def run_flags_experiments(config_file):
     """Run multiple experiments based on FLAGS section combinations."""
+    import time
+
+    # Start timing the flags experiments
+    start_time = time.time()
+
     config = configparser.ConfigParser()
     config.read(config_file)
-    
+
     # Check if FLAGS section exists
     if "FLAGS" not in config:
         print("ERROR: No [FLAGS] section found in configuration")
         return []
-    
+
     flags_config = config["FLAGS"]
-    
+
     # Parse flag parameters
     flag_params = {}
     for key, value in flags_config.items():
@@ -29,131 +34,201 @@ def run_flags_experiments(config_file):
             flag_params[key] = ast.literal_eval(value)
         except (ValueError, SyntaxError):
             flag_params[key] = [value]
-    
+
     print(f"Flag parameters found: {flag_params}")
-    
+
     # Generate all combinations
     param_names = list(flag_params.keys())
     param_values = list(flag_params.values())
-    
+
     combinations = []
     for combo in itertools.product(*param_values):
         param_dict = dict(zip(param_names, combo))
         combinations.append(param_dict)
-    
+
     print(f"Running {len(combinations)} experiment combinations...")
-    
+
+    # Optimize feature extraction: load data and extract features ONCE
+    print("Setting up experiment and extracting features (once for all experiments)...")
+    import nkululeko.experiment as exp
+
+    # Set up the config with the first feature type for feature extraction
+    setup_config = configparser.ConfigParser()
+    for section_name in config.sections():
+        setup_config.add_section(section_name)
+        for key, value in config.items(section_name):
+            setup_config.set(section_name, key, value)
+
+    # Set a valid feature type for initial extraction
+    if "features" in flag_params and flag_params["features"]:
+        first_feature = flag_params["features"][0]
+        if "FEATS" not in setup_config:
+            setup_config.add_section("FEATS")
+        setup_config["FEATS"]["type"] = f"['{first_feature}']"
+
+    # Create base experiment to load data and extract features
+    base_experiment = exp.Experiment(setup_config)
+    base_experiment.set_module("flags")
+
+    try:
+        base_experiment.load_datasets()
+        base_experiment.fill_train_and_tests()
+        base_experiment.extract_feats()
+        print(f"Features extracted once: {base_experiment.feats_train.shape}")
+    except Exception as e:
+        print(f"ERROR during feature extraction setup: {e}")
+        return []
+
     results = []
-    
+
     for i, combo in enumerate(combinations, 1):
         print(f"\n=== Experiment {i}/{len(combinations)} ===")
         print(f"Parameters: {combo}")
-        
-        # Create a copy of the original config
-        experiment_config = configparser.ConfigParser()
-        experiment_config.read_dict(config._sections)
-        
-        # Ensure required sections exist
-        if "MODEL" not in experiment_config:
-            experiment_config.add_section("MODEL")
-        if "FEATS" not in experiment_config:
-            experiment_config.add_section("FEATS")
-            
-        # Copy no_reuse from DATA to FEATS section if it exists to ensure proper feature reuse
-        if "DATA" in experiment_config and "no_reuse" in experiment_config["DATA"]:
-            experiment_config["FEATS"]["no_reuse"] = experiment_config["DATA"]["no_reuse"]
-        
-        # Apply combination parameters
-        for param, value in combo.items():
-            if param == "models":
-                experiment_config["MODEL"]["type"] = value
-            elif param == "features":
-                experiment_config["FEATS"]["type"] = f"['{value}']"
-            elif param == "balancing":
-                experiment_config["FEATS"]["balancing"] = value
[…]
