+++ examples/exp_polish_flags.ini
@@ -1,6 +1,6 @@
 [EXP]
-root = ./examples/results/
-name = exp_polish_flags3
+root = /tmp/results/
+name = exp_polish_flags1
 
 [DATA]
 databases = ['train', 'dev', 'test']
@@ -20,15 +20,15 @@ test.absolute_path = False
 test.split_strategy = test
 target = emotion
 
-[FEATS]
-type = []
-no_reuse = False
+; [FEATS]
+; type = []
+; no_reuse = False
 
-[MODEL]
-type = []
+; [MODEL]
+; type = []
 
 [FLAGS]
 models = ['xgb', 'svm']
-features = ['praat']   # can be multiple but best for one
-balancing = ['smote', 'clustercentroids']
-scale = ['standard', 'robust', 'minmax']
+features = ['praat', 'os']   
+balancing = ['none', 'ros', 'smote']  
+scale = ['none', 'standard', 'robust', 'minmax']
