+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.97.3"
+VERSION = "0.97.3"
 SAMPLING_RATE = 16000
