+++ nkululeko/feat_extract/tests/test_feats_bert.py
@@ -28,9 +28,12 @@ def dummy_data_df():
 @pytest.fixture
 def bert_instance(dummy_data_df, dummy_util):
     # Patch Featureset to inject util and patch the problematic lines in Bert.__init__
-    with patch(
-        "nkululeko.feat_extract.feats_bert.Featureset.__init__", return_value=None
-    ), patch("torch.cuda.is_available", return_value=False):
+    with (
+        patch(
+            "nkululeko.feat_extract.feats_bert.Featureset.__init__", return_value=None
+        ),
+        patch("torch.cuda.is_available", return_value=False),
+    ):
         bert = Bert.__new__(Bert)  # Create instance without calling __init__
         # Manually set required attributes
         bert.util = dummy_util
@@ -56,9 +59,11 @@ def test_init_sets_device_and_feat_type(dummy_data_df, dummy_util):
 
 
 def test_init_model_calls_transformers(bert_instance):
-    with patch("transformers.AutoConfig.from_pretrained") as mock_config, patch(
-        "transformers.BertTokenizer.from_pretrained"
-    ) as mock_tokenizer, patch("transformers.BertModel.from_pretrained") as mock_model:
+    with (
+        patch("transformers.AutoConfig.from_pretrained") as mock_config,
+        patch("transformers.BertTokenizer.from_pretrained") as mock_tokenizer,
+        patch("transformers.BertModel.from_pretrained") as mock_model,
+    ):
         mock_cfg = MagicMock()
         mock_cfg.num_hidden_layers = 12
         mock_config.return_value = mock_cfg
