+++ nkululeko/feat_extract/tests/test_feats_praat_core.py
@@ -8,308 +8,313 @@ import parselmouth
 import pytest
 from scipy.stats import lognorm
 
-from nkululeko.feat_extract.feats_praat_core import (AudioFeatureExtractor,
-                                                     add_derived_features,
-                                                     compute_features,
-                                                     get_speech_rate, run_pca,
-                                                     speech_rate)
+from nkululeko.feat_extract.feats_praat_core import (
+    AudioFeatureExtractor,
+    add_derived_features,
+    compute_features,
+    get_speech_rate,
+    run_pca,
+    speech_rate,
+)
 
 
 class TestAudioFeatureExtractor:
-    
+
     @pytest.fixture
     def extractor(self):
         return AudioFeatureExtractor(f0min=75, f0max=300)
-    
+
     @pytest.fixture
     def mock_sound(self):
         sound = Mock()
         sound.get_total_duration.return_value = 2.5
         return sound
-    
+
     def test_init(self):
         extractor = AudioFeatureExtractor(f0min=50, f0max=400)
         assert extractor.f0min == 50
         assert extractor.f0max == 400
-    
+
     def test_init_default_values(self):
         extractor = AudioFeatureExtractor()
         assert extractor.f0min == 75
         assert extractor.f0max == 300
-    
-    @patch('nkululeko.feat_extract.feats_praat_core.call')
+
+    @patch("nkululeko.feat_extract.feats_praat_core.call")
     def test_extract_pitch_features(self, mock_call, extractor, mock_sound):
         mock_pitch = Mock()
         mock_point_process = Mock()
-        
+
         # Mock call return values
         mock_call.side_effect = [
             150.0,  # mean_f0
-            25.0,   # stdev_f0
-            Mock(), # harmonicity object
-            0.8,    # hnr
-            0.01,   # local_jitter
-            0.05,   # localabsolute_jitter
-            0.02,   # rap_jitter
-            0.03,   # ppq5_jitter
-            0.04,   # ddp_jitter
-            0.1,    # local_shimmer
-            0.5,    # localdb_shimmer
-            0.15,   # apq3_shimmer
-            0.2,    # apq5_shimmer
-            0.25,   # apq11_shimmer
-            0.3     # dda_shimmer
+            25.0,  # stdev_f0
+            Mock(),  # harmonicity object
+            0.8,  # hnr
+            0.01,  # local_jitter
+            0.05,  # localabsolute_jitter
+            0.02,  # rap_jitter
+            0.03,  # ppq5_jitter
+            0.04,  # ddp_jitter
+            0.1,  # local_shimmer
+            0.5,  # localdb_shimmer
+            0.15,  # apq3_shimmer
+            0.2,  # apq5_shimmer
+            0.25,  # apq11_shimmer
+            0.3,  # dda_shimmer
         ]
-        
-        result = extractor._extract_pitch_features(mock_sound, mock_pitch, mock_point_process)
-        
-        assert result['meanF0Hz'] == 150.0
-        assert result['stdevF0Hz'] == 25.0
-        assert result['HNR'] == 0.8
-        assert result['localJitter'] == 0.01
+
+        result = extractor._extract_pitch_features(
+            mock_sound, mock_pitch, mock_point_process
+        )
+
+        assert result["meanF0Hz"] == 150.0
+        assert result["stdevF0Hz"] == 25.0
+        assert result["HNR"] == 0.8
[…]

[…]
