### 2025-08-12 92de045
add test for flags
