+++ .github/workflows/format_code.yml
@@ -10,4 +10,4 @@ jobs:
       - uses: psf/black@stable
       # - uses: psf/black@552baf822992936134cbd31a38f69c8cfe7c0f05
         with:
-                options: "--exclude nkululeko/constants.py"
+                options: "nkululeko --exclude nkululeko/constants.py"
