+++ examples/exp_emodb_os_mlp.ini
@@ -1,7 +1,7 @@
 [EXP]
 root = ./examples/results/
-name = exp_emodb_classifiers
-runs = 1
+name = exp_emodb_mlp
+runs = 3
 epochs = 500
 save = True
 [DATA]
@@ -12,17 +12,15 @@ emodb.test_tables = ['emotion.categories.test.gold_standard']
 emodb.train_tables = ['emotion.categories.train.gold_standard']
 target = emotion
 labels = ['anger', 'happiness']
-#no_reuse = True
 [FEATS]
 type = ['os']
 scale = standard
-#no_reuse = True
 [MODEL]
 type = mlp
 layers = {'l1':128, 'l2':16}
 drop = .4
-patience = 5
-save = True
+patience = 3
+random_state = False
 [PLOT]
 best_model = True
 epoch_progression = True
