### 2025-08-12 6303775
added speakers_stated
