+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.97.4"
+VERSION="0.97.5"
 SAMPLING_RATE = 16000
