+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Version 0.97.5 (25-08-12)
+-------------------------
+* added specification of speakers for splits
+
 Version 0.97.4 (25-08-12)
 -------------------------
 * fixed bug about traindevtest and feature scaling
