+++ examples/exp_emodb_traindevtest_split.ini
@@ -2,17 +2,23 @@
 root = ./examples/results/
 name = exp_emodb_traindevtest_split
 traindevtest = True
+save_test = examples/results/exp_emodb_traindevtest_split/test_preds.csv
+epochs = 5
 [DATA]
-databases = ['emodb']
+databases = ['emodb', 'polish']
 emodb = ./data/emodb/emodb
 emodb.split_strategy = speakers_stated
 emodb.dev = [14, 8]
+polish = ./data/polish/polish_train.csv
+polish.type = csv
+polish.absolute_path = False
+polish.split_strategy = test
 #emodb.train =[3, 9, 10, 11, 13, 16]
-emodb.test = [12, 15]
-labels = ['anger', 'happiness', 'neutral', 'sadness']
+labels = ['anger', 'neutral', 'sadness']
 target = emotion
 [FEATS]
 type = ['os']
 scale = standard
 [MODEL]
-type = xgb
+type = mlp
+layers = [64, 32]
