### 2025-08-18 9ced446
Update nkululeko/flags.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
