+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.97.6"
+VERSION="0.98.0"
 SAMPLING_RATE = 16000
