+++ data/meld-st/README.md
@@ -1,0 +1,183 @@
+# Pre-processing script for MELD-ST dataset
+
+This directory contains the pre-processing script and data files for the MELD-ST (Multimodal EmotionLines Dataset - Speech Translation) dataset.
+
+## Dataset Description
+
+MELD-ST is a multimodal speech translation dataset that extends the original MELD dataset with parallel audio files in multiple languages. It contains emotion and sentiment annotations for dialogues from the TV show "Friends".
+
+The dataset includes:
+- **ENG_DEU**: English-German parallel audio with emotion/sentiment labels
+- **ENG_JPN**: English-Japanese parallel audio with emotion/sentiment labels
+
+### Emotions
+7 emotion classes: anger, disgust, fear, joy, neutral, sadness, surprise
+
+### Sentiments
+3 sentiment classes: negative, neutral, positive
+
+## Dataset Statistics
+
+### ENG_DEU Language Pair
+#### English (ENG)
+- Total samples: 11,642
+- Speakers: 291 unique
+- Train: 9,314 samples
+- Valid: 1,164 samples  
+- Test: 1,164 samples
+
+#### German (DEU)
+- Total samples: 11,642
+- Speakers: 291 unique
+- Train: 9,314 samples./data/meld-st/MELD-ST/ENG_DEU/DEU
+- Valid: 1,164 samples
+- Test: 1,164 samples
+
+### ENG_JPN Language Pair
+#### English (ENG)
+- Total samples: 10,085
+- Speakers: 281 unique
+- Train: 8,069 samples
+- Valid: 1,008 samples
+- Test: 1,008 samples
+
+#### Japanese (JPN)
+- Total samples: 10,085
+- Speakers: 281 unique
+- Train: 8,069 samples
+- Valid: 1,008 samples
+- Test: 1,008 samples
+
+## Generated Files
+
+The preprocessing script generates the following CSV files in nkululeko-compatible format:
+
+### Language-Specific Split Files
+
+#### ENG_DEU Dataset
+- `meld_st_eng_deu_eng_train.csv` - English training split for ENG_DEU
+- `meld_st_eng_deu_eng_valid.csv` - English validation split for ENG_DEU  
+- `meld_st_eng_deu_eng_test.csv` - English test split for ENG_DEU
+- `meld_st_eng_deu_deu_train.csv` - German training split for ENG_DEU
+- `meld_st_eng_deu_deu_valid.csv` - German validation split for ENG_DEU
+- `meld_st_eng_deu_deu_test.csv` - German test split for ENG_DEU
+
+#### ENG_JPN Dataset
+- `meld_st_eng_jpn_eng_train.csv` - English training split for ENG_JPN
+- `meld_st_eng_jpn_eng_valid.csv` - English validation split for ENG_JPN
+- `meld_st_eng_jpn_eng_test.csv` - English test split for ENG_JPN
+- `meld_st_eng_jpn_jpn_train.csv` - Japanese training split for ENG_JPN
+- `meld_st_eng_jpn_jpn_valid.csv` - Japanese validation split for ENG_JPN
+- `meld_st_eng_jpn_jpn_test.csv` - Japanese test split for ENG_JPN
+
+### Combined Files (All Splits)
+- `meld_st_eng_deu_eng_all.csv` - All English splits combined for ENG_DEU
+- `meld_st_eng_deu_deu_all.csv` - All German splits combined for ENG_DEU
+- `meld_st_eng_jpn_eng_all.csv` - All English splits combined for ENG_JPN
+- `meld_st_eng_jpn_jpn_all.csv` - All Japanese splits combined for ENG_JPN
+
+## CSV Format
+
+Each CSV file contains the following columns:
[…]
