+++ data/meld-st/README.md
@@ -183,7 +183,7 @@ To run experiments on text emotion recognition, you must have "text" column (def
 [FEATS]
 type = ['bert']
 bert.model = line-corporation/line-distilbert-base-japanese
-bert.text_columnt = Japanese2  # if text column is not "text"
+bert.text_column = Japanese2  # if text column is not "text"
 ```
 Then run it with `nkululeko.nkululeko`. Example of results are below; highlights showed that it uses Japanese LINE distillbert instead of Google Bert.
 
@@ -218,4 +218,5 @@ DEBUG: modelrunner: run: 0 epoch: 1: result: test: 0.266 UAR
 ```
 
 ## Reference:  
-[1] Poria, S., Hazarika, D., Majumder, N., Naik, G., Cambria, E., & Mihalcea, R. (2019). MELD: A Multimodal Multi-Party Dataset for Emotion Recognition in Conversations. Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics, 527–536. https://doi.org/10.18653/v1/p19-1050
+[1] Poria, S., Hazarika, D., Majumder, N., Naik, G., Cambria, E., & Mihalcea, R. (2019). MELD: A Multimodal Multi-Party Dataset for Emotion Recognition in Conversations. Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics, 527–536. https://doi.org/10.18653/v1/p19-1050  
+[2] Chen, S., Yahata, S., Shimizu, S., Yang, Z., Li, Y., Chu, C., & Kurohashi, S. (2024). MELD-ST: An Emotion-aware Speech Translation Dataset. Findings of the Association for Computational Linguistics ACL 2024, 10118–10126. https://doi.org/10.18653/v1/2024.findings-acl.601
