+++ ini_file.md
@@ -14,7 +14,6 @@
     - [SEGMENT](#segment)
     - [FEATS](#feats)
     - [MODEL](#model)
-    - [MODEL](#model-1)
     - [EXPL](#expl)
     - [PREDICT](#predict)
     - [EXPORT](#export)
@@ -23,6 +22,7 @@
     - [RESAMPLE](#resample)
     - [REPORT](#report)
     - [OPTIM](#optim)
+    - [FLAGS](#flags)
 
 ## Sections
 
@@ -291,9 +291,6 @@
   * possible values:
     * **functional**: aggregated over the whole utterance
     * **lld**: low-level descriptor: framewise
-* **balancing**: balance the features with respect to [class distribution](https://imbalanced-learn.org/stable/)
-  * balancing=smote
-  * possible values: see above under **balancing**
 * **no_reuse**: don't re-use any feature files, but start fresh
   * no_reuse = False
 * **needs_feature_extraction**: force the features to be freshly extracted
@@ -319,8 +316,17 @@ Model and training specifications. In general, default values should work for cl
     * **cnn**: [Convolutional neural network](http://blog.syntheticspeech.de/2022/01/17/how-to-use-convolutional-neural-networks-with-nkululeko/)
     * **lstm**: [Long short-term memory](http://blog.syntheticspeech.de/2022/03/18/nkululeko-how-to-use-recurrent-neural-networks/) recurrent neural network
     * **gru**: Gated recurrent unit
-    * **finetune**: [Fine-tuning](http://blog.syntheticspeech.de/2022/10/07/nkululeko-how-to-fine-tune-a-wav2vec2-model/) for pre-trained models
-    * **auto**: Automated machine learning using [flaml](https://microsoft.github.io/FLAML/)
+    * **finetune**: [Fine-tuning](http://blog.syntheticspeech.de/2022/10/07/nkululeko-how-to-fine-tune-a-wav2vec2-model/) for pre-trained models:
+      - pretrained_model: HF for base model
+      - push_to_hub: True
+      - max_duration: 8 (in seconds, resit are disgarded)  
+      - balancing: smote (as in FEATS, only for finetune needs to be defined here)  
+* **class_weight**: add class_weight to the linear classifier (XGB, SVM) fit methods for imbalanced data (True or False)
+  * class_weight = False
+* **logo**: leave-one-speaker group out. Will disregard train/dev splits and split the speakers in *logo* groups and then do a LOGO evaluation. If you want LOSO (leave one speaker out), simply set the number to the number of speakers.
+  * logo = 10
+* **k_fold_cross**: k-fold-cross validation. Will disregard train/dev splits and do a stratified cross validation (meaning that classes are balanced across folds). speaker id is ignored.
+  * k_fold_cross = 10
 * **learning_rate**: learning rate for neural networks
   * learning_rate = 0.0001
 * **drop**: dropout rate for neural networks (0 to 1)  
@@ -332,8 +338,15 @@ Model and training specifications. In general, default values should work for cl
   * possible values:
     * **cross**: CrossEntropyLoss
     * **f1**: F1 loss  
+    * **1-ccc**: concordance correlation coefficient
     * **mse**: Mean squared error (for regression)
     * **mae**: Mean absolute error (for regression)
+* **measure**: A measure/metric to report progress with regression experiments (classification is UAR)
+  * measure = mse
+  * possible values:
+    * **mse**: mean squared error
+    * **mae**: mean absolute error
+    * **ccc**: concordance correlation coefficient
 * **layers**: specify the layer architecture for MLP
   * layers = [64, 16]
 * **C_val**: regularization value for SVM
@@ -365,125 +378,6 @@ Model and training specifications. In general, default values should work for cl
   * patience = 5
[…]
