### 2025-08-26 9a71028
add more exp files for meld-st
