+++ data/meld-st/exp_eng_jpn_eng.ini
@@ -1,0 +1,34 @@
+[EXP]
+root = ./results/
+name = meld_st_eng_jpn_eng_emotion_e2vlarge
+
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/meld-st/meld_st_eng_jpn_eng_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/meld-st/meld_st_eng_jpn_eng_valid.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/meld-st/meld_st_eng_jpn_eng_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+
+[FEATS]
+type = ['emotion2vec-large']
+scale = robust
+balancing = ros
+
+[MODEL]
+type = xgb
+; c_val = 10
+; C = 0.001
+; kernel = linear
+
+[PLOT]
+sample_selection = auto
+format = png
