+++ data/meld-st/exp_eng_jpn_jpn.ini
@@ -1,0 +1,47 @@
+[EXP]
+root = ./results/
+name = meld_st_eng_jpn_jpn_emotion_e2vlarge
+
+[DATA]
+databases = ['train', 'dev', 'test']
+audio_path = ./data/meld-st/
+train = ./data/meld-st/meld_st_eng_jpn_jpn_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/meld-st/meld_st_eng_jpn_jpn_valid.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/meld-st/meld_st_eng_jpn_jpn_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+
+[FEATS]
+type = ['emotion2vec-large']
+scaler = robust
+balancing = smoteenn
+
+[MODEL]
+type = xgb
+n_estimators = 50
+max_depth = 3
+learning_rate = 0.01
+subsample = 0.5
+; C = 0.001
+; kernel = linear
+
+[OPTIM]
+model = xgb
+search_strategy = random
+n_estimators = [50, 100, 200, 300, 500]
+max_depth = [3, 4, 5, 6, 7, 8, 9, 10]
+learning_rate = [0.0001, 0.001, 0.01, 0.1]
+subsample = [0.5, 0.6, 0.7, 0.8, 0.9, 1.0]
+metric = uar
+
+[PLOT]
+sample_selection = auto
+format = png
