### 2025-08-29 #8049 @faroit
@fakufaku can you please create your checklist that its documented here?
