+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.98.0"
+VERSION="0.98.1"
 SAMPLING_RATE = 16000
