### 2025-09-08 47d9500
Add show_percentages configuration option to control percentage display in confusion matrix plots
Co-authored-by: bagustris <3158361+bagustris@users.noreply.github.com>
