### 2025-09-11 220a0f1
add example for svmsmote
