+++ scripts/run_balancing_experiments.sh
@@ -28,6 +28,7 @@ echo ""
 balancing_methods=(
     "ros"
     "smote" 
+    "svmsmote"
     "adasyn"
     "smoteenc"
     "borderlinesmote"
