### 2025-09-11 81272b9
update docs and workshop
