+++ docs/source/balance.md
@@ -1,0 +1,462 @@
+# Data Balancing in Nkululeko
+
+Data imbalance is a common problem in machine learning, particularly in speech emotion recognition and other audio classification tasks. When some classes have significantly more samples than others, models tend to be biased toward the majority classes, resulting in poor performance on minority classes.
+
+Nkululeko provides a comprehensive set of balancing techniques to address this issue through the `DataBalancer` class, supporting various over-sampling, under-sampling, and combination methods.
+
+## Overview
+
+The balancing functionality in nkululeko allows you to:
+
+- **Automatically detect** class imbalance in your datasets
+- **Apply various balancing techniques** to improve model performance
+- **Compare different balancing methods** using the flags module
+- **Preserve data integrity** while addressing imbalance issues
+
+## Quick Start
+
+To quickly try balancing techniques, you can use the provided demo example:
+
+```bash
+# Clone the repository and navigate to it
+cd nkululeko
+
+# Run the balancing demo with SMOTE
+python -m nkululeko.nkululeko --config examples/exp_balancing_demo.ini
+```
+
+This demo uses the test dataset included with nkululeko and applies SMOTE balancing to show how class distribution changes.
+
+To use balancing in your own nkululeko experiment, simply add the `balancing` parameter to your `[FEATS]` section:
+
+```ini
+[FEATS]
+type = ['os']
+balancing = smote
+scale = standard
+```
+
+## Supported Balancing Methods
+
+Nkululeko supports three categories of balancing techniques:
+
+### 1. Over-sampling Methods
+These methods increase the number of minority class samples:
+
+- **`ros`** - Random Over-Sampling: Randomly duplicates minority class samples
+- **`smote`** - SMOTE: Generates synthetic samples using k-nearest neighbors
+- **`adasyn`** - ADASYN: Adaptive synthetic sampling with density-based generation
+- **`borderlinesmote`** - Borderline SMOTE: Focuses on borderline samples
+- **`svmsmote`** - SVM SMOTE: Uses SVM to identify support vectors for synthesis
+
+### 2. Under-sampling Methods
+These methods reduce the number of majority class samples:
+
+- **`randomundersampler`** - Random Under-Sampling: Randomly removes majority class samples
+- **`clustercentroids`** - Cluster Centroids: Replaces clusters with their centroids
+- **`editednearestneighbours`** - Edited Nearest Neighbours: Removes inconsistent samples
+- **`tomeklinks`** - Tomek Links: Removes Tomek link pairs
+
+### 3. Combination Methods
+These methods combine over-sampling and under-sampling:
+
+- **`smoteenn`** - SMOTE + Edited Nearest Neighbours
+- **`smotetomek`** - SMOTE + Tomek Links
+
+## Working Examples
+
+### Basic SMOTE Balancing
+
+```ini
+[EXP]
+root = ./examples/results/
+name = exp_smote_balancing
+runs = 1
+epochs = 10
+
+[DATA]
+databases = ['emodb']
+emodb = ./data/emodb/emodb
+emodb.split_strategy = specified
+emodb.test_tables = ['emotion.categories.test.gold_standard']
+emodb.train_tables = ['emotion.categories.train.gold_standard']
+target = emotion
+labels = ['anger', 'happiness', 'sadness', 'neutral']
+
+[FEATS]
+type = ['os']
+balancing = smote
+scale = standard
+
+[MODEL]
+type = xgb
+```
+
+### Random Over-Sampling Example
+
+```ini
+[EXP]
+root = ./examples/results/
+name = exp_ros_balancing
+
+[DATA]
+databases = ['polish_train', 'polish_dev', 'polish_test']
+polish_train = ./data/polish/polish_train.csv
[…]
