+++ examples/exp_scaling_demo.ini
@@ -1,0 +1,23 @@
+[EXP]
+root = ./examples/results/
+name = test_scaling_demo
+runs = 1
+epochs = 3
+
+[DATA]
+databases = ['testdb']
+testdb = ./data/test/samples.csv
+testdb.type = csv
+testdb.absolute_path = False
+testdb.split_strategy = train_test
+target = emotion
+
+[FEATS]
+type = ['os']
+scale = standard
+
+[MODEL]
+type = xgb
+
+[PLOT]
+name = scaling_demo_results

[…]
