### 2025-09-11 b9202bf
add more scripts
