+++ scripts/run_ensemble.sh
@@ -1,0 +1,6 @@
+python3 -m nkululeko.ensemble \
+# cocosda_2024/exp_kbes_audmodel.ini \
+cocosda_2024/exp_kbes_hubert.ini \
+cocosda_2024/exp_kbes_wavlm.ini \
+--method performance_weighted \
+--weights 0.775 0.758
