### 2025-09-11 4dda9f4
Update scripts/run_parallel.sh
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
