+++ scripts/run_ensemble.sh
@@ -1,6 +1,9 @@
+#/bi/bash 
+# ensemble based on performance
+# based on ococosda 2024 paper
 python3 -m nkululeko.ensemble \
-# cocosda_2024/exp_kbes_audmodel.ini \
-cocosda_2024/exp_kbes_hubert.ini \
-cocosda_2024/exp_kbes_wavlm.ini \
+examples/exp_kbes_audmodel.ini \
+examples/exp_kbes_hubert.ini \
+examples/exp_kbes_wavlm.ini \
 --method performance_weighted \
---weights 0.775 0.758
+--weights 0.77 0.775 0.758 
