### 2025-09-11 6cc0d65
added readme and data creation script
