+++ data/androids/exp.ini
@@ -1,6 +1,6 @@
 [EXP]
-root = ./
-name = /tests/results/exp_androids
+root = ./data/androids/
+name = results
 runs = 1
 epochs = 1
 save = True
@@ -16,10 +16,10 @@ data.split_strategy = random
 data.limit_samples_per_speaker = 8
 target = depression
 labels = ['depressed', 'control']
-no_reuse = True
-#filter = [['task', 'interview']]
+filter = [['task', 'interview']]
 [RESAMPLE]
 sample_selection = all
+replace = True
 [FEATS]
 type = ['os']
 scale = standard
