### 2025-09-11 104e23b
tqdm enhancement
