+++ data/emodb/README.md
@@ -1,12 +1,14 @@
-EmoDB can be downloaded from [Zenodo](https://zenodo.org/record/7447302).
+EmoDB can be downloaded from [Zenodo](https://zenodo.org/records/16273947).
 
 Simply download and unzip, it is already in audformat.
 
 ```bash
 # Download using wget
-wget https://zenodo.org/record/7447302/files/emodb.zip
+wget https://zenodo.org/record/16273947/files/emodb_2.0.zip
 # Unzip
-unzip emodb.zip
+unzip emodb_2.0.zip
+# rename to simply emodb
+mv emodb_2.0 emodb
 # change to Nkululeko parent directory
 cd ../..
 # run the nkululeko experiment
