+++ meta/demos/exp_emodb.ini
@@ -3,7 +3,7 @@ root = ./results/
 name = exp_emodb
 [DATA]
 databases = ['emodb']
-emodb = ./emodb
+emodb = ./data/emodb/emodb
 emodb.split_strategy = specified
 emodb.train_tables = ['emotion.categories.train.gold_standard']
 emodb.test_tables = ['emotion.categories.test.gold_standard']
