### 2025-09-16 4c169c9
added a warning for broken files
