### 2025-09-16 d04c0c3
added file name for processing errors
