+++ nkululeko/models/model_svm.py
@@ -13,7 +13,7 @@ class SVM_model(Model):
     def __init__(self, df_train, df_test, feats_train, feats_test):
         super().__init__(df_train, df_test, feats_train, feats_test)
         self.name = "svm"
-        c = float(self.util.config_val("MODEL", "C_val", "0.001"))
+        c = float(self.util.config_val("MODEL", "C_val", "1"))
         if eval(self.util.config_val("MODEL", "class_weight", "False")):
             class_weight = "balanced"
         else:
