### 2025-09-16 b9d9725
update scaler and balancer example and slide
