+++ meta/workshop/voice_wellness_2025/slide_balance.md
@@ -1,7 +1,7 @@
-# Data Balancing in Nkululeko
-## Addressing Class Imbalance in Speech Emotion Recognition
-
----
+# Data Balancing in Machine Learning and its implementation in Nkululeko
+  
+### Addressing Class Imbalance Machine Learning  
+Study Case: Speech Processing
 
 ## Slide 1: Why We Need Data Balancing
 
@@ -25,7 +25,7 @@ Class Distribution:
 
 ### The Accuracy Paradox
 - **Scenario**: Medical diagnosis with 99% healthy, 1% diseased patients
-- **Naive Model**: Always predict "healthy" → 99% accuracy!
+- **Naive Model**: Always predict "healthy" -> 99% accuracy!
 - **Reality**: Completely useless for detecting disease
 
 ---
@@ -56,19 +56,17 @@ Class Distribution:
 
 **Before Balancing:**
 ```
-Angry: ████ (200 samples)
-Sad:   ████████ (400 samples)  
-Happy: ████████████████ (800 samples)
+Angry: #### (200 samples)
+Sad:   ######## (400 samples)  
+Happy: ################ (800 samples)
 ```
 
 **After SMOTE Balancing:**
 ```
-Angry: ████████████████ (800 samples)
-Sad:   ████████████████ (800 samples)
-Happy: ████████████████ (800 samples)
+Angry: ################ (800 samples)
+Sad:   ################ (800 samples)
+Happy: ################ (800 samples)
 ```
-
----
 
 ## Slide 3: Nkululeko's Balancing Implementation
 
@@ -204,20 +202,20 @@ scale = standard
 
 ```
 Dataset Size?
-├── Small (<1000 samples)
-│   └── Use Over-sampling
-│       ├── Start with: SMOTE
-│       └── Try: ROS (baseline), ADASYN (severe imbalance)
-│
-├── Medium (1000-10000 samples)  
-│   └── Use Combination Methods
-│       ├── Start with: SMOTE + Tomek
-│       └── Try: SMOTE + ENN (noisy data)
-│
-└── Large (>10000 samples)
-    └── Use Under-sampling
-        ├── Start with: Cluster Centroids
-        └── Try: Random Under-sampling (fast)
++-- Small (<1000 samples)
+|   +-- Use Over-sampling
+|       +-- Start with: SMOTE
+|       +-- Try: ROS (baseline), ADASYN (severe imbalance)
+|
++-- Medium (1000-10000 samples)  
+|   +-- Use Combination Methods
+|       +-- Start with: SMOTE + Tomek
+|       +-- Try: SMOTE + ENN (noisy data)
+|
++-- Large (>10000 samples)
+    +-- Use Under-sampling
+        +-- Start with: Cluster Centroids
+        +-- Try: Random Under-sampling (fast)
 ```
 
 ### Guidelines by Dataset Characteristics
@@ -263,16 +261,18 @@ type = xgb
 ### Comparing Multiple Methods
 ```ini
 [FLAGS]
-balancing = ['none', 'smote', 'adasyn', 'clustercentroids', 'smoteenn']
+balancing = ['none', 'smote', 'adasyn', 'smoteenn']
 ```
 
 ### Expected Output
-```
+```bash
 Balancing features with: smote
 Original dataset size: 1200
-Original class distribution: {'happy': 400, 'sad': 300, 'angry': 300, 'neutral': 200}
+Original class distribution: 
+ {'happy': 400, 'sad': 300, 'angry': 300, 'neutral': 200}
 Balanced dataset size: 1600 (was 1200)
-New class distribution: {'happy': 400, 'sad': 400, 'angry': 400, 'neutral': 400}
+New class distribution: 
+ {'happy': 400, 'sad': 400, 'angry': 400, 'neutral': 400}
 ```
 
[…]
