+++ meta/workshop/voice_wellness_2025/slide_scaler.md
@@ -57,7 +57,8 @@ SVM Decision Boundary influenced by:
 - Energy: 0.1 (small impact)
 - Spectral Centroid: 2000 (dominates!)
 - F0: 150 (medium impact)
-Result: Model ignores energy, focuses only on spectral features
+Result: Model ignores energy, 
+focuses only on spectral features
 ```
 
 **With Scaling:**
@@ -123,9 +124,12 @@ After:  [-1.2, -0.4, 0.8, 1.6] (standardized)
 - **Use Case**: When data contains outliers or non-normal distribution
 
 ```
-Before: [80, 150, 300, 450, 2000] (F0 with outlier)
-Robust: [-0.8, -0.3, 0.5, 1.0, 4.2] (outlier less dominant)
-Standard: [-0.9, -0.7, -0.3, 0.1, 4.8] (outlier more dominant)
+Before: [80, 150, 300, 450, 2000] 
+(F0 with outlier)
+Robust: [-0.8, -0.3, 0.5, 1.0, 4.2] 
+(outlier less dominant)
+Standard: [-0.9, -0.7, -0.3, 0.1, 4.8] 
+(outlier more dominant)
 ```
 
 ### 3. Min-Max Scaling
@@ -153,13 +157,13 @@ After:  [-0.5, -0.1, 0.25, 1.0] (scaled by absolute max)
 ## Slide 5: Advanced Distribution Methods
 
 ### 5. Normalizer (L2 Norm)
-- **Formula**: `x / ||x||₂` (per sample)
-- **Result**: Each sample has unit norm
-- **Use Case**: When direction matters more than magnitude
+- **Formula**: $\dfrac{x}{||x||_2}$ (per sample)  
+- **Result**: Each sample has unit norm  
+- **Use Case**: When direction matters more than magnitude  
 
 ```
 Sample: [F0=200, Energy=0.5, MFCC1=10]
-Norm = √(200² + 0.5² + 10²) = 200.25
+Norm = sqrt(200^2 + 0.5^2 + 10^2) = 200.25
 Result: [0.999, 0.0025, 0.05]
 ```
 
@@ -169,9 +173,10 @@ Result: [0.999, 0.0025, 0.05]
 - **Use Case**: Skewed features, improving normality
 
 ```
-Before: Highly skewed energy distribution
-After:  More bell-shaped, normal-like distribution
-Effect: Better performance for algorithms assuming normality
+Before: Highly skewed energy distribution  
+After:  More bell-shaped, normal-like distribution  
+Effect: Better performance for algorithms  
+assuming normality
 ```
 
 ### 7. Quantile Transformer
@@ -180,9 +185,9 @@ Effect: Better performance for algorithms assuming normality
 - **Use Case**: Heavy outliers, unknown distribution
 
 ```
-Before: Any distribution shape
-After:  Perfect uniform [0,1] or normal N(0,1)
-Benefit: Completely removes outlier effects
+Before: Any distribution shape  
+After:  Perfect uniform [0,1] or normal N(0,1)  
+Benefit: Completely removes outlier effects  
 ```
 
 ---
@@ -195,32 +200,35 @@ Benefit: Completely removes outlier effects
 - **Output**: String values "0", "0.5", "1"
 
 ```
-Before: [80, 120, 150, 200, 300, 400, 450] (F0 values)
-Percentiles: 33% = 150, 66% = 350
-After:  ["0", "0", "0", "0.5", "0.5", "1", "1"]
+Before: [80, 120, 150, 200, 300, 400, 450] 
+(F0 values)   
+Percentiles: 33% = 150, 66% = 350  
+After:  ["0", "0", "0", "0.5", "0.5", "1", "1"]  
 ```
 
 **Benefits:**
-- Robust to outliers
[…]
