+++ examples/exp_balancing_svmsmote.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_svmsmote_balancing
+name = exp_polish_os
 runs = 1
 epochs = 1
 save = True
