+++ examples/exp_balancing_borderlinesmote.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_borderlinesmote_balancing
+name = exp_polish_os
 runs = 1
 epochs = 1
 save = True
