+++ examples/exp_balancing_smotetomek.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_smotetomek_balancing
+name = exp_polish_os
 runs = 1
 epochs = 1
 save = True
