+++ examples/exp_balancing_clustercentroids.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_clustercentroids_balancing
+name = exp_polish_os
 runs = 1
 epochs = 1
 save = True
