+++ examples/exp_polish_bayes.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_os
+name = exp_polish_os1
 [DATA]
 databases = ['train', 'dev', 'test']
 train = ./data/polish/polish_train.csv
@@ -20,6 +20,7 @@ test.split_strategy = test
 target = emotion
 [FEATS]
 type = ['os']
+; balancing = ros
 ; type = ['hubert-xlarge-ll60k']
 ; no_reuse = False
 ; scale = standard
@@ -30,3 +31,5 @@ type = bayes
 replace = True
 [PLOT]
 format = svg
+[FLAGS]
+scale = ["none","standard","minmax","robust","maxabs"]
