### 2025-09-17 8e9d8d7
added the section that you should remove the audio folder from interview
