+++ examples/exp_polish_bayes.ini
@@ -1,6 +1,6 @@
 [EXP]
 root = ./examples/results/
-name = exp_polish_os1
+name = exp_polish_os
 [DATA]
 databases = ['train', 'dev', 'test']
 train = ./data/polish/polish_train.csv
@@ -23,7 +23,7 @@ type = ['os']
 ; balancing = ros
 ; type = ['hubert-xlarge-ll60k']
 ; no_reuse = False
-; scale = standard
+scale = standard
 [MODEL]
 type = bayes
 ; save = True
