+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.98.1"
+VERSION="0.98.2"
 SAMPLING_RATE = 16000
