### 2025-09-17 fcdf786
various bug fixes and added warnings
