+++ examples/exp_polish_bayes.ini
@@ -32,4 +32,4 @@ replace = True
 [PLOT]
 format = svg
 [FLAGS]
-scale = ["none","standard","minmax","robust","maxabs"]
+scale = ["none","standard","minmax","robust","maxabs", "speaker"]
