+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.98.2"
+VERSION="0.98.3"
 SAMPLING_RATE = 16000
