### 2025-09-22 ed23af4
Update nkululeko/reporting/run_plotter.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
