+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Version 0.98.4 (25-09-22)
+-------------------------
+* added comparison plots for run results
+
 Version 0.98.3 (25-09-18)
 -------------------------
 * updated balancing and scaling examples and slides
