+++ nkululeko/utils/stats.py
@@ -3,6 +3,7 @@ import math
 
 import numpy as np
 import pandas as pd
+from scipy import stats
 
 
 def check_na(a):
@@ -94,6 +95,214 @@ def cohens_D_to_string(val: float) -> str:
     else:
         rval = "large effect"
     return f"Cohen's d: {rval}"
+
+
+def p_value_to_string(p_val: float) -> str:
+    """Convert p-value to interpretable significance string.
+
+    Args:
+        p_val: p-value from statistical test
+    Returns:
+        String describing statistical significance level
+    """
+    if p_val < 0.001:
+        return f"highly significant ({p_val:.3f} < 0.001)"
+    elif p_val < 0.01:
+        return f"significant ({p_val:.3f} < 0.01)"
+    elif p_val < 0.05:
+        return f"significant ({p_val:.3f}< 0.05)"
+    elif p_val < 0.1:
+        return f"marginally significant ({p_val:.3f} < 0.1)"
+    else:
+        return f"not significant ({p_val:.3f} >= 0.1)"
+
+
+def get_t_test_effect(
+    variable1: np.array, variable2: np.array, equal_var: bool = True
+) -> tuple[float, float, str]:
+    """Get t-test statistics for two real-numbered distributions.
+
+    Performs independent samples t-test between two continuous variables.
+
+    Args:
+        variable1: first real-numbered distribution
+        variable2: second real-numbered distribution
+        equal_var: whether to assume equal variances (default True)
+                  If False, uses Welch's t-test
+    Returns:
+        t-statistic
+        p-value
+        significance interpretation string
+    """
+    # Handle NaN values
+    variable1 = check_na(variable1)
+    variable2 = check_na(variable2)
+
+    # Perform t-test
+    t_stat, p_value = stats.ttest_ind(variable1, variable2, equal_var=equal_var)
+
+    # Create interpretation string
+    significance = p_value_to_string(p_value)
+
+    return float(t_stat), float(p_value), significance
+
+
+def get_mannwhitney_effect(
+    variable1: np.array, variable2: np.array
+) -> tuple[float, float, str]:
+    """Get Mann-Whitney U test statistics for two real-numbered distributions.
+
+    Performs Mann-Whitney U test between two continuous variables.
+    This is a non-parametric test that doesn't assume normal distributions.
+
+    Args:
+        variable1: first real-numbered distribution
+        variable2: second real-numbered distribution
+    Returns:
+        U-statistic
+        p-value
+        significance interpretation string
+    """
+    # Handle NaN values
+    variable1 = check_na(variable1)
+    variable2 = check_na(variable2)
+
+    # Perform Mann-Whitney U test
+    u_stat, p_value = stats.mannwhitneyu(variable1, variable2, alternative="two-sided")
+
+    # Create interpretation string
+    significance = p_value_to_string(p_value)
+
+    return float(u_stat), float(p_value), significance
+
+
+def find_most_significant_difference(
+    distributions: dict, equal_var: bool = True
+) -> tuple[str, float, float, str, dict]:
+    """Find the combination with the most significant t-test difference among n distributions.
+
+    Args:
+        distributions: dictionary with distribution names as keys and numpy arrays as values
+        equal_var: whether to assume equal variances (default True)
+                  If False, uses Welch's t-test
+    Returns:
+        Most significant pair as string "dist1-dist2"
+        t-statistic for most significant pair
+        p-value for most significant pair
+        significance interpretation for most significant pair
+        Dictionary with all pairwise comparisons containing t_stat, p_value, significance
+    """
+    if len(distributions) < 2:
+        raise ValueError("Need at least 2 distributions for comparison")
+
+    dist_names = list(distributions.keys())
[…]
