### 2025-09-23 130728f
changed digit num of floats to 3
