### 2025-09-23 fe764e2
debugging stats results
