+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 Changelog
 =========
 
-Version 0.98.4 (25-09-22)
+Version 0.98.4 (25-09-24)
 -------------------------
 * added comparison plots for run results
+* added statistic tests for feature distribution plots
 
 Version 0.98.3 (25-09-18)
 -------------------------
