+++ nkululeko/plots.py
@@ -628,7 +628,6 @@ class Plots:
         )
         ignore_gender = eval(self.util.config_val("EXPL", "ignore_gender", "False"))
         sample_num = df_labels.shape[0]
-        title = f"{title} ({sample_num})"
         if self.util.is_categorical(df_labels[label]):
             p_val = ""
             cat_num = df_labels[label].nunique()
