+++ data/ogvc/exp.ini
@@ -1,0 +1,34 @@
+[EXP]
+root = ./results_ogvc_hubert/
+name = exp_ogvc_hubert_allemotions
+; runs = 1
+; epochs = 1
+; save = True
+[DATA]
+databases = ['train', 'dev', 'test']
+train = ./data/ogvc/ogvc_train.csv
+train.type = csv
+train.absolute_path = False
+train.split_strategy = train
+dev = ./data/ogvc/ogvc_dev.csv
+dev.type = csv
+dev.absolute_path = False
+dev.split_strategy = train
+test = ./data/ogvc/ogvc_test.csv
+test.type = csv
+test.absolute_path = False
+test.split_strategy = test
+target = emotion
+; JOY, ACC, FEA, SUR, SAD, DIS, ANG, ANT, NEU, OTH
+labels = ['JOY', 'ANG', 'SAD', 'ANT', 'FEA', 'SUR', 'ACC', 'DIS']
+; get the number of classes from the target column automatically
+[FEATS]
+type = ['hubert-xlarge-ll60k']
+; no_reuse = False
+scale = standard
+[MODEL]
+type = svm
+; save = True
+C_val = 10
+[RESAMPLE]
+replace = True
