+++ data/uudb/exp.ini
@@ -1,0 +1,34 @@
+;"Pleasantness", "Arousal", "Dominance", "Credibility", "Interest", "Positivity"
+
+[EXP]
+root = ./results_UUDB/
+name = exp_UUDB_Pleasantness_os_svm_1
+; runs = 1
+; epochs = 1
+; save = True
+
+[DATA]
+databases = ['train','test']
+train = C:\Users\Atoz_\.nkululeko\data\UUDB\UUDB\datasets\dataset_Pleasantness_train.csv
+train.type = csv
+train.absolute_path = True
+train.split_strategy = train
+test = C:\Users\Atoz_\.nkululeko\data\UUDB\UUDB\datasets\dataset_Pleasantness_test.csv
+test.type = csv
+test.absolute_path = True
+test.split_strategy = test
+target = label
+stratify = True
+labels = ['weak', 'strong']
+
+[FEATS]
+type = ['os']
+scale = standard
+balancing = smote
+
+[MODEL]
+type = svm
+C_val = 10
+
+[RESAMPLE]
+replace = True
