+++ data/uudb/exp.ini
@@ -9,13 +9,13 @@ name = exp_UUDB_Pleasantness_os_svm_1
 
 [DATA]
 databases = ['train','test']
-train = C:\Users\Atoz_\.nkululeko\data\UUDB\UUDB\datasets\dataset_Pleasantness_train.csv
+train = data/uudb/UUDB/datasets/dataset_Pleasantness_train.csv
 train.type = csv
-train.absolute_path = True
+train.absolute_path = False
 train.split_strategy = train
-test = C:\Users\Atoz_\.nkululeko\data\UUDB\UUDB\datasets\dataset_Pleasantness_test.csv
+test = data/uudb/UUDB/datasets/dataset_Pleasantness_test.csv
 test.type = csv
-test.absolute_path = True
+test.absolute_path = False
 test.split_strategy = test
 target = label
 stratify = True
