+++ data/eatd/mpl_exp.ini
@@ -1,0 +1,58 @@
+[EXP]
+root = ./results/
+name = eatd_mlp_experiment1
+runs = 1
+epochs = 5
+
+[DATA]
+databases = ['train', 'dev','test']
+
+train = ./data/eatd/eatd_train.csv
+train.type = csv
+train.absolute_path = True
+train.split_strategy = train
+
+dev = ./data/eatd/eatd_dev.csv
+dev.type = csv
+dev.absolute_path = True
+dev.split_strategy = train
+
+test = ./data/eatd/eatd_test.csv
+test.type = csv
+test.absolute_path = True
+test.split_strategy = test
+
+# depression classification
+target = task
+labels = ['depressed', 'non-depressed']
+balance = True
+
+[FEATS]
+# try OpenSMILE and MFCC features
+type = ['hubert-xlarge-ll60k']
+scale = standard
+#scaler = True
+balancing = ros
+
+[MODEL]
+type = mlp
+layers = [128,64]   # two hidden layers
+activation = relu
+dropout = 0.3
+
+batch_size = 32
+optimizer = adam
+learning_rate = 0.001
+class_weight = balanced
+#patience=10
+
+
+#[FLAGS]
+#alancing = ['ros','smote','clustercentroids']
+#scale = ['standard','robust','minmax','normalizer']
+
+
+[PLOT]
+name = mlp_trial
+confusion_matrix = True
+tsne = True
