+++ data/neurovoz/README.md
@@ -1,0 +1,16 @@
+# Neurovoz
+
+*Study of the automatic detection of Parkison’s Disease based on
+speaker recognition technologies and allophonic distillation*
+
+* [github](https://github.com/BYO-UPM/Neurovoz_Dababase/)
+* [paper](https://oa.upm.es/55198/1/INVE_MEM_2018_298847.pdf)
+* [dataset](https://zenodo.org/records/13647600)
+
+In order to publish this for nkululeko
+
+1) request the dataset from zenodo
+2) unzip into folder *data* (should be the default name)
+3) run *python process_database.py*
+4) cd ../..
+5) python -m nkululeko.nkululeko --config data/neurovoz/exp.ini
