+++ data/neurovoz/exp.ini
@@ -1,0 +1,17 @@
+[EXP]
+root = data/neurovoz/
+name = results
+[DATA]
+databases = ['data']
+data = ./data/neurovoz/neurovoz.csv
+data.type = csv
+data.absolute_path = False
+data.split_strategy = speaker_split
+data.colnames = {'Group':'depression', 'ID': 'speaker'}
+labels = ['HC', 'PD']
+target = depression
+[FEATS]
+type = ['os']
+scale = standard
+[MODEL]
+type = svm
