+++ data/neurovoz/exp.ini
@@ -7,11 +7,13 @@ data = ./data/neurovoz/neurovoz.csv
 data.type = csv
 data.absolute_path = False
 data.split_strategy = speaker_split
-data.colnames = {'Group':'depression', 'ID': 'speaker'}
+data.colnames = {'Group':'depression', 'ID': 'speaker', 'Sex': 'gender', 'Age': 'age'}
 labels = ['HC', 'PD']
 target = depression
 [FEATS]
 type = ['os']
 scale = standard
+[EXPL]
+value_counts = [['age'],['gender'],['duration']]
 [MODEL]
 type = svm
