### 2025-10-01 a530726
added general mappings
