+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.98.4"
+VERSION="0.98.5"
 SAMPLING_RATE = 16000
