+++ nkululeko/constants.py
@@ -1,2 +1,2 @@
-VERSION="0.98.5"
+VERSION="0.98.6"
 SAMPLING_RATE = 16000
