+++ data/neurovoz/exp.ini
@@ -8,6 +8,7 @@ data.type = csv
 data.absolute_path = False
 data.split_strategy = speaker_split
 data.colnames = {'Group':'depression', 'ID': 'speaker', 'Sex': 'gender', 'Age': 'age'}
+data.mapping = {'gender': {'0.0': "female", '1.0': "male"}}
 labels = ['HC', 'PD']
 target = depression
 [FEATS]
