### 2025-10-08 #8049 @faroit
@felixbur, I went ahead and did a grammar check where I found some minor things. Feel free to merge my PR [Fix grammar and punctuation in paper.md felixbur/nkululeko#301](https://github.com/felixbur/nkululeko/pull/301)
