### 2025-10-16 #8049 @faroit
@felixbur can you check the references again please and add two missing DOIs (maybe you can find more?)

- speechbrain: `10.48550/arXiv.2106.04624`
- wav2vec2: `10.48550/arXiv.2006.11477`
-
